### 2023-08-01 #5721 @editorialbot
[PRE REVIEW]: FrEDI: An R package for estimating future climate impacts within the United States
**Submitting author:** @knoiva-indecon (<a href="http://orcid.org/0000-0002-4412-5302"></a>)
**Repository:** https://github.com/USEPA/FrEDI/
**Branch with paper.md** (empty if default branch): joss
**Version:** v3.4.0
