### 2023-08-02 37909b3
Updated DOI info and corrected missing affiliations
Updated: DOI for Hartin 2023 and Kopp 2016. Added missing affiliations.
