+++ paper/paper.bib
@@ -2,7 +2,7 @@
 	title = {The {Long}-{Term} {Strategy} of the {United} {States}: {Pathways} to {Net}-{Zero} {Greenhouse} {Gas} {Emissions} by 2050},
 	url = {https://www.whitehouse.gov/wp-content/uploads/2021/10/US-Long-Term-Strategy.pdf},
 	institution = {U.S. Department of State and the Executive Office of the President},
-	author = {{Department of State}},
+	author = {{DOS}},
 	year = {2021}
 }
 
@@ -10,7 +10,7 @@
 	title = {Technical {Documentation} on the {Framework} for {Evaluating} {Damages} and {Impacts} ({FrEDI})},
 	url = {https://www.epa.gov/sites/default/files/2021-03/documents/cira2.0_technicalappendixfornca4.pdf},
 	institution = {U.S. Environmental Protection Agency},
-	author = {{Environmental Protection Agency}},
+	author = {{EPA}},
 	year = {2021},
 	keywords = {CIRA, EPA},
 }
@@ -19,7 +19,7 @@
 	title = {Climate {Change} and {Social} {Vulnerability} in the {United} {States}: {A} {Focus} on {Six} {Impacts}},
 	url = {https://www.epa.gov/cira/social-vulnerability-report},
 	institution = {U.S. Environmental Protection Agency},
-	author = {{Environmental Protection Agency}},
+	author = {{EPA}},
 	year = {2021},
 	keywords = {CIRA, EPA},
 }
@@ -27,8 +27,7 @@
 @article{Hartin:2023,
 	title = {Advancing the estimation of future climate impacts within the {United} {States}},
 	volume = {[preprint]},
-	url = {https://doi.org/10.5194/egusphere-2023-114},
-	abstract = {Evidence of the physical and economic impacts of climate change is a critical input to policy development and decision making. The potential magnitude of climate change damages, where, when, and to whom those damages may occur across the country, the types of impacts that will be most damaging, and the ability of adaptation to reduce potential risks are all important and interconnected. This study utilizes the reduced-complexity model, Framework for Evaluating Damages and Impacts (FrEDI), to rapidly assess economic and physical impacts of climate change in the contiguous United States (U.S.). Results from FrEDI show that net national damages increase overtime, with mean climate-driven damages estimated to reach \$2.9 trillion USD (95 \% CI: \$510 billion to \$12 trillion) annually by 2090. Climate-driven damages are largest for the health category, with the majority of damages in this category from the valuation estimates of premature mortality attributable to climate-driven changes in extreme temperature and air quality (O3 and PM2.5). Results from FrEDI also show that climate-driven damages vary by geographical region, with the Southeast experiencing the largest annual damages per capita (mean: \$9,300 per person annually, 95 \% CI: \$1,800–\$37,000 per person annually), whereas the smallest damages per capita are expected in the Southwest region (mean: \$6,300 per person annually, 95 \% CI: \$840–\$27,000 per person annually). Climate change impacts may also broaden existing societal inequalities, with Black or African Americans disproportionately affected by additional premature mortality from changes in air quality. This work significantly advances our understanding of the impacts from climate change to the U.S., in what U.S. regions impacts are happening, what sectors are being impacted, and which population groups being impacted the most.},
+	doi = {10.5194/egusphere-2023-114},
 	journal = {EGUsphere},
 	author = {Hartin, Corinne and McDuffie, Erin E. and Noiva, Karen and Sarofim, Marcus and Parthum, Bryan and Martinich, Jeremy and Barr, Sarah and Neumann, Jim and Willwerth, Jacqueline and Fawcett, Allen},
 	year = {2023},
[…]
