### 2023-08-02 28e6ada
Fixed DOI for Kopp 2016
Fixed DOI for Kopp 2016
