+++ paper/paper.bib
@@ -36,8 +36,7 @@
 @article{Kopp:2016,
 	title = {Temperature-driven global sea-level variability in the {Common} {Era}},
 	volume = {113},
-	doi = {10.1073/pnas.15170561},
-	abstract = {We assess the relationship between temperature and global sea-level (GSL) variability over the Common Era through a statistical metaanalysis of proxy relative sea-level reconstructions and tide-gauge data. GSL rose at 0.1 ± 0.1 mm/y (2σ) over 0–700 CE. A GSL fall of 0.2 ± 0.2 mm/y over 1000–1400 CE is associated with ∼0.2 °C global mean cooling. A significant GSL acceleration began in the 19th century and yielded a 20th century rise that is extremely likely (probability P≥0.95) faster than during any of the previous 27 centuries. A semiempirical model calibrated against the GSL reconstruction indicates that, in the absence of anthropogenic climate change, it is extremely likely (P=0.95) that 20th century GSL would have risen by less than 51\% of the observed 13.8±1.5 cm. The new semiempirical model largely reconciles previous differences between semiempirical 21st century GSL projections and the process model-based projections summarized in the Intergovernmental Panel on Climate Change’s Fifth Assessment Report.},
+	doi = {10.1073/pnas.1517056113},
 	number = {11},
 	journal = {PNAS},
 	author = {Kopp, Robert E. and Kemp, Andrew C. and Bittermann, Klaus and Rahmstorf, Stefan},
@@ -50,7 +49,6 @@
 	volume = {165},
 	issn = {1573-1480},
 	doi = {10.1007/s10584-021-03048-6},
-	abstract = {Characterizing the future risks of climate change is a key goal of climate impacts analysis. Temperature binning provides a framework for analyzing sector-specific impacts by degree of warming as an alternative or complement to traditional scenario-based approaches in order to improve communication of results, comparability between studies, and flexibility to facilitate scenario analysis. In this study, we estimate damages for nine climate impact sectors within the contiguous United States (US) using downscaled climate projections from six global climate models, at integer degrees of US national warming. Each sector is analyzed based on socioeconomic conditions for both the beginning and the end of the century. The potential for adaptive measures to decrease damages is also demonstrated for select sectors; differences in damages across adaptation response scenarios within some sectors can be as much as an order of magnitude. Estimated national damages from these sectors based on a reactive adaptation assumption and 2010 socioeconomic conditions range from \$600 million annually per degree of national warming for winter recreation to \$8 billion annually per degree of national warming for labor impacts. Results are also estimated per degree of global temperature change and for 2090 socioeconomic conditions.},
 	number = {1},
 	urldate = {2023-04-19},
 	journal = {Climatic Change},
