### 2023-08-03 #5721 @knoiva-indecon
@kthyng Thanks for the heads up and update. I fixed the DOIs that I could and added the missing DOIs that were available. Also fixed the missing affiliations. Will the @editorialbot try compiling the PDF again so I can confirm that the issues were fixed?
