### 2023-09-08 #5784 @pdebuyl
Hello @tonyewong thanks a lot for mentioning this, I had indeed not noticed it.

I will check with other editors about this and let you know if we may proceed as is or not.
