### 2023-10-02 #5784 @hahsan1
I have successfully installed the package and tested out the first example available in the link provided ([usepa.github.io/FrEDI](https://usepa.github.io/FrEDI)). I tried to use a dummy file I made for `tempInputFile`, but I'm getting an error after running the `import_inputs()` function:
```
In import_inputs():
	User supplied temperature input...
		Importing data from c(2000, 2001, 2002, 2003, 2004, 2005)...		Importing data from c(25, 26.4, 24, 23.6, 33, 23)...
Error in file.exists(.) : invalid 'file' argument
```

Is the format incorrect?

Also in Step 3, the write output commands should use `output_df$results` since it's a list.

In terms of the manuscript, can you speak to any other packages or code that is used in this application and how it compares to FrEDI?
