+++ FrEDI/data_tests/R/plot_DOW_byImpactTypes.R
@@ -1,0 +1,262 @@
+###### plot_DOW_byImpactType ######
+### This function plots degrees of warming by sector, variant, impact year, and type
+plot_DOW_byImpactTypes <- function(
+    data,
+    sector,
+    xCol      = "driverValue",
+    yCol      = "annual_impacts",
+    modelType = "GCM",
+    nTicks    = 5,
+    silent    = TRUE,
+    options   = list(
+      title      = "Impacts by Degrees of Warming",
+      # subtitle   = NULL,
+      xTitle     = expression("Degrees of Warming (°C)"),
+      yTitle     = "Impacts ($2015)",
+      lgdTitle   = "Model",
+      margins    = c(0, 0, .15, 0),
+      marginUnit = "cm",
+      theme      = NULL
+    )
+){
+  ###### Messaging ######
+  print_msg <- !silent
+  if(print_msg){ "Running plot_DOW_byImpactType()..." |> message()}
+  ### Model Type
+  # modelType %>% print
+  do_gcm    <- "gcm" %in% (modelType |> tolower())
+  do_slr    <- "slr" %in% (modelType |> tolower())
+  # modelType |> print(); do_gcm |> print(); do_slr |> print()
+  ###### Format Data ######
+  ### Filter to sector and convert to data frame
+  # data |> glimpse()
+  # data[["sector"]] |> unique() |> print(); modelType |> print()
+  sector0   <- sector; rm(sector)
+  df0       <- data  ; rm(data  )
+  df0       <- df0 |> filter(model_type == modelType)
+  df0       <- df0 |> filter(sector     == sector0  )
+  # df0 |> glimpse()
+
+  ###### Defaults ######
+  ### Defaults
+  def_titles  <- list(GCM="Impacts by Degrees of Warming", SLR="Impacts by GMSL (cm)")
+  def_xTitles <- list(GCM=expression("Degrees of Warming (°C)"), SLR="GMSL (cm)")
+  def_lgdLbls <- list(GCM="Model", SLR="Scenario")
+  def_margins <- list(GCM=c(0, 0, .15, 0), SLR=c(0, .2, .15, 0))
+  ### Values
+  title0      <- options[["title"     ]]
+  xTitle      <- options[["xTitle"    ]]
+  yTitle      <- options[["yTitle"    ]]
+  lgdLbl      <- options[["lgdTitle"  ]]
+  lgdPos      <- options[["lgdPos"    ]]
+  heights     <- options[["heights"   ]]
+  margins     <- options[["margins"   ]]
+  mUnit       <- options[["marginUnit"]]
+  theme0      <- options[["theme"     ]]
+  # xTitle |> print()
+  ### Plot options
+  hasTitle    <- !(is.null(title0  ))
+  hasXTitle   <- !(is.null(xTitle  ))
+  hasYTitle   <- !(is.null(yTitle  ))
+  hasLgdLbl   <- !(is.null(lgdLbl  ))
+  hasLgdPos   <- !(is.null(lgdPos  ))
+  hasHeights  <- !(is.null(heights ))
+  hasMargins  <- !(is.null(margins ))
+  hasMUnits   <- !(is.null(mUnit   ))
+  hasTheme    <- !(is.null(theme0  ))
+  ### Defaults: Default Heights Below
+  def_title   <- do_gcm |> ifelse(def_titles [["GCM"]], def_titles [["SLR"]])
+  def_xTitle  <- do_gcm |> ifelse(def_xTitles[["GCM"]], def_xTitles[["SLR"]])
+  def_margin  <- do_gcm |> ifelse(def_margins[["GCM"]], def_margins[["SLR"]])
[…]
