+++ FrEDI/data_tests/R/image_scale_utils.R
@@ -1,0 +1,417 @@
+###### get_p10Labels ######
+### Get power of 10 labels
+get_p10Labels <- function(
+    values,
+    type0 = c("p10") ### Power 10, power 1000, "p1000"
+){
+  ### Data tible
+  vals0 <- values; rm("values")
+  # dfx   <- tibble(value=vals0, type=type0)
+  dfx   <- tibble(value=vals0)
+  dfx   <- dfx |> rename_at(.vars=c("value"), ~c(type0))
+  ### Tibble to match to
+  c_p10 <- seq(0, 21)
+  df0   <- tibble(p10=c_p10)
+  df0   <- df0 |> mutate(p1000 = p10 %/% 3)
+  df0   <- df0 |> mutate(p1000 = p10 %/% 3)
+  df0   <- df0 |> mutate(label = case_when(
+    p1000 >= 0 & p1000 < 1 ~ "",
+    p1000 >= 1 & p1000 < 2 ~ "Thousands",
+    p1000 >= 2 & p1000 < 3 ~ "Millions",
+    p1000 >= 3 & p1000 < 4 ~ "Billions",
+    p1000 >= 4 & p1000 < 5 ~ "Trillions",
+    p1000 >= 5 & p1000 < 6 ~ "Quadillions",
+    p1000 >= 6 & p1000 < 7 ~ "Quintillions",
+    .default = ""
+  ))
+  ### Join data
+  dfx   <- dfx |> left_join(df0, by=c(type0))
+  ### Return
+  return(dfx)
+}
+
+###### fun_getScale ######
+### This function creates a set of breaks for a particular column
+### It returns a list of breaks, the power of 10, and the limits
+fun_getScale <- function(
+    data,
+    scaleCol = "driverValue",
+    nTicks   = 5
+){
+  ###### Defaults ######
+  # if(is.null(scaleCol)) scaleCol <- "driverValue"
+  ### Default is not to zero out in case there are negative numbers
+  # if(is.null(zero))     zero <- F
+  # if(is.null(nTicks)){nTicks <- 5}
+
+  ###### Min/Max Values ######
+  data <- data
+  xMin <- data[[scaleCol]] |> min(na.rm=T)
+  xMax <- data[[scaleCol]] |> max(na.rm=T)
+  # c(xMin, xMax)|> print()
+
+  ### Set minimum to zero unless the minimum is less than zero
+  ### Does the series have negative values? If not, zero out.
+  # zero     <- ifelse(lowValue < 0, F, zero)
+  if(xMin > 0){xMin <- 0}
+  if(xMax < 0){xMax <- 0}
+
+  ###### Bounds
+  bound_min <- xMin |> floor()
+  bound_max <- xMax |> ceiling()
+
+  ###### Min/Max Tibble ######
+  ### Limit names, values, bounds, etc
+  df_minMax <- tibble(
+    name  = c("min", "max"),
+    value = c(xMin, xMax),
+    bound = c(bound_min, bound_max),
+    boundType = c("floor", "ceiling")
+  )
+
+  ###### Log10 Values ######
+  ### Absolute value, Power of 10 and y-scale info
+  df_minMax <- df_minMax |> mutate(bound_abs = bound |> abs())
+  ### Calculate log 10 and replace values of infinity with 0
+  df_minMax <- df_minMax |> mutate(log10 = bound_abs |> log10())
+  df_minMax <- df_minMax |> mutate(log10 = log10 |> abs())
+  df_minMax <- df_minMax |> mutate(log10 = log10 |> na_if(Inf))
+  df_minMax <- df_minMax |> mutate(log10 = log10 |> replace_na(0))
+
+  ###### Power-10 Values ######
[…]

[…]
