### 2023-10-04 f10c196
Deleted createSystemData
Deleted createSystemData since it is no longer required
