### 2023-10-05 e6debe0
Deleted functions for getting cowplot elements...no longer using package cowplot in figure creation
