+++ FrEDI/R/utils_plot_DOW_bySector.R
@@ -1,63 +1,3 @@
-### Get Plot element
-get_cowplot_element <- function(
-    plot0,  ### Reference plot
-    type0, ### Plot element type: spacer, xlab-b, ylab-l
-    return_all = FALSE
-){
-  ### Get first list element
-  obj0     <- plot0[[1]]
-  type0 |> print()
-  ### Get cowplot component
-  # doComp   <- type0 %in% c("spacer", "xlab-b", "ylab-l")
-  doSpace  <- type0 %in% c("spacer")
-  doComp   <- type0 %in% c("xlab-b", "ylab-l")
-  doTitle  <- type0 %in% c("title" )
-  doLegend <- type0 %in% c("legend")
-  doUnlist <- type0 %in% c("spacer")
-
-  ### Get element
-  # plot_spacer   <- (ref_plot[[1]] %>% cowplot::get_plot_component(pattern="spacer", return_all=T))[[1]] %>% (function(x){ggplotify::as.grob(x)})
-  # plot_xTitle   <- (ref_plot[[1]] %>% cowplot::get_plot_component(pattern="xlab-b", return_all=F)) %>% (function(x){ggplotify::as.grob(x)})
-  if     (doComp  ){elem0 <- obj0 |> cowplot::get_plot_component(pattern=type0, return_all=return_all)}
-  else if(doSpace ){elem0 <- (obj0 |> cowplot::get_plot_component(pattern=type0, return_all=return_all))[[1]]}
-  else if(doTitle ){elem0 <- obj0 |> cowplot::get_title ()}
-  else if(doLegend){elem0 <- obj0 |> ggpubr::get_legend()}
-  else             {elem0 <- list()}
-  ### If unlist, get first list element
-  # if(doUnlist){elem0 <- elem0[[1]]}
-  ### Mutate to grob
-  if(doComp|doSpace  ){grob0 <- ggplotify::as.grob(elem0)}
-  else        {grob0 <- elem0}
-  # [[1]] %>% (function(x){ggplotify::as.grob(x)})
-  ### Return
-  return(grob0)
-} ### End get_cowplot_element
-
-### Get list of cowplot elements
-get_cowplot_elements <- function(
-    plot0, ### Plot object
-    types0 = c("spacer", "xlab-b", "ylab-l", "title", "legend")
-){
-  ### Get types
-  types0   <- types0 |> tolower()
-  ### Which types
-  doSpacer <- "spacer" %in% types0
-  doXlab   <- "xlab-b" %in% types0
-  doYlab   <- "ylab-l" %in% types0
-  doTitle  <- "title"  %in% types0
-  doLegend <- "legend" %in% types0
-  ### Initiate list
-  list0  <- list()
-  ### Add list elements
-  if(doSpacer){list0[["spacer"]] <- get_cowplot_element(plot0=plot0, type0="spacer", return_all=TRUE )}
-  if(doXlab  ){list0[["xlab-b"]] <- get_cowplot_element(plot0=plot0, type0="xlab-b", return_all=FALSE)}
-  if(doYlab  ){list0[["ylab-l"]] <- get_cowplot_element(plot0=plot0, type0="ylab-l", return_all=FALSE)}
-  if(doTitle ){list0[["title" ]] <- get_cowplot_element(plot0=plot0, type0="title" , return_all=TRUE )}
-  if(doLegend){list0[["legend"]] <- get_cowplot_element(plot0=plot0, type0="legend", return_all=TRUE )}
-
-  ### Return
-  return(list0)
-}
 
 ###### plot_DOW_bySector ######
 ### Get plot_DOW_byModelType_sector
