### 2023-10-05 #5784 @tonyewong
Regarding the license file -- I see that it's present in the `joss` branch, but not in `main`. Should that be included in the `main` branch too? Since presumably that will be the one that users would access and need to know the license info.
