### 2023-10-05 a990542
Update utils_report_figures_scales.R
Edited fun_limitsByGroup to work for sector-only grouping
Edited get_sector_plotInfo to work for sector-only grouping
