### 2023-10-05 9cd2c3e
Update utils_plot_DOW_byModelType.R
Added Y Title, simplified plot code
