### 2023-10-05 f776c1a
Update utils_plot_DOW_byImpactTypes.R
Added Y Title, simplified plot grid code
