### 2023-10-05 50de078
Update utils_create_report_figures.R
Added sector list to scenario code. Edited get_ciraDataObject to be able to be used in nonFrEDI packages.
