+++ FrEDI/R/utils_create_report_figures.R
@@ -1,22 +1,56 @@
-### Add list names
+### Add names to list object
 addListNames <- function(
     list0, ### List object
     names0 ### Names to give to list or data frame
 ){
   names(list0) <- names0
   return(list0)
-}
+} ### End addListNames
 
 ### This function makes it easier to get data objects from the sysdata.rda file
-get_ciraDataObj <- function(x, listall=FALSE){
-  x_listName <- "rDataList"
-  if(exists(x_listName)){new_x <- parse(text=x_listName) |> eval()}
-  else                  {new_x <- utils::getFromNamespace(x_listName, "FrEDI")}
+get_ciraDataObj <- function(
+    x,       ### Object name
+    listall  = FALSE,
+    listName = "rDataList",
+    pkg      = "FrEDI",
+    lib.loc  = .libPaths()[1] ### Library path to look for packages
+    ){
+  ### Messaging
+  msg0    <- "\t"
+  ### Check if list name exists
+  exists0 <- listName |> exists()
+  ### If the listname exists in the name space, parse it
+  ### Otherwise, grab it from a package name space
+  if(exists0){new_x <- parse(text=listName) |> eval()}
+  else       {
+    ### Check if package & list name
+    pkgList0    <- lib.loc |> installed.packages()
+    pkgExists0  <- pkg %in% pkgList0
+    if(!pkgExists0){
+      msg0 |> paste0("Package doesn't exist...") |> message()
+      msg0 |> paste0("Exiting...") |> message()
+      return()
+    } ### End if(!pkgExists0)
+    else           {new_x <- getFromNamespace(listName, ns=pkg)}
+  } ### End else(exists0)
 
+  ### Whether to list all items in data object or not
   if(listall) {return_x <- new_x |> names()}
   else        {return_x <- new_x[[x]]}
+  ### Return
   return(return_x)
-}
+} ### End get_ciraDataObj
+
+### Get column values from a tibble
+get_column_values <- function(
+    df0,    ### Tibble
+    col0,   ### Column
+    unique0 = FALSE ### Unique values
+){
+  vals0 <- df0[[col0]]
+  if(unique0){vals0 <- vals0 |> unique()}
+  return(vals0)
+} ### End get_column_values
 
 #### Summarize missing values
 sum_with_na <- function(
@@ -52,18 +86,27 @@ sum_with_na <- function(
   if(drop0){df0 <- df0 |> select(-c("is_NA"))}
   ### Return
   return(df0)
-
-}
-
+} ### End sum_with_na
 
 ### Filter to five year values
 filter_years <- function(
     df0, ### data
-    years = seq(2010, 2090, by=5)
+    byYears = 5, ### Number of years
+    years   = NULL ### Override by years
 ){
+  ### Check whether user provided values to years argument
+  nullYears <- years |> is.null()
+  hasYears  <- !nullYears
+  ### If user did not provide values to years argument
+  if(!hasYears){
+    range0 <- df0[["year"]] |> range(na.rm=TRUE)
+    years  <- seq(range0[1], range0[2], by=byYears)
+  }
+  ### Filter to years
   df0 <- df0 |> filter(year %in% years)
+  ### Return
   return(df0)
-}
+} ### End filter_years
 
 ### Filter values
 ### Format values to specific number of decimal places
@@ -82,41 +125,45 @@ create_constant_temp_scenario <- function(
     type0   = "conus",
     prefix0 = "Other_Integer" ### Prefix for scenario
 ){
[…]
