### 2023-10-05 #5784 @tonyewong
The some of the vignettes from the GitHub Pages documentation site, the `kableExtra` package is loaded and used. I may have missed if elsewhere this dependency is noted and the pacakge is installed, and this isn't a core part of the FrEDI package itself. But, could be nice to include in the initial package installations so it doesn't trip folks up later potentially.

I also noticed that the top-level README.md in the Github repo is very bare bones. This is probably fine since the documentation site is *very* thorough. The "about" in the upper-right of the github repo links to it, but could we add a statement in the README.md that explicitly tells newcomers to check out the documentation site for further info and examples?
