### 2023-10-05 #5784 @tonyewong
>  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

For community guidelines, it looks like the [contributions part of the documentation](https://usepa.github.io/FrEDI/articles/contributing.html) is still under construction. Is this ready to be added?
