### 2023-10-05 cc8a977
Update create_report_figures.R
Turned script for creating report figures into a function, create_DoW_results
