+++ FrEDI/scripts/create_report_figures.R
@@ -1,406 +1,483 @@
-###### Set Up Environment ######
-###### ** Load Packages ######
+###### Function to run Degree of Warming (DOW) scenarios and create figures
+### 2023.10.05: Renamed to create_DoW_results from create_report_figures
+###### Load Packages ######
 require(tidyverse)
 require(ggpubr)
-require(cowplot)
-require(FrEDI)
+# require(cowplot)
+# require(FrEDI)
 
+###### create_DoW_results ######
+create_DoW_results <- function(
+    fpath    = "." ,  ### Path to main FrEDI directory
+    sectors  = FrEDI::get_sectorInfo(), ### Which sectors
+    gcmYears = c(2010, 2050, 2090), ### Which years to report on for GCM sectors
+    slrYears = c(2010, 2050, 2090), ### Which years to report on for SLR sectors
+    img_dev  = "pdf", ### Image device
+    saveFile = FALSE, ### Save file
+    silent   = TRUE,  ### Degree of messaging
+    testing  = FALSE  ### Whether to print out extra diagnostic values
+){
+  ###### Messaging ######
+  do_msg          <- !silent
 
-###### ** Set Paths ######
-projectPath      <- getwd()     |> file.path("FrEDI")
-testsPath        <- projectPath |> file.path("data_tests")
-codePath         <- testsPath   |> file.path("R")
+  ###### Initialize Return List ######
+  resultsList     <- list()
 
-### Output Paths
-mainResultsPath <- projectPath     |> file.path("report_figures")
-dowResultsPath  <- mainResultsPath |> file.path("DoW")
-fig7ResultsPath <- mainResultsPath |> file.path("fig7")
-appxResultsPath <- mainResultsPath |> file.path("appendix_figures")
-# ### Check and create paths
-# mainResultsPath |> check_and_create_path()
-# dowResultsPath  |> check_and_create_path()
-# fig7ResultsPath |> check_and_create_path()
-# appxResultsPath |> check_and_create_path()
+  ###### Set Up Environment ######
+  ###### ** Set Paths ######
+  # projectPath      <- getwd()     |> file.path("FrEDI")
+  projectPath     <- fpath; rm(fpath)
+  codePath        <- projectPath |> file.path("R")
+  # projectPath |> list.files() |> print()
+  # codePath |> list.files() |> print()
+  ### Output Paths
+  mainResultsPath <- projectPath
+  if(testing){mainResultsPath <- mainResultsPath |> file.path("data_tests")}
+  mainResultsPath <- mainResultsPath |> file.path("report_figures")
+  dowResultsPath  <- mainResultsPath |> file.path("DoW")
+  fig7ResultsPath <- mainResultsPath |> file.path("fig7")
+  appxResultsPath <- mainResultsPath |> file.path("appendix_figures")
+  # ### Check and create paths
+  # mainResultsPath |> check_and_create_path()
+  # dowResultsPath  |> check_and_create_path()
+  # fig7ResultsPath |> check_and_create_path()
+  # appxResultsPath |> check_and_create_path()
 
-###### ** Load Code ######
-### Custom function to load code from a specified path
-# codeFiles <- codePath |> list.files(pattern=".R", full.names = T); codeFiles |> basename()
-# for(code_i in codeFiles){code_i |> source()}
-loadCustomFunctions <- function(fpath=codeFiles, local=FALSE){
[…]
