### 2023-10-05 11ade8d
Update DESCRIPTION
Edited description: removed dplyr, ggplot2, tidyr (which are included in tidyverse)
Added ggpubr (which is required for plotting)
Added devtools to suggested packages
