### 2023-10-05 e0d0501
Renamed create_report_figures
Renamed create_report_figures and added argument to return the data and plot list
