### 2023-10-05 cdeeb3e
Update create_DoW_results.R
Added code for whether to load the image code from source, package, or project
