### 2023-10-05 066fb72
Update testing_general_fredi_test.R
Changed function pipe to %>%
