### 2023-10-05 9f1633a
Update create_DoW_results.R
Added code to specify how to load the image code
