+++ FrEDI/DESCRIPTION
@@ -2,25 +2,28 @@ Package: FrEDI
 Title: The Framework for Evaluating Damages and Impacts (FrEDI)
 Version: 3.4
 Authors@R: c(person("Corinne", "Hartin",
-                    email = "hartin.corinne@epa.gov",
-                    role = c("aut","cre"),
+                    email   = "hartin.corinne@epa.gov",
+                    role    = c("aut","cre"),
                     comment = c(ORCID = "0000-0003-1834-6539",
-                    github = "cahartin")),
-              person("Erin","McDuffie",
-                      email = "mcduffie.erin.e@epa.gov",
-                      role = c("aut","ctb"),
-                      comment = c(ORCID = "0000-0002-6845-6077",
-                                  github = "emcduffie")),
-              person("Karen","Noiva",
-                      role = c("aut","ctb"),
-                      comment = c(github = "knoiva-indecon")),
-              person("Will","Maddock",
-                      role = "ctb",
-                      comment = c(github = "maddockw")),
-              person("Tony","Gardella",
-                      role = "ctb",
-                      comment = c(github = "tonygard-indecon"))
-              )
+                                github = "cahartin")),
+             person("Erin","McDuffie",
+                     email   = "mcduffie.erin.e@epa.gov",
+                     role    = c("aut","ctb"),
+                     comment = c(ORCID = "0000-0002-6845-6077",
+                                 github = "emcduffie")),
+             person("Karen","Noiva",
+                     role    = c("aut","ctb"),
+                     comment = c(ORCID = "0000-0002-4412-5302", 
+                                 github = "knoiva-indecon")),
+             person("Will","Maddock",
+                     role    = "ctb",
+                     comment = c(ORCID = "0000-0002-6107-0315",
+                                 github = "maddockw")),
+             person("Tony","Gardella",
+                     role    = "ctb",
+                     comment = c(ORCID = "0000-0003-4380-3412",
+                                 github = "tonygard-indecon"))
+             )
 URL: https://usepa.github.io/FrEDI, https://epa.gov/cira/FrEDI/, https://github.com/USEPA/FrEDI/
 Description: This R package models damages and impacts from climate change in the contiguous United States (CONUS) 
     using the Framework for Evaluating Damages and Impacts (FrEDI) developed as part of the Climate Change Impacts and 
