### 2023-10-06 #5784 @pdebuyl
Hi all, thanks for the progress in the review!

Regarding the license it should be in the repository. BTW, it is indeed nice to have a self-sufficient (even small) README file, pointing to the docs for more info.
