### 2023-10-06 da4ca71
New function create_fig_scale_note
File "utils_create_report_figures.R":
- Created function `create_fig_scale_note` to create a note for multiple impact types

File: "utils_plot_DOW_byImpactTypes.R":
- Made note on impact type contingent on multiple impact types and variants

File: "create_DoW_results.R":
- Set default loadCode to "project"
