+++ FrEDI/scripts/create_DoW_results.R
@@ -16,7 +16,7 @@ create_DoW_results <- function(
     img_dev  = "pdf", ### Image device
     silent   = TRUE,  ### Degree of messaging
     testing  = FALSE, ### Whether to print out extra diagnostic values
-    loadCode = "source", ### Whether to load code as source or devtools
+    loadCode = "project", ### Whether to load code as source or devtools
     saveFile = FALSE, ### Save file
     return   = TRUE   ### Whether to return list object
 ){
@@ -24,6 +24,7 @@ create_DoW_results <- function(
   ### Messaging
   do_msg          <- !silent
   ### Initialize Return List
+  return0         <- return; rm(return)
   resultsList     <- list()
   ### How to load code
   loadProject     <- "project" %in% (loadCode |> tolower())
@@ -88,7 +89,7 @@ create_DoW_results <- function(
   newSectorNames <- c_sectorNames |> format_sectorNames(thresh0 = breakChars)
   ### Message and save to list
   if(testing|do_msg) "Formatting sector names for plotting..." |> message()
-  if(testing) resultsList[["sectorNames"]] <- c_sectorNames
+  if(return0) resultsList[["sectorNames"]] <- c_sectorNames
   if(testing) c_sectorNames |> print(); newSectorNames |> print()
 
   ###### ** Constants ######
@@ -114,7 +115,7 @@ create_DoW_results <- function(
     mutate(scenario  = prefix |> paste0("_", tempLabel))
   ### Glimpse & save to list
   if(testing) "Creating tibble of integer scenario information..." |> message()
-  if(testing) resultsList[["df_scenarios"]] <- df_scenarios
+  if(return0) resultsList[["df_scenarios"]] <- df_scenarios
   if(testing) df_scenarios |> glimpse()
   ### Vector of scenarios
   c_scen_con     <- df_scenarios |> filter(tempType == "conus" ) |> get_column_values(col0="scenario")
@@ -136,7 +137,7 @@ create_DoW_results <- function(
     )
   }) %>% (function(x){do.call(rbind, x)})
   ### Glimpse, message, & save
-  if(testing) resultsList[["inputs_df_int"]] <- inputs_df_int
+  if(return0) resultsList[["inputs_df_int"]] <- inputs_df_int
   if(testing) inputs_df_int |> glimpse()
 
   ###### Run Scenarios ######
@@ -154,7 +155,7 @@ create_DoW_results <- function(
     joinCols  = c("year")
   )
   ### Glimpse results
-  if(testing) resultsList[["df_int_byType"]] <- df_int_byType
+  if(return0) resultsList[["df_int_byType"]] <- df_int_byType
   if(testing) df_int_byType |> glimpse()
   ### Save results
   if(saveFile){
@@ -181,7 +182,7 @@ create_DoW_results <- function(
     joinCols  = c("year")
   )
   ### Glimpse results
-  if(testing) resultsList[["df_int_totals"]] <- df_int_totals
+  if(return0) resultsList[["df_int_totals"]] <- df_int_totals
   if(testing) df_int_totals |> glimpse()
   ### Save results
   if(saveFile){
@@ -214,7 +215,7 @@ create_DoW_results <- function(
     adjCol      = "impact_billions"
   )
   ### Glimpse
-  if(testing) resultsList[["sum_gcm_totals"]] <- sum_gcm_totals
+  if(return0) resultsList[["sum_gcm_totals"]] <- sum_gcm_totals
[…]
