### 2023-10-06 2908173
Update utils_report_figures_format_sector_names.R
File "utils_report_figures_format_sector_names.R":
- Created function `repCollapse` to repeat a new line a certain number of times and then collapse it
- Changed extra lines to be added all at beginning of sector name
