### 2023-10-06 6a9adf0
Update utils_report_figures_format_sector_names.R
Fixed new line string
