### 2023-10-06 7a8d5f8
Fixes to report figure code
File "utils_create_report_figures.R"
- Changed functions `sum_impacts_by_DoW` and `sum_impacts_by_DoW_years` and `get_fig7_slrImpacts` to take argument `aggOnly` to determine whether to only include sectors in Figure 7 for which `includeaggregate==1`
- Changed function `plot_DoW_by_modelYear` to remove lines of code factoring sectors for labeling...moved lines of code to `utils_plot_DOW_byImpactTypes`
- Moved formatting of sector names into `utils_plot_DOW_byModelType`
- Fixed years for which plots are printed for Fig7 and appendix figures
