+++ FrEDI/R/utils_create_report_figures.R
@@ -359,7 +359,6 @@ run_scenarios <- function(
 } ### End run_scenarios
 
 
-
 ### Summarize results by degree of warming for a single specified year
 sum_impacts_byDoW <- function(
     df0,
@@ -370,21 +369,19 @@ sum_impacts_byDoW <- function(
     sumCol      = "annual_impacts",
     groupVars   = c("variant", "impactType", "impactYear"),
     impactYears = c("Interpolation", "NA", "2010", "2090"),
-    primary     = FALSE,
+    aggOnly     = FALSE,
     adjVal      = 1/10**9, ### Factor to multiply by
     adjCol      = "impact_billions",
     silent      = FALSE
 ){
   ### Filter to includeaggregate==1
-  # aggOnly         <- TRUE
-  aggOnly         <- !bySector
   if(aggOnly){df0 <- df0 |> filter(includeaggregate==1)}
   ### Filter to sector primary
+  primary         <- !bySector
   if(primary){df0 <- df0 |> filter(sectorprimary   ==1)}
-
   ### Filter to scenarios
   scenarios0 <- scenarios; rm("scenarios")
-  df0        <- df0 |> filter(scenario %in% scenarios0)
+  df0        <- df0 |> filter(scenario   %in% scenarios0)
   ### Filter to appropriate models
   df0        <- df0 |> filter(model_type %in% models)
   ### Filter to appropriate impact years
@@ -424,20 +421,19 @@ sum_impacts_byDoW <- function(
   # df0 %>% glimpse
   ### Return
   return(df0)
-}
+} ### End sum_impacts_byDoW
 
 ### Summarize results by degree of warming for multiple specified years
 sum_impacts_byDoW_years <- function(
     df0, ### Outputs of sum_impactsByDegree
     scenarios,
     bySector    = FALSE,
-    primary     = FALSE,
-    years       = c(2010, 2050, 2090),
+    years       = c(2010, 2090),
     models      = c("GCM", "SLR"),
     sumCol      = "annual_impacts",
     groupVars   = c("variant", "impactType", "impactYear"),
     impactYears = c("Interpolation", "NA", "2010", "2090"),
-    # aggOnly     = TRUE,
+    aggOnly     = FALSE,
     adjVal      = 1/10**9, ### Factor to multiply by
     adjCol      = "impact_billions",
     silent      = FALSE
@@ -445,15 +441,15 @@ sum_impacts_byDoW_years <- function(
   ### Filter to includeaggregate==1
   ### Filter to includeaggregate==1
   # aggOnly         <- TRUE
-  aggOnly    <- !bySector
   if(aggOnly){df0 <- df0 |> filter(includeaggregate==1)}
   ### Filter to sector primary
+  primary    <- !bySector
   if(primary){df0 <- df0 |> filter(sectorprimary   ==1)}
   ### Run scenarios
   nYears <- years |> length()
 
   ### Get list
-  list0   <- years %>% map(function(.x){
+  list0   <- years |> map(function(.x){
     paste0("Summarizing values for ", which(years == .x), "/" , nYears, " years...") |> message()
     df_x <- sum_impacts_byDoW(
       df0         = df0,
@@ -464,14 +460,13 @@ sum_impacts_byDoW_years <- function(
       impactYears = impactYears,
       year        = .x,
       models      = models,
-      # aggOnly     = aggOnly,
+      aggOnly     = aggOnly,
       adjVal      = adjVal,
       adjCol      = adjCol,
       silent      = silent
[…]
