### 2023-10-06 c7c109b
change artifact path with *
