### 2023-10-06 a06d366
take out conditionals of upload
