+++ .github/workflows/test_fredi.yml
@@ -130,7 +130,7 @@ jobs:
         with:
           name: Test Data
           path: |
-            ./data_tests/general/
+            ./FrEDI/data_tests/general/
 
       - name: Upload Report Figures
         uses: actions/upload-artifact@v3
@@ -139,7 +139,7 @@ jobs:
         with:
           name: Report Data
           path: |
-            ./data_tests/general/
+            ./FrEDI/data_tests/general/
 # #
 #       - name: Commit results
 #         run: |
