### 2023-10-07 d129638
Improved figure code
Added additional code to format legend depending on the number of variants (i.e., plot grid width)
Added a try catch to some of the saving functions
