+++ FrEDI/R/utils_save_report_objects.R
@@ -13,7 +13,7 @@ check_and_create_path <- function(
   if(create1) {
     "Directory does not exist. " |> message()
     "\t" |> paste0("Creating directory...") |> message()
-    fpath |> dir.create()
+    try0 <- fpath |> dir.create() |> try()
     "\t" |> paste0("...Directory created.") |> message()
   } ### End if(create1)
 } ### End check_and_create_path
@@ -67,7 +67,7 @@ save_image <- function(
   ### Plot options
   dev0      <- device |> tolower()
   h0        <- options[["height"]]
-  w0        <- options[["height"]]
+  w0        <- options[["width" ]]
   res0      <- options[["res"]]
   units0    <- options[["units"]]
   ### File name
@@ -78,7 +78,7 @@ save_image <- function(
   fpath     <- fdir  |> file.path(fname)
 
   ### Save Image
-  saved0    <- ggsave(fname, plot = obj0, device=dev0, path=fdir, width=w0, height=h0, units=units0)
+  saved0    <- ggsave(fname, plot = obj0, device=dev0, path=fdir, width=w0, height=h0, units=units0) |> try()
   ### Return
 } ### End save_image
 
@@ -97,51 +97,81 @@ save_appendix_figures <- function(
   ### Create directory if it doesn't exist
   fdir      <- fpath; rm("fpath")
   fdir      <- fdir |> file.path("images")
-  created0  <- fdir  |> check_and_create_path(createDir=createDir)
+  created0  <- fdir |> check_and_create_path(createDir=createDir)
   ### Prepare data
-  df0       <- df0 |> filter(model_type == modelType)
+  df0       <- df0  |> filter(model_type %in% modelType)
   list0     <- plotList[[modelType]]
   ### Unique values
   names0    <- list0  |> names()
   sectors0  <- names0 |> map(function(.x){str_split(string=.x, pattern="_")[[1]][1]}) |> unlist() |> unique()
   refYears0 <- names0 |> map(function(.x){str_split(string=.x, pattern="_")[[1]][2]}) |> unlist() |> unique()
-  # names0 |> print(); sectors0 |> print(); refYears0 |> print(); impYears0 |> print()
+  # names0 |> print(); #sectors0 |> print(); refYears0 |> print()
 
   ### Iterate over sectors
   names0 |> map(function(.x){
     ### Plot .x
     list_x    <- list0[[.x]]
-    .x |> print(); #list_x |> names() |> print()
-    # .x |> c(list_x |> names()) |> print()
-    # "got here" |> print()
+    .x |> print()
+
     ### Split name into sector and ref year
     sector_x  <- .x |> map(function(.y){str_split(string=.y, pattern="_")[[1]][1]}) |> unlist() |> unique()
-    refYear_x <- .x |> map(function(.y){str_split(string=.y, pattern="_")[[1]][2]}) |> unlist() |> unique()
+    year_x    <- .x |> map(function(.y){str_split(string=.y, pattern="_")[[1]][2]}) |> unlist() |> unique()
     fname_x   <- sector_x
-    ### Filter to data
-    df_x      <- df0 |> filter(sector == .x)
-    ### Unique sector values
-    # impYears0 <- list0[[1]] |> names()
-    c_years <- list_x |> names()
-    # c_years <- df_x[["impactYear"]] |> unique()
-    c_types <- df_x[["impactType"]] |> unique()
-    c_vars  <- df_x[["variant"   ]] |> unique()
+    # sector_x |> c(year_x) |> print()
 
-    # c_types |> print(); c_types |> print(); c_vars |> print();
+    ### Filter to data
[…]
