### 2023-10-07 f9068ed
Update test_fredi.yml
Updated FrEDI_test results to add additional tests
