### 2023-10-11 USEPA/FrEDI#94 @tonyewong
license file in the repo
per https://github.com/openjournals/joss-reviews/issues/5784#issuecomment-1750160254
