### 2023-10-12 220faf6
Update test_fredi.yml
make boolean check true
