+++ .github/workflows/test_fredi.yml
@@ -137,7 +137,7 @@ jobs:
 
       - name: Upload Tests
         if: |
-          inputs.run_tests == 'yes'
+          inputs.run_tests == 'true'
         uses: actions/upload-artifact@v3
         with:
           name: Test Data
@@ -147,7 +147,7 @@ jobs:
 
       - name: Upload Report Figures
         if: |
-          inputs.make_figures == 'yes'
+          inputs.make_figures == 'true'
         uses: actions/upload-artifact@v3
         with:
           name: Report Data
