### 2023-10-12 d9d44e8
Update utils_report_figures_scales.R
Added some lines to function "lim_bySector" in "utils_report_figures_scales.R" to convert infinite values to NA (produced, for instance, if there are no non-missing values when calculating `min(na.rm=TRUE)`)
