+++ FrEDI/R/utils.R
@@ -48,12 +48,14 @@ interpolate_annual <- function(
     column    = NULL, ### Column to create results for
     rule      = NULL, ### for interpolation,
     method    = NULL, ### method for interpolation; default=linear
-    region    = "National.Total" ### Region if "addRegion"
+    region    = "National.Total", ### Region if "addRegion"
+    byState   = FALSE ### If breakdown by state
 ){
   ###### Data Info ######
   ##### Columns
   dataCols  <- data %>% names
   defCols   <- c("year", "region")
+  if (byState) {defCols <- defCols |> c("state", "postal")}
   defCol0   <- dataCols[!(dataCols %in% defCols)][1]
   column0   <- ifelse(is.null(column), defCol0, column)
   othCols   <- dataCols[!(dataCols %in% c(defCols, column0))]
@@ -71,9 +73,12 @@ interpolate_annual <- function(
   ##### Regions
   regions0  <- region
   addRegion <- !("region" %in% dataCols)
-  if(addRegion) {data      <- data %>% mutate(region = regions0[1])}
-  else          {regions0  <- data %>% get_uniqueValues("region")}
+  if(addRegion)   {data      <- data %>% mutate(region = regions0[1])}
+  else            {regions0  <- data %>% get_uniqueValues("region")}
   rm("addRegion")
+
+  #### States
+  if (byState) {states0 <- data |> get_uniqueValues("state"); postals0 <- data |> get_uniqueValues("postal")}
 
   ###### Interpolation Info ######
   ### - Return NA values outside of extremes
@@ -90,19 +95,39 @@ interpolate_annual <- function(
   ### Filter to the region and then interpolate missing values
   cols0     <- c("x", "y");
   cols1     <- c("year", column0)
-  df_interp <- regions0 %>% lapply(function(region_i){
-    ### Values
-    which_i <- (data$region==region_i) %>% which
-    x_i     <- years0[which_i]
-    y_i     <- values0[which_i]
-    ### Approximate
-    new_i   <- approx(x = x_i, y = y_i, xout = years, rule = rule, method = method)
-    new_i   <- new_i %>% as.data.frame
-    new_i   <- new_i %>% rename_at(.vars=c(all_of(cols0)), ~cols1)
-    new_i   <- new_i %>% mutate(region = region_i)
-    ### Return
-    return(new_i)
-  }) %>% (function(i){do.call(rbind, i)})
+
+  if (!byState) {
+    df_interp <- regions0 %>% lapply(function(region_i){
+      ### Values
+      which_i  <- (data$region==region_i) %>% which
+      x_i      <- years0[which_i]
+      y_i      <- values0[which_i]
+      ### Approximate
+      new_i   <- approx(x = x_i, y = y_i, xout = years, rule = rule, method = method)
+      new_i   <- new_i %>% as.data.frame
+      new_i   <- new_i %>% rename_at(.vars=c(all_of(cols0)), ~cols1)
+      new_i   <- new_i %>% mutate(region = region_i)
+      ### Return
+      return(new_i)
+    }) %>% (function(i){do.call(rbind, i)})
+
+  } else { ### By state
+    df_interp <- states0 %>% lapply(function(state_i){
+      ### Values
+      which_i <- (data$state==state_i) %>% which
+      postal_i <- data$postal[which_i]
+      region_i <- data$region[which_i]
+      x_i     <- years0[which_i]
+      y_i     <- values0[which_i]
+      ### Approximate
+      new_i   <- approx(x = x_i, y = y_i, xout = years, rule = rule, method = method)
+      new_i   <- new_i %>% as.data.frame
+      new_i   <- new_i %>% rename_at(.vars=c(all_of(cols0)), ~cols1)
[…]
