### 2023-10-17 86e8c46
Adjusted utils for loading data to accomodate state
