+++ createSystemData/R/createSystemData.R
@@ -12,11 +12,11 @@ createSystemData <- function(
   ###### Set up the environment ######
   ### Level of messaging (default is to message the user) and save behavior
   msgUser     <- !silent
-  
+
   ###### Initialize Data List ######
   ### Initialize list of data to save
   rDataList   <- list()
-  
+
   ###### Create File Paths ######
   ### Output file
   outPath     <- ifelse(outPath %>% is.null, "." %>% file.path("data", "sysdata.rda"), outPath)
@@ -30,7 +30,7 @@ createSystemData <- function(
   configPath <- configPath %>% dirname()
   configFile <- configPath %>% file.path(configFile)
   # configFile %>% print
-  
+
   ###### Configuration Data ######
   ### Read in configuration data and assign data tables to objects in the list
   if(msgUser) {paste0("Loading config info from '", configFile, "'...") %>% message()}
@@ -45,30 +45,30 @@ createSystemData <- function(
   get_uniqueValues      <- utils::getFromNamespace("get_uniqueValues"     , "FrEDI")
   get_scenario_id       <- utils::getFromNamespace("get_scenario_id"      , "FrEDI")
   extend_slr            <- utils::getFromNamespace("extend_slr"           , "FrEDI")
-  
+
   get_impactFunctions   <- utils::getFromNamespace("get_impactFunctions"  , "FrEDI")
   interpolate_annual    <- utils::getFromNamespace("interpolate_annual"   , "FrEDI")
   match_scalarValues    <- utils::getFromNamespace("match_scalarValues"   , "FrEDI")
   convertTemps          <- utils::getFromNamespace("convertTemps"         , "FrEDI")
   temps2slr             <- utils::getFromNamespace("temps2slr"            , "FrEDI")
-  
-  ###### Assign Data Objects ###### 
+
+  ###### Assign Data Objects ######
   ###### This section reads in data from the data file and returns a list of tables
   ###### Table names physScalars_table1, econScalars_table1, outputs_table1, co_sectors
   ### Add loaded tables to data list
   ### Load data from file
-  loadDataList <- dataList 
+  loadDataList <- dataList
   rDataList    <- rDataList %>% c(loadDataList)
   if(msgUser) {paste0("\t", messages_data[["loadInputs"]]$success) %>% message()}
-  for(name_i in names(loadDataList)) {assign(name_i, loadDataList[[name_i]]); rm("name_i")} 
-  
+  for(name_i in names(loadDataList)) {assign(name_i, loadDataList[[name_i]]); rm("name_i")}
+
   ###### Sector Info ######
   ### Exclude some sectors, get the number of sectors and sector info
   ### Sector info with additional sector info: df_sectorsInfo
   ### Sector info with models: df_sectorsModels
   num_sectors  <- co_sectors %>% nrow #; num_sectors
   sector_ids   <- co_sectors$sector_id
-  
+
   ###### Default Driver Scenarios ######
   ### Get reference years and add to fredi_config
   refYear_temp <- (co_modelTypes %>% filter(modelUnitType=="temperature"))$modelRefYear %>% unique
@@ -92,7 +92,7 @@ createSystemData <- function(
   temp_default <- temp_default    %>% select(-c(all_of(drop0)))
   rDataList[["temp_default"]] <- temp_default
   # temp_default %>% glimpse
-  
+
   ### Calculate annual values for SLR from global temperatures and add to list
   ### Ref year not needed since temps2slr will zero out values
   slr_default  <- temps2slr(temps=temp_default$temp_C_global, years=temp_default$year)
[…]
