### 2023-10-18 36edf38
Update test_fredi.yml
Changed test0 to tests0 in rm command
