### 2023-10-24 5ca2634
Updated utils to cooperate with new state data configuration in FrEDI_Data
