### 2023-10-25 bf09e21
Update utils.R
Replaced some magrittR pipe operators with native pipe operators
