### 2023-10-27 9c24dcc
Updated get_econAdjValues and get_annual_model_stats
