### 2023-11-03 c69f904
First batch of byState changes to run_fredi
