+++ FrEDI/R/run_fredi.R
@@ -15,6 +15,7 @@
 #' @param maxYear=2090 A numeric value indicating the maximum year for the analysis.
 #' @param thru2300 A ` TRUE/FALSE` shortcut that overrides the maxYear argument to run the model to 2300.
 #' @param outputList A ` TRUE/FALSE` value indicating whether to output results as a data frame object (`outputList=FALSE`, default) or to return a list of objects (`outputList=TRUE`) that includes information about model provenance (including input arguments and input scenarios) along with the data frame of results.
+#' @param byState A `TRUE/FALSE` value indicating whether to run at the state level for available sectors (default=`FALSE`).
 #' @param silent A `TRUE/FALSE` value indicating the level of messaging desired by the user (default=`TRUE`).
 #'
 #'
@@ -146,6 +147,7 @@ run_fredi <- function(
     maxYear    = 2090,
     thru2300   = FALSE,
     outputList = FALSE, ### Whether to return input arguments as well as results. [If TRUE], returns a list instead of a data frame
+    byState    = FALSE, ### Whether to run at state level for available sectors
     silent     = TRUE   ### Whether to message the user
 ){
 
@@ -174,6 +176,12 @@ run_fredi <- function(
   ### Configuration and data list
   for(i in 1:length(fredi_config)) assign(names(fredi_config)[i], fredi_config[[i]])
   for(i in 1:length(rDataList   )) assign(names(rDataList)[i], rDataList[[i]])
+  for(i in 1:length(frediData   )) assign(names(frediData)[i], frediData[[i]])
+  if(byState){
+    for(i in 1:length(stateData   )) assign(names(stateData)[i], stateData[[i]])
+  } else{
+    for(i in 1:length(regionData  )) assign(names(regionData)[i], regionData[[i]])
+  }
 
   ### Years
   maxYear    <- maxYear0
@@ -183,7 +191,7 @@ run_fredi <- function(
   ###### Aggregation level  ######
   ### Types of summarization to do: default
   doPrimary <- ifelse(is.null(doPrimary), FALSE, doPrimary)
-  aggList0     <- c("national", "modelaverage", "impactyear", "impacttype")
+  aggList0  <- c("national", "modelaverage", "impactyear", "impacttype")
   if(!is.null(aggLevels)){
     ### Aggregation levels
     aggLevels    <- aggLevels |> tolower()
@@ -200,8 +208,15 @@ run_fredi <- function(
 
   ###### Sectors List ######
   ### Sector names
-  sector_names  <- co_sectors$sector_id
-  sector_labels <- co_sectors$sector_label
+  ### Limit to just available sectors by state/region
+  if(byState){
+    sector_names  <- co_stateSectors$sector_id
+    sector_labels <- co_stateSectors$sector_label    
+  } else{
+    sector_names  <- co_sectors$sector_id
+    sector_labels <- co_sectors$sector_label    
+  }
+
   ### Initialize sector list if the sectors list is null
   if(is.null(sectorList)){
     sectorList       <- sector_names
@@ -239,7 +254,7 @@ run_fredi <- function(
   list_inputs     <- co_inputScenarioInfo$inputName
   num_inputNames  <- co_inputScenarioInfo |> nrow()
 
-  if(is.null(inputsList)) {inputsList <- list()}else{message("Checking input values...")}
+  if(is.null(inputsList)) {inputsList <- list()} else{message("Checking input values...")}
   ### Iterate over the input list
   # if(!is.null(inputsList)){
   ### Assign inputs to objects
@@ -258,7 +273,7 @@ run_fredi <- function(
[…]
