+++ FrEDI/R/run_fredi.R
@@ -273,7 +273,7 @@ run_fredi <- function(
     valueCol_i  <- inputInfo_i$valueCol |> unique()
     ### Initialize column names
     numCols_i   <- colNames_i <- c("year", valueCol_i) #; print(colNames_i)
-    ### Add region column ##wm need to be state if byState?
+    ### Add region column ##wm not used elsewhere so fine to leave/delete
     if(region_i == 1){
       colNames_i  <- c(colNames_i[1], "region", colNames_i[2])
     }
@@ -320,7 +320,7 @@ run_fredi <- function(
       minYear_x <- x$year |> min()
       interpYrs <- refYear_temp:maxYear
       ### Interpolate
-      x_interp  <- x |> interpolate_annual(#wm I think this is fine to leave with region, no byState = T for now unless we plan to allow state-level temps
+      x_interp  <- x |> interpolate_annual(#wm fine to leave with region, no byState = T for now unless we plan to allow state-level temps
         years  = interpYrs,
         column = "temp_C",
         rule   = 1:2
@@ -359,7 +359,7 @@ run_fredi <- function(
       minYear_x <- x$year |> min()
       interpYrs <- refYear_slr:maxYear
       ### Interpolate annual values
-      x_interp  <- x |> interpolate_annual(#wm same as temps above, I *think* this is fine to leave
+      x_interp  <- x |> interpolate_annual(#wm same as temps above, I think fine to leave
         years  = interpYrs,
         column = "slr_cm",
         rule   = 1:2
@@ -484,14 +484,20 @@ run_fredi <- function(
   ### Filter main scalars to correct years and filter out regional population
   ### Join with regPopScalar
   df_mainScalars <- df_mainScalars |> filter(year >= minYear) |> filter(year <= maxYear)
-  df_mainScalars <- df_mainScalars |> filter(scalarName!="reg_pop")
+  df_mainScalars <- df_mainScalars |> filter(scalarName!="reg_pop", scalarName!="state_pop")
   df_mainScalars <- df_mainScalars |> (function(df0, pop0 = pop_df, popCols = popCols0){
     ### Format population
     pop0 <- pop0 |> select(c(all_of(popCols)))
-    pop0 <- pop0 |> rename(value=reg_pop)
-    pop0 <- pop0 |> mutate(scalarName           = "reg_pop")
-    pop0 <- pop0 |> mutate(scalarType           = "physScalar")
-    pop0 <- pop0 |> mutate(national_or_regional = "regional")
+    if(byState){
+      pop0 <- pop0 |> rename(value=state_pop)
+      pop0 <- pop0 |> mutate(scalarName = "state_pop")
+      pop0 <- pop0 |> mutate(national_or_regional = "state")
+    } else{
+      pop0 <- pop0 |> rename(value=reg_pop)
+      pop0 <- pop0 |> mutate(scalarName = "reg_pop")
+      pop0 <- pop0 |> mutate(national_or_regional = "regional")
+    }
+    pop0 <- pop0 |> mutate(scalarType = "physScalar")
     ### Bind regional population with other scalars
     df0  <- df0  |> rbind(pop0)
     return(df0)
@@ -537,27 +543,38 @@ run_fredi <- function(
     ### Join with regional population
     npdScalars <- npdScalars |> (function(npd0, pop0 = pop_df, refYear0 = refYear){
       ### Columns
-      join0 <- c("year", "region", "reg_pop")
+      join0 <- c("year", "region")
+      if(byState){join0 <- c(join0, "state", "postal", "state_pop")} else{join0 <- c(join0, "reg_pop")}
       ### Separate population
       pop1  <- pop0 |> filter(year >= refYear0) |> select(c(all_of(join0)))
[…]
