### 2023-11-07 a3c49f4
Initial fixes
Updated:
sysdata.rda with fixes from FrEDI_Data
changes to main and helper functions to get code to work
