+++ FrEDI/R/aggregate_impacts.R
@@ -50,70 +50,80 @@ aggregate_impacts <- function(
     aggLevels   = c("national", "modelaverage", "impactyear", "impacttype"),  ### Levels of aggregation
     columns     = c("annual_impacts"), ### Columns to aggregate
     groupByCols = c("sector", "variant", "impactYear", "impactType", "model_type", "model", "region"), ### Columns to group by
-    mode = "all"
+    mode        = "all"
 ){
-
-
   ###### Defaults ######
   ### Not used currently; preserving it in messaging logicals for the future
-  msgUser <- T
+  msgUser    <- T
+
+  ###### Set Values ######
+  ###### - Args ######
+  null_groupByCols <- groupByCols |> is.null()
+  ###### - Modes ######
   ### Modes...specifically SLR interpolation only
-  xMode   <- ifelse(is.null(mode), "all", mode) |> tolower()
+  xMode      <- mode |> is.null() |> ifelse( "all", mode) |> tolower()
+  doSlr      <- "slrinterpolation" == xMode
+  ### Aggregation if xMode = "slrInterpolation"
+  if(doSlr){aggLevels <- "modelaverage"}
 
-  ###### Ungroup Data ######
-  data    <- data |> ungroup() #; names(data) |> print()
+  ####### By State  ######
+  byState    <- c("state") %in% groupByCols
+  if(byState){stateCols0 <- c("state", "postal")} else{stateCols0 <- c()}
+  popCol0    <- byState |> ifelse("state_pop", "reg_pop")
+  # byState |> print()
 
+  ###### Format Data ######
+  ### Ungroup
+  data       <- data |> ungroup() #; names(data) |> print()
+  # data |> glimpse()
   ###### Years Info ######
   ### Years in data
   # c_npdRefYear <- 2090
   c_dataYears  <- data$year |> unique()
   # has_plus2090vals <- (c_dataYears[which(c_dataYears > c_npdRefYear)] |> length()) > 0
 
-  ###### SLR Info ######
+  ###### Get FrEDI Data Objects ######
+  co_models   <- "co_models"   |> get_frediDataObj("frediData")
+  co_sectors  <- "co_sectors"  |> get_frediDataObj("frediData")
+  co_variants <- "co_variants" |> get_frediDataObj("frediData")
+
+  ###### - Formate SLR Info ######
   # assign("slr_cm", rDataList[["slr_cm"]])
-  assign("co_models", rDataList[["co_models"]])
+  # assign("co_models", rDataList[["co_models"]])
   co_slrs <- co_models |> filter(modelType=="slr")
-  co_slrs <- co_slrs   |>  mutate_at(.vars=c("model_dot", "model_label"), as.character)
-  co_slrs <- co_slrs   |> as.data.frame()
+  co_slrs <- co_slrs   |> mutate_at(.vars=c("model_dot", "model_label"), as.character)
+  # co_slrs <- co_slrs   |> as.data.frame()
 
-  ###### Load Sector Info ######
-  name_dfSectors   <- "co_sectors"
-  sectorCols_old   <- c("sector_id", "sector_label", "modelType")
-  sectorCols_new   <- c("sector_id", "sector", "model_type")
-
-  assign(name_dfSectors, rDataList[[name_dfSectors]])
-
-  co_sectors  <- co_sectors |> select(all_of(sectorCols_old))
-  co_sectors  <- co_sectors |> rename_with(~sectorCols_new, .cols=sectorCols_old)
-  # co_sectors  <- co_sectors |> (function(y){names(y) <- c(sectorCols_new); return(y)})
-  co_sectors  <- co_sectors |> mutate(model_type = model_type |> toupper())
+  ###### - Format Sector Info ######
[…]
