+++ FrEDI/R/utils.R
@@ -37,6 +37,42 @@ get_msgPrefix <- function(level=1){
   return(msg_x)
 }
 
+### This function makes it easier to get data objects from the sysdata.rda file
+get_frediDataObj <- function(
+    x        = NULL,       ### Object name
+    listSub  = "frediData", ### Sublist name
+    listall  = FALSE,
+    listName = "rDataList",
+    pkg      = "FrEDI",
+    lib.loc  = .libPaths()[1] ### Library path to look for packages
+){
+  ### Messaging
+  msg0    <- "\t"
+  ### Check if list name exists
+  exists0 <- listName |> exists()
+  ### If the listname exists in the name space, parse it
+  ### Otherwise, grab it from a package name space
+  if(exists0){new_x <- parse(text=listName) |> eval()}
+  else       {
+    ### Check if package & list name
+    pkgList0    <- lib.loc |> installed.packages()
+    pkgExists0  <- pkg %in% pkgList0
+    if(!pkgExists0){
+      msg0 |> paste0("Package doesn't exist...") |> message()
+      msg0 |> paste0("Exiting...") |> message()
+      return()
+    } ### End if(!pkgExists0)
+    else           {new_x <- getFromNamespace(listName, ns=pkg)}
+  } ### End else(exists0)
+
+  ### Whether to list all items in data object or not
+  # if(listall) {return_x <- new_x |> names()}
+  # else        {return_x <- new_x[[x]]}
+  if(listall) {return_x <- new_x |> names()}
+  else        {return_x <- new_x[[listSub]][["data"]][[x]]}
+  ### Return
+  return(return_x)
+} ### End get_frediDataObj
 
 ###### interpolate_annual ######
 ### Created 2021.02.05. Updated 2021.02.05
@@ -52,44 +88,46 @@ interpolate_annual <- function(
     byState   = FALSE ### If breakdown by state
 ){
   ###### Data Info ######
+  ##### By state
+  if (byState) {stateCols0 <- c("state", "postal")} else{stateCols0 <- c()}
   ##### Columns
   dataCols  <- data |> names()
-  defCols   <- c("year", "region")
-  if (byState) {defCols <- defCols |> c("state", "postal")}
+  defCols   <- c("year", "region") |> c(stateCols0)
   defCol0   <- dataCols[!(dataCols %in% defCols)][1]
-  column0   <- ifelse(is.null(column), defCol0, column)
+  column0   <- column |> is.null() |> ifelse(defCol0, column)
   othCols   <- dataCols[!(dataCols %in% c(defCols, column0))]
+  # defCol0 |> print(); column0 |> print()
   rm("defCol0")
   ###### Format data
   data      <- data |> filter(!is.na(column0))
-  values0   <- data |> get_vector(column0)
-  years0    <- data |> get_vector(defCols[1])
+  values0   <- data[[column0]]
+  # column0 |> print()
+  years0    <- data[["year" ]]
   ### Interpolation years
-  doYears   <- is.null(years)
+  doYears   <- years |> is.null()
   if(doYears){
-    years  <- years0 |> range(na.rm=TRUE);
+    years  <- years0 |> range(na.rm=TRUE)
     years  <- years[1]:years[2]
   }; rm("doYears")
   ##### Regions
-  regions0  <- region
+  region0   <- region
   addRegion <- !("region" %in% dataCols)
-  if(addRegion) {data      <- data |> mutate(region = regions0[1])}
-  else          {regions0  <- data |> get_uniqueValues("region")}
+  if (addRegion) {data     <- data |> mutate(region = region0)}
+  if (byState  ) {states0  <- data |> get_uniqueValues("state" )}
[…]

[…]
