### 2023-11-08 ffe3f32
Update temps2slr.R
Updated to use correct item from rDataList
