### 2023-11-08 07068a9
Update utils.R
Simplified some lines of code in `extend_slr`
