### 2023-11-08 b16031a
Update run_fredi.R
Changed some section headers to fix some nesting issues when viewing outline in RStudio
Cleaned up some comments
