+++ FrEDI/R/run_fredi.R
@@ -272,8 +272,8 @@ run_fredi <- function(
     popCol0      <- "reg_pop"
   } ### End byState
 
-
-  ###### Load Inputs ######
+  ###### Inputs ######
+  ###### ** Load Inputs ######
   ### Create logicals and initialize inputs list
   list_inputs     <- co_inputScenarioInfo$inputName
   num_inputNames  <- co_inputScenarioInfo |> nrow()
@@ -295,7 +295,7 @@ run_fredi <- function(
     ### Min and Max Values
     min_i       <- inputInfo_i$inputMin |> unique()
     max_i       <- inputInfo_i$inputMax |> unique()
-    ###### Column Info ######
+    ### Column Info
     region_i    <- inputInfo_i$region |> unique()
     valueCol_i  <- inputInfo_i$valueCol |> unique()
     ### Initialize column names
@@ -308,7 +308,7 @@ run_fredi <- function(
     # has_update_i <- is.null(inputsList[[inputName_i]])
     df_input_i   <- inputsList[[inputName_i]]
     has_update_i <- !is.null(df_input_i)
-    ###### Assign inputs to objects ######
+    ### Assign inputs to objects
     assign(has_i,    has_update_i)
     assign(inputName_i, df_input_i)
   } ### End iterate over inputs
@@ -324,7 +324,7 @@ run_fredi <- function(
   returnList[["arguments"]] <- argsList
   rm("argsList")
 
-  ###### Temperature Scenario ######
+  ###### ** Temperature Scenario ######
   ### User inputs: temperatures have already been converted to CONUS temperatures. Filter to desired range.
   ### Name the reference year temperature
   ### Add the point where impacts are zero
@@ -366,7 +366,7 @@ run_fredi <- function(
   } ### End else(has_tempUpdate)
   # temp_df |> nrow() |> print()
 
-  ###### SLR Scenario ######
+  ###### ** SLR Scenario ######
   ### Year where SLR impacts are zero
   refYear_slr <- (co_modelTypes |> filter(modelUnitType=="slr"))$modelRefYear |> unique()
   # co_modelTypes |> names() |> print()
@@ -407,7 +407,7 @@ run_fredi <- function(
   # slr_df |> head() |> print()
   # slr_df$year |> range() |> print()
 
-  ###### Driver Scenario  ######
+  ###### ** Driver Scenario  ######
   ### Format the temperature and SLR values
   temp_df <- temp_df |> select(c("year", "temp_C_conus")) |>
     rename(modelUnitValue = temp_C_conus) |> mutate(modelType="gcm")
@@ -430,7 +430,7 @@ run_fredi <- function(
   ### Remove intermediate values
   rm("co_modelType0", "temp_df", "slr_df")
 
-  ###### Socioeconomic Scenario ######
+  ###### ** Socioeconomic Scenario ######
   ### Update the socioeconomic scenario with any GDP or Population inputs and message the user
   ### Reformat GDP inputs if provided, or use defaults
   gdpCols0 <- c("year", "gdp_usd")
@@ -486,7 +486,8 @@ run_fredi <- function(
   join0             <- "year"
   arrange0          <- "region" |> c(stateCols0) |> c(join0)
   updatedScenario   <- national_scenario |> left_join(pop_df, by=join0)
-  updatedScenario   <- updatedScenario   |> arrange_at(.vars=c(arrange0))
+  updatedScenario   <- updatedScenario   |> arrange_at(c(arrange0))
+  # updatedScenario   <- updatedScenario   |> mutate_at (c("region"), function(x){gsub(" ", ".", x)})
   rm(join0, arrange0)
[…]
