### 2023-11-08 628cf75
Update aggregate_impacts.R
Cleaned up crufty code, fixed issues with aggregating national totals
