+++ FrEDI/R/aggregate_impacts.R
@@ -49,81 +49,76 @@ aggregate_impacts <- function(
     data,             ### Data frame of outputs from temperature binning
     aggLevels   = c("national", "modelaverage", "impactyear", "impacttype"),  ### Levels of aggregation
     columns     = c("annual_impacts"), ### Columns to aggregate
-    groupByCols = c("sector", "variant", "impactYear", "impactType", "model_type", "model", "region"), ### Columns to group by
-    mode        = "all"
+    groupByCols = c("sector", "variant", "impactYear", "impactType", "model_type", "model", "region")
 ){
   ###### Defaults ######
   ### Not used currently; preserving it in messaging logicals for the future
-  msgUser    <- T
+  msgUser        <- T
 
   ###### Set Values ######
   ###### - Args ######
-  null_groupByCols <- groupByCols |> is.null()
-  ###### - Modes ######
-  ### Modes...specifically SLR interpolation only
-  xMode      <- mode |> is.null() |> ifelse( "all", mode) |> tolower()
-  doSlr      <- "slrinterpolation" == xMode
-  ### Aggregation if xMode = "slrInterpolation"
-  if(doSlr){aggLevels <- "modelaverage"}
+  null_groupCols <- groupByCols |> is.null()
 
   ####### By State  ######
-  byState    <- c("state") %in% groupByCols
+  byState        <- c("state") %in% groupByCols
   if(byState){stateCols0 <- c("state", "postal")} else{stateCols0 <- c()}
-  popCol0    <- byState |> ifelse("state_pop", "reg_pop")
+  popCol0        <- byState |> ifelse("state_pop", "reg_pop")
   # byState |> print()
 
   ###### Format Data ######
   ### Ungroup
-  data       <- data |> ungroup() #; names(data) |> print()
+  data           <- data |> ungroup() #; names(data) |> print()
   # data |> glimpse()
   ###### Years Info ######
   ### Years in data
   # c_npdRefYear <- 2090
-  c_dataYears  <- data$year |> unique()
+  c_dataYears    <- data$year |> unique()
   # has_plus2090vals <- (c_dataYears[which(c_dataYears > c_npdRefYear)] |> length()) > 0
 
   ###### Get FrEDI Data Objects ######
-  co_models   <- "co_models"   |> get_frediDataObj("frediData")
-  co_sectors  <- "co_sectors"  |> get_frediDataObj("frediData")
-  co_variants <- "co_variants" |> get_frediDataObj("frediData")
+  co_models      <- "co_models"   |> get_frediDataObj("frediData")
+  co_sectors     <- "co_sectors"  |> get_frediDataObj("frediData")
+  co_variants    <- "co_variants" |> get_frediDataObj("frediData")
 
   ###### - Formate SLR Info ######
   # assign("slr_cm", rDataList[["slr_cm"]])
   # assign("co_models", rDataList[["co_models"]])
-  co_slrs <- co_models |> filter(modelType=="slr")
-  co_slrs <- co_slrs   |> mutate_at(.vars=c("model_dot", "model_label"), as.character)
+  co_slrs        <- co_models |> filter(modelType=="slr")
+  co_slrs        <- co_slrs   |> mutate_at(.vars=c("model_dot", "model_label"), as.character)
   # co_slrs <- co_slrs   |> as.data.frame()
 
   ###### - Format Sector Info ######
-  sectorCols_old <- c("sector_id", "sector_label", "modelType")
-  sectorCols_new <- c("sector_id", "sector", "model_type")
-  co_sectors     <- co_sectors |> select(all_of(sectorCols_old))
-  co_sectors     <- co_sectors |> rename_at(c(sectorCols_old), ~sectorCols_new)
[…]
