### 2023-11-09 324d285
Updates to combine region & state-level data
Updated sysdata.rda with combined region & state-level data
Edited fun_formatScalars to determine whether to run a particular scalar as byState or not depending on the level of data provided
Edited run_fredi to run for all sectors. Also cleaned up code
