+++ FrEDI/R/run_fredi.R
@@ -150,8 +150,6 @@ run_fredi <- function(
     byState    = FALSE, ### Whether to run at state level for available sectors
     silent     = TRUE   ### Whether to message the user
 ){
-
-
   ###### Set up the environment ######
   ### Level of messaging (default is to message the user)
   silent    <- silent |> is.null() |> ifelse(T, silent)
@@ -201,63 +199,59 @@ run_fredi <- function(
   ### Types of summarization to do: default
   doPrimary <- doPrimary |> is.null() |> ifelse(FALSE, doPrimary)
   aggList0  <- c("national", "modelaverage", "impactyear", "impacttype")
-  if(!(aggLevels |> is.null())){
+  if(aggLevels |> is.null()){
+    aggLevels <- aggList0
+  } else{
     ### Aggregation levels
     aggLevels    <- aggLevels |> tolower()
     aggLevels    <- aggLevels[which(aggLevels %in% c(aggList0, "all", "none"))]
     doAgg        <- "none" %in% aggLevels
     ### If none specified, no aggregation (only SLR interpolation)
     ### Otherwise, aggregation depends on length of agg levels
-    if     ("none" %in% aggLevels) {
-      aggLevels   <- c()
-    } else if("all"  %in% aggLevels) {
-      aggLevels   <- aggList0
-    }
-    else                           {
-      requiresAgg <- aggLevels |> length() > 0
-    }
-  } else{
-    aggLevels <- aggList0
+    if     ("none" %in% aggLevels) {aggLevels   <- c()}
+    else if("all"  %in% aggLevels) {aggLevels   <- aggList0}
+    else                           {requiresAgg <- aggLevels |> length() > 0}
   } ### End else(!is.null(aggLevels))
   requiresAgg <- aggLevels |> length() > 0
 
   ###### Sectors List ######
   ### Sector names
   ### Limit to just available sectors by state/region
-  if(byState){
-    sector_names  <- co_stateSectors$sector_id
-    sector_labels <- co_stateSectors$sector_label
-  } else{
-    sector_names  <- co_sectors$sector_id
-    sector_labels <- co_sectors$sector_label
-  }
+  # if(byState){
+  #   sector_names  <- co_stateSectors$sector_id
+  #   sector_labels <- co_stateSectors$sector_label
+  # } else{
+  #   sector_names  <- co_sectors$sector_id
+  #   sector_labels <- co_sectors$sector_label
+  # }
+  sector_names  <- co_sectors$sector_id
+  sector_labels <- co_sectors$sector_label
+  # sector_names |> print()
 
   ### Initialize sector list if the sectors list is null
   if(sectorList |> is.null()){
-    sectorList       <- sector_names
+    sectorList    <- sector_names
   } else{
     ### Compare inputs to names and labels in the data
     ### Subset to sector list in sector names
-    which_sectors    <- which(
-      (tolower(sector_names) %in% tolower(sectorList)) |
-        (tolower(sector_labels) %in% tolower(sectorList))
-    )
-    sectorList       <- sector_names[which_sectors]
+    sectors0   <- sector_names  |> tolower()
+    sectors1   <- sector_labels |> tolower()
+    sectors2   <- sectorList    |> tolower()
+    which0     <- (sectors0 %in% sectors2) | (sectors1 %in% sectors2)
+    sectorList <- sector_names[which0]
     ### Message users about missing sectors
-    which_notSectors <- which(!(
-      (tolower(sectorList) %in% tolower(sector_names)) |
[…]
