### 2023-11-15 bc9c9bc
Add state data import functionality to import_inputs
