### 2023-11-16 7152705
Initialized fixes to combined data code
Initialized some fixes to the combined data branch
These fixes fix some things and break others (national level scalars not processing correctly...not sure if it is on this side or the FrEDI code side)
