+++ FrEDI/R/utils.R
@@ -4,7 +4,7 @@ get_vector <- function(
     data, column = NULL
 ){
   ### Select column and get values as vector
-  col0  <- ifelse(is.null(column), c(), column)
+  col0  <- column |> is.null() |> ifelse(c(), column)
   vals0 <- data[[column]] |> as.vector()
   ### Return
   return(vals0)
@@ -89,7 +89,8 @@ interpolate_annual <- function(
 ){
   ###### Data Info ######
   ##### Other values
-  region0   <- region; rm(region)
+  region0   <- region
+  rm(region)
   ##### By state
   if (byState) {stateCols0 <- c("state", "postal")} else{stateCols0 <- c()}
   ##### Columns
@@ -99,25 +100,26 @@ interpolate_annual <- function(
   column0   <- column |> is.null() |> ifelse(defCol0, column)
   othCols   <- dataCols[!(dataCols %in% c(defCols, column0))]
   # defCol0 |> print(); column0 |> print()
-  rm("defCol0")
+  rm(defCol0)
+
   ###### Format data
-  data      <- data |> filter(!is.na(column0))
-  values0   <- data[[column0]]
   # column0 |> print()
+  data      <- data |> filter(!(column0 |> is.na()))
+  values0   <- data[[column0]]
   years0    <- data[["year" ]]
+
   ### Interpolation years
   doYears   <- years |> is.null()
   if(doYears){
     years  <- years0 |> range(na.rm=TRUE)
     years  <- years[1]:years[2]
-  }; rm("doYears")
+  } ### End if(doYears)
+  rm(doYears)
+
   ##### Regions
   addRegion <- !("region" %in% dataCols)
-  if (addRegion) {data     <- data |> mutate(region = region0)}
-  if (byState  ) {states0  <- data |> get_uniqueValues("state" )}
-  else           {regions0 <- data |> get_uniqueValues("region")}
-  rm("addRegion")
-
+  if (addRegion) {data <- data |> mutate(region = region0)}
+  rm(addRegion)
 
   ###### Interpolation Info ######
   ### - Return NA values outside of extremes
@@ -130,27 +132,14 @@ interpolate_annual <- function(
   if(repRule ){rule <- rule |> rep(2)}
   method    <- method |> is.null() |> ifelse("linear", method)
 
-  ###### Interpolate missing values for each region ######
+  ###### Interpolate NA values ######
   ### Filter to the region and then interpolate missing values
-  cols0     <- c("x", "y");
-  cols1     <- c("year", column0)
-
-  if (!byState) {
-    df_interp <- regions0 |> map(function(region_i){
-      ### Values
-      df_i     <- data |> filter(region==region_i)
-      x_i      <- df_i[["year" ]]
-      y_i      <- df_i[[column0]]
-      ### Approximate
-      new_i   <- approx(x=x_i, y=y_i, xout=years, rule=rule, method=method)
-      new_i   <- new_i |> as_tibble()
-      new_i   <- new_i |> rename_at(.vars=c(cols0), ~cols1)
-      new_i   <- new_i |> mutate(region = region_i)
-      # new_i |> names() |> print()
-      ### Return
-      return(new_i)
-    }) |> bind_rows()
-  } else { ### By state
+  cols0     <- c("x", "y")
+  cols1     <- c("year") |> c(column0)
+  ### Iterate over states if byState=T
+  ### Otherwise, iterate over regions
+  if (byState) {
+    states0   <- data |> get_uniqueValues("state" )
     df_interp <- states0 |> map(function(state_i){
       ### Values
       df_i     <- data |> filter(state==state_i)
@@ -159,8 +148,24 @@ interpolate_annual <- function(
       ### Approximate
[…]
