### 2023-11-29 344759c
fix to duplicate issue and some testing
