### 2023-12-01 1c5e3f7
fix econ adj value columns
