### 2023-12-05 fa919d2
Update utils.R
Changed function "match_scalarValues" in "utils.R":
- Changed join columns for national values (removed "byState" column from join columns). The previous join columns were causing national values to go missing.
- Removed step converting missing values to one (which was hiding the fact that the national values were missing)
