### 2023-12-05 dd4bef8
Update utils.R
Fixed remaining issues in calculating scalars:
Edits to match_scalarValues:
Corrected issue in joining national-level scalars

Edits to extend_slrScalars:
- Simplified workflow to make better use of physical scalars
- Corrected column names and join columns

Edits to initialize_resultsDf:
- Added additional conditional so SLR scalars only extended if the analysis period runs to years > 2090
