+++ FrEDI/R/utils.R
@@ -332,15 +332,16 @@ update_popScalars <- function(
 ###### extend_slrScalars ######
 ### Scalars for SLR past 2090
 extend_slrScalars <- function(
+    df0,        ### Dataframe of initial results
     df_se,      ### Socioeconomic scenario
     df_scalars, ### Main scalar values: df_mainScalars
-    # df_info,    ### SLR scalar info: co_slrScalars
     elasticity  = NULL
 ){
   ###### Elasticity ######
   ### Update exponent
   if(!(elasticity |> is.null())){
-    df_scalars <- df_scalars |> mutate(exp0 = (exp0 == 1) |> ifelse(exp0, elasticity))
+    # df_scalars <- df_scalars |> mutate(exp0 = (exp0 == 1) |> ifelse(exp0, elasticity))
+    df_scalars <- df_scalars |> mutate(exp0 = (econScalarName=="vsl_usd") |> ifelse(elasticity, exp0))
   } ### End if(!(elasticity |> is.null()))
 
   ###### By State ######
@@ -353,135 +354,126 @@ extend_slrScalars <- function(
   ### Filter to years >= refYear (2090)
   df_info    <- "co_slrScalars" |> get_frediDataObj("frediData")
   refYear0   <- (df_info[["refYear"]] |> unique())[1]
-  df_se      <- df_se      |> filter(year >= refYear0)
   df_scalars <- df_scalars |> filter(year >= refYear0)
+  df_se      <- df_se      |> filter(year >= refYear0)
+  df0        <- df0        |> filter(year >= refYear0)
+  # df_se      <- df_se      |> filter(year >= refYear0)
 
-  ###### Gather Multipliers ######
+  ### Drop columns from data
+  drop0      <- c("physScalar", "physAdj", "damageAdj", "econScalar", "econMultiplier", "econAdj")
+  drop1      <- drop0 |> paste0("Name") |> c(drop0 |> paste0("Value"))
+  drop1      <- drop1 |> c("physScalar", "econScalar", "econMultiplier", "physEconScalar")
+  drop1      <- drop1 |> c("c1", "exp0")
+  df0        <- df0   |> select(-all_of(drop1))
+  rm(drop0, drop1)
+
+  ###### Join Data & Scalar Info ######
+  ### Join initial results & scalar info
+  drop0      <- c("byState")
+  join0      <- c("sector")
+  df_info    <- df_info |> select(-all_of(drop0))
+  df0        <- df0     |> left_join(df_info, by=c(join0))
+  rm(drop0, join0)
+
+  ###### Set Default Columns ######
+  ### Make some columns none and 1
+  mutate0    <- c("damageAdj", "econScalar") |> paste0("Name")
+  mutate1    <- c("damageAdj", "econScalar") |> paste0("Value")
+  df0[,mutate0] <- "none"
+  df0[,mutate1] <- 1
+  rm(mutate0, mutate1)
+
+  ###### Get Multiplier Values ######
   ### Gather econMultiplier scalars
   gather0    <- c("gdp_usd", "gdp_percap")
-  select0    <- gather0 |> c("byState", "year")
-  df_gdp0    <- df_se   |> summarize_seScenario(national=T)
-  df_gdp0    <- df_gdp0 |> select(all_of(select0))
-  df_gdp0    <- df_gdp0 |> pivot_longer(
+  select0    <- gather0 |> c("year")
+  # select0    <- gather0 |> c("byState", "year")
+  df_mult0   <- df_se   |> summarize_seScenario(national=T)
+  df_mult0   <- df_mult0 |> filter(byState==0)
+  df_mult0   <- df_mult0 |> select(all_of(select0))
+  df_mult0   <- df_mult0 |> pivot_longer(
     cols      = all_of(gather0),
[…]
