### 2023-12-14 685eda5
make state_pop NA during National level aggregation
