### 2023-12-14 ae5cc01
make physcialmeasure NA during National level aggregation
