### 2023-12-15 5505849
seperate drop column conditionals from msg consitionals
