### 2023-12-15 1c5a725
fix aggregation for national using existing column strings
