+++ FrEDI/R/aggregate_impacts.R
@@ -509,16 +509,15 @@ aggregate_impacts <- function(
     ### Ungroup first
     df_agg      <- df_agg |> ungroup()
     ### Grouping columns
-    group0      <- groupByCols |> (function(y){y[!(y %in% c("region", stateCols0, yearCol0))]})()
+    group0      <- groupByCols |> (function(y){y[!(y %in% c("region", stateCols0,popCol0, yearCol0))]})()
     group0      <- group0 |> c("year")
     ### Calculate number of non missing values
     df_national <- df_agg |> (function(w){
       w |> mutate(not_isNA = 1 * (!(w[[summaryCol1]] |> is.na())))
     })()
     ### Group data, sum data, calculate averages, and drop NA column
-    sum0        <- summaryCols |> c("not_isNA")
+    sum0        <- summaryCols |> c(popCol0,"not_isNA")
     df_national <- df_national |>
-      mutate(state_pop = "N/A") |>
       group_by_at(c(group0)) |>
       summarize_at(vars(sum0), sum, na.rm=T) |> ungroup()
     rm(sum0)
@@ -531,13 +530,16 @@ aggregate_impacts <- function(
     ### Drop columns, adjust values
     df_national <- df_national |> select(-c("not_isNA"))
     df_national <- df_national |> mutate(region="National Total")
+    ### Join with National Pop
+    #join0       <- natPopCols |> (function(y){y[!(y %in% c("national_pop"))]})()
+    #df_national <- df_national |> left_join(nationalPop, by = c(join0))
     if(byState){
       df_national   <- df_national |> mutate(state ="All")
       df_national   <- df_national |> mutate(postal="US")
     } ### End if(byState)
 
     ### Add back into regional values and bind national population to impact types
-    # df_agg |> glimpse(); df_national |> glimpse()
+    #df_agg |> glimpse(); df_national |> glimpse()
     df_agg      <- df_agg |> rbind(df_national);
 
     ### Add national to total populations
