### 2023-12-18 097967c
add dome prints statements
