### 2023-12-18 1ae0855
add dome print/list statements
