+++ .github/workflows/test_fredi.yml
@@ -98,9 +98,9 @@ jobs:
             ### Check if path exists and, if not, create it
             exists0 <- oPath1 |> dir.exists()
             print(oPath1)
-            prints(exists0)
+            print(exists0)
             if(!exists0){oPath1 |> dir.create()}
-
+            list.files(oPath1)
             ### Run FrEDI
             results0 <- run_fredi()
             ### Run tests
@@ -128,7 +128,9 @@ jobs:
             exists0 <- oPath2 |> dir.exists()
             if(!exists0){oPath2 |> dir.create()}
             print(oPath2)
-            prints(exists0)
+            print(exists0)
+            list.files(oPath2)
+
             ### Create report figures
             reports0 <- create_DoW_results(
               outPath  = oPath2,
