### 2023-12-18 f6aeeba
Fixed default groups in aggregate_impacts and started adding state level to report figures
