+++ FrEDI/R/utils_create_report_figures.R
@@ -127,18 +127,25 @@ filter_years <- function(
   df0 <- df0 |> filter(year %in% years)
   ### Return
   return(df0)
-} ### End filter_years
+} ### End filter_years()
 
 ### Filter values
 ### Format values to specific number of decimal places
 format_values <- function(
     df0, ### data
-    cols0  = c("driverValue", "gdp_usd", "national_pop", "gdp_percap", "reg_pop", "annual_impacts"), ### Columns to format
+    byState = TRUE,
+    # cols0  = c("driverValue", "gdp_usd", "national_pop", "gdp_percap", "reg_pop", "annual_impacts"), ### Columns to format
     digits = 16
 ){
-  df0 <- df0 |> mutate_at(.vars=c(cols0), function(x){format(x, digits=digits)})
+  ### Pop columns
+  if(byState){popCols <- c("state", "postal")} else{c()}
+  popCol <- byState |> ifelse("state_pop", "reg_pop")
+  ### Columns
+  cols0 <- c("driverValue", "gdp_usd", "national_pop", "gdp_percap", popCol, "annual_impacts")
+  ### Mutate
+  df0 <- df0 |> mutate_at(vars(cols0), function(x){format(x, digits=digits)})
   return(df0)
-}
+} ### End format_values()
 
 ### Run CONUS scenarios
 create_constant_temp_scenario <- function(
@@ -189,7 +196,8 @@ create_constant_temp_scenario <- function(
 #### Get scenario inputs
 #### Get inputs list for a single scenario
 get_scenario_inputsList <- function(
-    df0   ### Data
+    df0,   ### Data
+    byState = TRUE
 ){
   ### df0 names
   names0  <- df0 |> names()
@@ -199,12 +207,15 @@ get_scenario_inputsList <- function(
   slr0    <- NULL
   gdp0    <- NULL
   pop0    <- NULL
+  ### Pop columns
+  if(byState){popCols <- c("state", "postal")} else{c()}
+  popCol <- byState |> ifelse("state_pop", "reg_pop")
   ### Columns for scenarios
   cTemp0  <- c("year", "temp_C_conus")
   cTemp1  <- c("year", "temp_C")
   cSlr    <- c("year", "slr_cm")
   cGdp    <- c("year", "gdp_usd")
-  cPop    <- c("year", "region", "reg_pop")
+  cPop    <- c("year", "region") |> c(popCols, popCol)
   ### Whether to create scenarios
   doTemp0 <- (cTemp0 %in% names0) |> all()
   doTemp1 <- (cTemp1 %in% names0) |> all()
@@ -217,28 +228,28 @@ get_scenario_inputsList <- function(
   if(doTemp){
     if(doTemp0){cTemp <- cTemp0}
     else       {cTemp <- cTemp1}
-    temp0   <- df0 |> select(c(all_of(cTemp)))
+    temp0   <- df0 |> select(all_of(cTemp))
     if(doTemp0){
-      temp0 <- temp0 |> rename_at(.vars=c("temp_C_conus"), ~c("temp_C"))
-    }
+      temp0 <- temp0 |> rename_at(vars("temp_C_conus"), ~c("temp_C"))
+    } ### End if(doTemp0)
     list0[["tempInput"]] <- temp0
     rm("temp0")
   } ### End if(doTemp)
 
   if(doSlr){
-    slr0   <- df0 |> select(c(all_of(cSlr)))
+    slr0   <- df0 |> select(all_of(cSlr))
     list0[["slrInput"]] <- slr0
     rm("slr0")
   } ### End if(doSlr)
 
   if(doGdp){
-    gdp0   <- df0 |> select(c(all_of(cGdp)))
[…]
