### 2023-12-18 95b649f
Update utils_create_report_figures.R
Fixed references to deprecated functions
