### 2023-12-18 52cbe77
Update create_DoW_results.R
Made changes to run by state. Commented out plots that aren't working yet
