+++ FrEDI/scripts/create_DoW_results.R
@@ -9,18 +9,32 @@ require(ggpubr)
 ###### create_DoW_results ######
 create_DoW_results <- function(
     sectors  = FrEDI::get_sectorInfo(), ### Which sectors
-    gcmYears = c(2090), ### Which years to report on for GCM sectors
+    gcmYears = c(2090),   ### Which years to report on for GCM sectors
     slrYears = c(2050, 2090), ### Which years to report on for SLR sectors
-    silent   = TRUE,  ### Degree of messaging
-    testing  = FALSE, ### Whether to print out extra diagnostic values
-    aggOnly  = TRUE, ### Whether to only include sectors for which "includeaggregate==1" in Fig 7 plots
+    byState  = TRUE       ### Whether values are by state or just by region
+    silent   = TRUE,      ### Degree of messaging
+    testing  = FALSE,     ### Whether to print out extra diagnostic values
+    aggOnly  = TRUE,      ### Whether to only include sectors for which "includeaggregate==1" in Fig 7 plots
     loadCode = "project", ### Whether to load code as source or devtools
-    fpath    = "." ,  ### Path to main FrEDI directory to load code from if loadCode == "project" or loadCode == "package"
-    saveFile = FALSE, ### Save file
+    fpath    = "." ,      ### Path to main FrEDI directory to load code from if loadCode == "project" or loadCode == "package"
+    saveFile = FALSE,     ### Save file
     outPath  = "." |> file.path("report_figures"),  ### Path to save results if saveFile == TRUE
-    img_dev  = "pdf", ### Image device if saveFile == TRUE
-    return   = TRUE   ### Whether to return list object
+    img_dev  = "pdf",     ### Image device if saveFile == TRUE
+    return   = TRUE       ### Whether to return list object
 ){
+  # sectors  = FrEDI::get_sectorInfo() ### Which sectors
+  # gcmYears = c(2090) ### Which years to report on for GCM sectors
+  # slrYears = c(2050, 2090) ### Which years to report on for SLR sectors
+  # silent   = TRUE    ### Degree of messaging
+  # testing  = TRUE    ### Whether to print out extra diagnostic values
+  # byState  = TRUE    ### Whether values are by state or just by region
+  # aggOnly  = TRUE    ### Whether to only include sectors for which "includeaggregate==1" in Fig 7 plots
+  # loadCode = "project" ### Whether to load code as source or devtools
+  # fpath    = "."     ### Path to main FrEDI directory to load code from if loadCode == "project" or loadCode == "package"
+  # saveFile = FALSE   ### Save file
+  # outPath  = "." |> file.path("report_figures")  ### Path to save results if saveFile == TRUE
+  # img_dev  = "pdf"   ### Image device if saveFile == TRUE
+  # return   = TRUE    ### Whether to return list object
   ###### Initial values ######
   ### Messaging
   do_msg          <- !silent
@@ -100,7 +114,9 @@ create_DoW_results <- function(
   ###### ** Constants ######
   ### Numeric columns: Specify so that we can print out the associated data
   ### Number of digits to format
-  c_numVars      <- c("driverValue", "gdp_usd", "national_pop", "gdp_percap", "reg_pop", "annual_impacts")
[…]
