### 2023-12-18 dff8818
Update create_DoW_results.R
Fixed comma in function definition
