+++ .github/workflows/test_fredi.yml
@@ -94,6 +94,11 @@ jobs:
 
           ### Test results
           if(do_tests){
+
+            ### Check if path exists and, if not, create it
+            exists0 <- oPath1 |> dir.exists()
+            if(!exists0){oPath1 |> dir.create(recursive = TRUE)}
+
             ### Run FrEDI
             results0 <- run_fredi()
             ### Run tests
@@ -119,7 +124,7 @@ jobs:
 
             ### Check if path exists and, if not, create it
             exists0 <- oPath2 |> dir.exists()
-            if(!exists0){oPath2 |> dir.create()}
+            if(!exists0){oPath2 |> dir.create(recursive = TRUE)}
 
             ### Create report figures
             reports0 <- create_DoW_results(
