### 2023-12-19 ae5cef4
Update run_fredi.R
Changed default elasticity from 0.4 to 1
