### 2023-12-19 2036a49
Update utils_plot_DOW_bySector.R
Added fix to ggplot call
