### 2023-12-19 ec4e499
Update utils_plot_DOW_bySector.R
Added fix to aesthetics on dashed lines feature in DoW plots
