+++ FrEDI/R/utils_plot_DOW_bySector.R
@@ -125,10 +125,10 @@ plot_DOW_bySector <- function(
     plot0  <- plot0 + geom_point(data = df0_2, aes(x = .data[[xCol]], y = .data[[yCol]], color=.data[["model"]], shape = .data[["model"]]), alpha=0.65)
   } else{
     ### Initialize plot
-    plot0  <- df0 |> ggplot(x=.data[[xCol]], y=.data[[yCol]], color=.data[["model"]])
+    plot0  <- df0 |> ggplot(aes(x=.data[[xCol]], y=.data[[yCol]], color=.data[["model"]]))
     ### Add geoms
-    plot0  <- plot0 + geom_line (alpha=0.65)
-    plot0  <- plot0 + geom_point(aes(shape = model), alpha=0.65)
+    plot0  <- plot0 + geom_line (aes(x=.data[[xCol]], y=.data[[yCol]], color=.data[["model"]]), alpha=0.65)
+    plot0  <- plot0 + geom_point(aes(x=.data[[xCol]], y=.data[[yCol]], color=.data[["model"]], shape = model), alpha=0.65)
   } ### End if(do_gcm)
 
   ### Add Scales
