### 2024-01-05 149e80b
Update utils.R
Moved step for updating elasticity with a custom elasticity.
