### 2024-01-10 e824453
Update utils_sv.R
Updated use of "all_of"
