+++ FrEDI/R/utils_sv.R
@@ -51,40 +51,40 @@
 ###
 ###
 get_sv_sectorInfo <- function(
-    description=F,
-    gcmOnly=F,
-    slrOnly=F
+    description = F,
+    gcmOnly     = F,
+    slrOnly     = F
 ){
-  if(is.null(description)){description<-F}
-  if(is.null(gcmOnly    )){gcmOnly    <-F}
-  if(is.null(slrOnly    )){slrOnly    <-F}
+  if(description |> is.null()){description <- F}
+  if(gcmOnly     |> is.null()){gcmOnly     <- F}
+  if(slrOnly     |> is.null()){slrOnly     <- F}
   #svDataList <- load(file = "C:/Users/wmaddock/Documents/GitHub/FrEDI/createSystemData/data/sv/svDataList.rda")
   # co_sectorsRef$sector_label
   cData <- c("sectorInfo", "svSectorInfo")
   for(item_i in cData){assign(item_i, svDataList[[item_i]])}
   # print(sectorInfo)
-
   select0 <- c("sector", "modelType", "impactUnit")
   select1 <- c("sector", "modelType", "driverUnit", "variants", "impactUnit")
+
   ### - Select appropriate columns
   ### - Format modelType as uppercase
   ### - Add variant label
   ### - Add driver value
-  sectorInfo <- sectorInfo |> select(c(all_of(select0)))
+  sectorInfo <- sectorInfo |> select(all_of(select0))
   sectorInfo <- sectorInfo |> mutate(modelType  = modelType |> toupper())
   sectorInfo <- sectorInfo |> mutate(
-    variants = sector |> lapply(function(sector_i){
+    variants = sector |> map(function(sector_i){
       variants_i <- (svSectorInfo |> filter(sector==sector_i))$variant_label |> paste(collapse=", ")
       return(variants_i)
     }) |> unlist())
   sectorInfo <- sectorInfo |> mutate(driverUnit=ifelse(modelType=="GCM", "degrees Celsius", "cm"))
   ### Select and arrange
-  sectorInfo <- sectorInfo |> select(c(all_of(select1)))
-  sectorInfo <- sectorInfo |> arrange_at(.vars=c("sector"))
+  sectorInfo <- sectorInfo |> select(all_of(select1))
+  sectorInfo <- sectorInfo |> arrange_at(vars("sector"))
   ### GCM or SLR
   doFilter   <- (gcmOnly | slrOnly)
   if(doFilter){
-    cFilter    <- ifelse(gcmOnly, TRUE, FALSE)
+    cFilter    <- gcmOnly
     gcm_string <- "GCM"
     sectorInfo <- sectorInfo |> mutate(is_gcm =  modelType==gcm_string)
     sectorInfo <- sectorInfo |> filter(is_gcm == cFilter)
@@ -93,7 +93,7 @@ get_sv_sectorInfo <- function(
 
   ### If not description, return names only
   if(!description) {return_obj <- sectorInfo$sector}
-  else             {return_obj <- sectorInfo |> as.data.frame()}
+  else             {return_obj <- sectorInfo       }
   ### Return
   return(return_obj)
 }
@@ -103,7 +103,6 @@ get_sv_sectorInfo <- function(
 ### Created 2022.02.14. Last updated 2022.02.14
 ### This function attempts to load a user-specified input file
 ### Use this function to calculate tract level impacts
-
 calc_countyPop <- function(
     regPop,  ### Dataframe of population projection
     funList, ### Dataframe of population projections
@@ -112,7 +111,7 @@ calc_countyPop <- function(
 ){
   c_regions <- regPop$region |> unique()
[…]
