### 2024-01-10 48676be
Update run_fredi_sv.R
Updated use of "all_of", converted to use ofbind_rows instead of do.call(rbind, x)
