+++ FrEDI/R/run_fredi_sv.R
@@ -97,17 +97,13 @@ run_fredi_sv <- function(
     driverInput = NULL,
     popInput    = NULL,
     silent      = TRUE,  ### Whether to message the user
-    save        = FALSE,
-    outpath     = getwd(),
-    overwrite   = FALSE,
-    addDate     = FALSE,
     .testing    = FALSE
     # addDate     = FALSE, ### Whether to add the date to the file name
     # libPath     = .libPaths()[1]
 ){
   ###### Set up the environment ######
   pkgPath     <- NULL
-  pkgPath     <- ifelse(is.null(pkgPath), system.file(package="FrEDI"), pkgPath);
+  pkgPath     <- (pkgPath |> is.null()) |> ifelse(system.file(package="FrEDI"), pkgPath);
   rDataType   <- "rds"
   #pkgPath |> print()
   impactsPath <- pkgPath |> file.path("extdata", "sv", "impactLists")
@@ -118,13 +114,15 @@ run_fredi_sv <- function(
   ### Group types
   c_svGroupTypes <- svDataList$c_svGroupTypes
   ### Update years,
-  minYear <- 2010; maxYear <- 2090; list_years_by5 <- seq(minYear, maxYear, by=5)
+  minYear <- 2010
+  maxYear <- 2090
+  list_years_by5 <- seq(minYear, maxYear, by=5)
 
   ### Testing
-  save    <- ifelse(.testing, FALSE, save)
+  save    <- .testing |> ifelse(FALSE, save)
 
   ### Level of messaging (default is to message the user)
-  silent  <- ifelse(is.null(silent), T, silent)
+  silent  <- (silent |> is.null()) |> ifelse(T, silent)
   msgUser <- !silent
   msg0    <- ""
   msg1    <- msg0 |> paste0("\t")
@@ -141,7 +139,8 @@ run_fredi_sv <- function(
   co_formatting <- svDataList$co_formatting
   # co_modelTypes <- rDataList$co_modelTypes
   ### Sector names
-  if(is.null(sector)){
+  hasSector <- !(sector |> is.null())
+  if(!hasSector){
     sector_msg1   <- paste0("Please select a sector: ") |> print()
     sector_msg2   <- 1:nrow(sectorInfo) |> lapply(function(i){
       sector_i    <- sectorInfo$sector[i]
@@ -153,8 +152,8 @@ run_fredi_sv <- function(
     sector_msg3   <- "Enter a number:"
     sector_input  <- readline(prompt = sector_msg3) |> as.numeric()
     sector        <- sectorInfo$sector[sector_input]
-    rm("sector_msg1", "sector_msg2", "sector_msg3", "sector_input")
-  }
+    rm(sector_msg1, sector_msg2, sector_msg3, sector_input)
+  } ### End if(hasSector)
   c_sector        <- sector
   paste0("Running FrEDI SV for sector '", c_sector, "':") |> message()
   ### Sector info
@@ -162,23 +161,23 @@ run_fredi_sv <- function(
   df_sectorInfo   <- svSectorInfo[which_sector,]
   c_variants      <- df_sectorInfo[["variant_abbr" ]][which_sector]
   c_variantLabels <- df_sectorInfo[["variant_label"]][which_sector]
-  rm("which_sector")
+  rm(which_sector)
   ###### Invalid Sectors ######
   which_sector    <- (sectorInfo$sector == c_sector) |> which() #; which_sector |> print()
   c_popWtCol      <- sectorInfo[["popWeightCol"]][which_sector] |> tolower()
   c_modelType     <- sectorInfo[["modelType"   ]][which_sector] |> tolower()
-  rm("which_sector")
+  rm(which_sector)
[…]
