### 2024-01-10 798093b
Update run_fredi_sv.R
Updated call to FrEDI config file data
