### 2024-01-10 f93a0d2
delete refYear hardcode
