+++ .github/workflows/update_base_data.yml
@@ -5,7 +5,7 @@ on:
     inputs:
       fredi_data_branches:
         type: string
-        description: Which data only branch of FrEDI_data to pull in data from? (i.e. data_only_branch_state_initial)
+        description: Which data-only branch of FrEDI_data to pull in data from? (e.g., data_only_branch_state_initial)
 
 
 jobs:
@@ -18,7 +18,7 @@ jobs:
           git config --global user.email "actions@github.com"
           git config --global user.name "GitHub Actions"
           git pull --depth=1 --ff https://github.com/USEPA/FrEDI_Data.git ${{ github.event.inputs.fredi_data_branches }} --allow-unrelated-histories
-          mv data/tmp_sysdata.rda FrEDI/R/sysdata.rda
+          mv data/sysdata.rda FrEDI/R/sysdata.rda
           git branch --show-current
           git add FrEDI/R/sysdata.rda
           git rm -r data
