### 2024-01-11 1517360
Small fixes to SV functions
Small fixes to SV functions to get them to work. They work!
