### 2024-02-01 93dd655
Update test_fredi.yml
Changed location of where to save results for Figure H in report figures action
