### 2024-02-01 d990bba
Update test_fredi.yml
Fixed file name for saving Fig H results
