### 2024-02-01 03adb79
Update test_fredi.yml
Updating FigH results to interpolate over impact years
