### 2024-02-13 b70cef5
Improved messaging
Modified some lines in utils to make sure that results are filtered to appropriate max year. Edited FrEDI to do the same. Also changed some messaging
