+++ FrEDI/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: FrEDI
 Title: The Framework for Evaluating Damages and Impacts (FrEDI)
-Version: 3.4
+Version: 4.0
 Authors@R: c(person("Corinne", "Hartin",
                     email   = "hartin.corinne@epa.gov",
                     role    = c("aut","cre"),
@@ -31,7 +31,8 @@ Description: This R package models damages and impacts from climate change in th
     implement FrEDI for projecting impacts from climate change and sea level rise for a selected set of 
     sectors. Version 3 and above includes a module for estimating impacts on socially vulnerable populations for 
     select sectors. Versions > 3.3.0 extend calculations out to 2300 to facilitate the calculation of Net Present Values 
-    (NPV) but only contains default values through 2090--users must provide their own inputs through 2300. 
+    (NPV) but only contains default values through 2090--users must provide their own inputs through 2300. Versions 4.0 and 
+    above calculate values at the state-level for most sectors.  
 License: file LICENSE
 Encoding: UTF-8
 Depends: R (>= 3.5.0),
