### 2024-02-16 d2efa03
Update build_fredi.yml
Edited out double "))" that was causing an error
