+++ FrEDI/man/get_plots.Rd
@@ -1,80 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_plots.R
-\name{get_plots}
-\alias{get_plots}
-\title{Create plots of FrEDI outputs}
-\usage{
-get_plots(
-  data = NULL,
-  column = "annual_impacts",
-  undiscounted = TRUE,
-  plotTypes = "all",
-  save = FALSE,
-  directory = NULL,
-  groupVars = c("sector", "variant")
-)
-}
-\arguments{
-\item{data}{Data frame of summarized outputs produced by \code{\link[=run_fredi]{run_fredi()}}. Do not change column names of the \code{\link[=run_fredi]{run_fredi()}} output before running \code{\link[=get_plots]{get_plots()}}.}
-
-\item{column}{A character string indicating the name of the column in the data for which to create plots.}
-
-\item{undiscounted}{A \code{TRUE/FALSE} value indicating whether the values in column represent undiscounted values or discounted values (i.e., present values). Defaults to \code{undiscounted=TRUE}.}
-
-\item{plotTypes}{Character string or character vector indicating which types of plots to produce. Options are \code{c("heatmaps", "ribbon", "all")}. Set \code{plotTypes="all"} (default) to produce both types of plots.}
-
-\item{save}{A \code{TRUE/FALSE} value indicating whether to save results. If a directory value is supplied (i.e., \code{!is.null(directory)}), defaults to \code{save=TRUE}. Otherwise, default is \code{save=FALSE}.}
-
-\item{directory}{A character string indicating the location of a directory in which to save the plot objects. No default (i.e., \code{directory=NULL}).}
-
-\item{groupVars}{A character vector indicating columns to use for grouping. Defaults to \code{groupVars=c("sector", "variant")}}
-}
-\value{
-\code{\link[=get_plots]{get_plots()}} returns a nested list of plots:
-
-\tabular{ll}{
-\strong{List Index} \tab \strong{Description} \cr
-\code{heatmaps} \tab List of heatmaps with list elements for all unique values for model types present in data (i.e., \code{"GCM"} for results calculated with temperature as the driver value and \code{"SLR"} for results calculated with GMSL as the driver value.\cr
-\code{ribbon} \tab List of lists of annual results. List of annual plots contains a list of sectors. Each sector contains a list of nested regional plots (\code{regional}, with all regions except national results and \code{national}, for national results). \cr
-}
-}
-\description{
-`This function creates plots for the summarized FrEDI outputs. \code{\link[=get_plots]{get_plots()}} returns a list with heatmaps for model types present in the data (GCMs and SLR scenarios) and ribbon plots, by sector.
-}
-\details{
-This function creates plots for the summarized FrEDI outputs. \code{\link[=get_plots]{get_plots()}} returns a list with heatmaps for model types present in the data (GCMs and SLR scenarios) and annual results for all sectors and variants. Results from FrEDI must be summed across impact types before using \code{\link[=get_plots]{get_plots()}} (use \code{\link[=run_fredi]{run_fredi()}} with the defaults or run \code{aggregate_impacts(aggLevels="impactType")} on the output from \code{\link[=run_fredi]{run_fredi()}}).
-
-By default, \code{\link[=get_plots]{get_plots()}} plots results from the \code{"annual_impacts"} column (or users can specify a column name in the data with \code{column}).
-
-The argument \code{undiscounted} is used by \code{\link[=get_plots]{get_plots()}} for plot labels and in file and directory names for saving results.
-
[…]
