### 2024-02-16 7f47863
Update build_fredi.yml
Added default results to commit
