### 2024-02-16 68e50a1
Update build_fredi.yml
Modified line causing a fatal error
