+++ .github/workflows/build_fredi.yml
@@ -64,7 +64,9 @@ jobs:
 
           ###### Build Package ######
           ###### - Build Package but don't include vignettes
-          devtools::build(pkg=pPath0, path=rPath0)
+          # devtools::build(pkg=pPath0, path=rPath0)
+          pPath0 |> print()
+          devtools::build(pkg=pPath0)
 
           '
 
