### 2024-02-16 fcdb5b5
Update build_fredi.yml
Fixed strange symbol that might be causing isues
