### 2024-02-16 b2cd406
Update build_fredi.yml
Removed apostrophes in comments
