### 2024-02-16 1222459
Removed default results
Removed default results from data. Package is now too large with the state-level results
