+++ FrEDI/DESCRIPTION
@@ -40,7 +40,7 @@ Depends: R (>= 3.5.0),
     ggpubr,
     openxlsx
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.1
 LazyData: TRUE
 Suggests: knitr, rmarkdown, prettydoc, devtools, kableExtra
 VignetteBuilder: knitr
