### 2024-02-20 0420ae9
Improve plot legend appearance
Edited routine for creating plots to improve plot legend appearance
