+++ FrEDI/testing/utils_create_report_figures.R
@@ -912,11 +912,9 @@ plot_DoW_by_sector <- function(
         df1      <- df0 |> filter(model_type=="GCM")
         sectors0 <- df1[["sector"]] |> unique()
         df_x     <- sectors0 |> map(function(.y){tibble(type=.x, sector=.y, year=years, label=.y |> paste0("_", years))})
-        # df_x     <- df_x %>% (function(y){do.call(rbind, y)})
         df_x     <- df_x |> bind_rows()
         return(df_x)
       })
-    # df_gcm   <- df_gcm %>% (function(y){do.call(rbind, y)})
     df_gcm   <- df_gcm |> bind_rows()
     df_types <- df_types |> rbind(df_gcm)
     rm(df_gcm)
