### 2024-02-20 ffe3cbb
Update test_fredi.yml
Edited lines to load test code
