### 2024-02-20 bab7a2f
Update test_fredi.yml
Fixed reference to test code path
