+++ FrEDI/DESCRIPTION
@@ -32,7 +32,7 @@ Description: This R package models damages and impacts from climate change in th
     sectors. Version 3 and above includes a module for estimating impacts on socially vulnerable populations for 
     select sectors. Versions > 3.3.0 extend calculations out to 2300 to facilitate the calculation of Net Present Values 
     (NPV) but only contains default values through 2090--users must provide their own inputs through 2300. Versions 4.0 and 
-    above calculate values at the state-level for most sectors.  
+    above calculate values at the state-level.  
 License: file LICENSE
 Encoding: UTF-8
 Depends: R (>= 3.5.0),
