### 2024-02-21 cd342ec
Update utils_save_report_objects.R
Edits to determining heights for appendix plots
