### 2024-02-21 5e2b2d7
Update DESCRIPTION
Updated version number and info about Version 4.0+
