+++ FrEDI/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: FrEDI
 Title: The Framework for Evaluating Damages and Impacts (FrEDI)
-Version: 4.0
+Version: 4.0-public-review
 Authors@R: c(person("Corinne", "Hartin",
                     email   = "hartin.corinne@epa.gov",
                     role    = c("aut","cre"),
