+++ FrEDI/R/import_inputs.R
@@ -70,8 +70,7 @@
 #'   tempfile = tempInputFile,
 #'   slrfile  = slrInputFile,
 #'   popfile  = popInputFile,
-#'   temptype = "global",
-#'   popform  = "wide"
+#'   temptype = "global"
 #' )
 #'
 #' ### Use imports with FREDI:
@@ -84,9 +83,6 @@
 #' @md
 #'
 #'
-# #' @param popform A character string indicating whether the populations in the population input file specified by `popfile` are spread across multiple columns (i.e., `popform="wide"`) or are combined in a single column (i.e., `popform="long"`). By default, the model assumes `popform="wide"`. For both formats (`popform="wide"` or `popform="long"`), the first column contains years. If `popform="wide"` (default), the second through eighth columns of `popfile ` must contain population values for each NCA region, with the associated NCA region as the column name. If `popform="long"`, the second column must contain NCA region names and the third column must contain values for the associated region population.
-# #' Users can specify whether the temperature input is for the contiguous U.S. (CONUS) or global using `temptype` and specify the format of the population scenario using `popform`.
-# #' If the population input is spread across multiple columns (i.e., `popform="wide"`), columns must be named according to the NCA regions. If the population input is in the long format, the region value must be in the second column. The NCA region names for population inputs must match the following values: `"Midwest"`, `"Northeast"`, `"Northern.Plains"`, `"Northwest"`, `"Southeast"`, `"Southern.Plains"`, `"Southwest"`. All regions must be present in the population input file.
 
 ###### import_inputs ######
 ### Created 2021.02.08. Last updated 2021.02.08
@@ -96,37 +92,31 @@ import_inputs <- function(
   slrfile  = NULL,
   popfile  = NULL,
   gdpfile  = NULL,
-  temptype = "conus", ### "global", or "conus" (default)
-  popform  = "wide", ### "wide" or "long" ### Previously: "gather", "spread"
-  byState  = FALSE
+  temptype = "conus" ### "global", or "conus" (default)
 ){
   ###### Messaging ######
-  hasAnyInputs <- list(tempfile, slrfile, popfile, gdpfile) |>
-    lapply(function(x){!is.null(x)}) |>
-    unlist() |> any()
+  namesInputs  <- c("tempfile", "slrfile", "popfile", "gdpfile")
+  hasAnyInputs <- list(tempfile, slrfile, popfile, gdpfile)
+  hasAnyInputs <- namesInputs  |> map(function(x, list0=hasAnyInputs){!(list0[[x]] |> is.null())})
+  hasAnyInputs <- hasAnyInputs |> unlist() |> any()
   silent  <- TRUE
-  msgUser <- ifelse(silent, FALSE, TRUE)
+  msgUser <- !silent
   msg0    <- ""
   msg1    <- msg0 |> paste0("\t")
   msg2    <- msg1 |> paste0("\t")
   msg3    <- msg2 |> paste0("\t")
-  if(hasAnyInputs){
+  if(hasAnyInputs) {
     "\n" |> paste0(msg0) |> paste0("In import_inputs():") |> message()
-  }
+  } ### End if(hasAnyInputs)
 
   ###### Defaults ######
-  ### Set popform default to "wide" (previously "spread")
-  ### Set popform to "wide" if none is specified
-  popform_default  <- "wide"
-  popform          <- ifelse(is.null(popform), popform_default, tolower(popform))
-  wide_pop         <- popform=="wide"
-  if(byState){geo_msg <- " state..."} else{geo_msg <- " region..."}
+  geo_msg <- " state..."
 
   ### Set temperature type default and set temperature type to default if none
   ### is declared. Check whether inputs temperatures are already in CONUS degrees
   temptype_default <- "conus"
[…]
