### 2024-02-21 819c904
Delete configTests.R
Deleted config tests
