+++ FrEDI/R/configTests.R
@@ -1,891 +1,0 @@
-require(tidyverse)
-###### Dimension Tests ######
-#' test_loadData
-#'
-#' @param data_list A list of named tables to be examined
-#' @param outPath A filepath to where test files are saved
-#' @param save  A TRUE/FALSE option to save files or not
-#' @param return A TRUE/FALSE option to return the test_table
-#'
-#' @return Returns a table object if "return" option is TRUE
-
-dataInfo_test <- function(
-    dataList = list(),
-    outPath  = ".",
-    csvName  = "dataInfo_test",
-    save     = TRUE, ### Whether to save results to file
-    return   = TRUE  ### Whether to return results
-) {
-  #dataList <- list_reshapeData
-  ###### List Info ######
-  ### List names
-  ### Length of list and list names
-  listNames <- dataList  |> names()
-  lenList   <- dataList  |> length()
-  lenName   <- listNames |> length()
-  # c(lenList, lenName, lenList==lenName) %>% print
-
-  ###### List Object Types ######
-  ### Get info on object types and add names
-  ### Simplify list of types
-  # cTypes    <- c("data.frame", "list", "character", "numeric")
-  listTypes <- listNames %>% map(~ (dataList[[.]] %>% class()))
-  ### Add names back to list
-  listTypes <- listTypes |> set_names(listNames)
-  # listTypes[1] %>% print
-  ### Simplify types
-  listTypes <- listNames %>% map(~ case_when(
-    ("data.frame" %in% listTypes[[.]]) ~ "data.frame",
-    ("list"       %in% listTypes[[.]]) ~ "list",
-    ("character"  %in% listTypes[[.]]) ~ "character",
-    ("numeric"    %in% listTypes[[.]]) ~ "numeric",
-    TRUE ~ "N/A"
-  ))
-  ### Add names back to list
-  listTypes <- listTypes %>% (function(x){names(x) <- listNames; return(x)})
-  # c(length(listTypes), names(listTypes) %>% length) %>% print
-
-  ###### Initial Table Info ######
-  ### Initialize table of info...make methods specific to class
-  ### Get class of table object
-  df_info   <- tibble(table = listNames)
-  df_info   <- df_info %>% mutate(itemClass = listTypes %>% unlist)
-
-  ### Count number of columns in each table
-  ### Count number of rows in each table
-  ### Count number of distinct rows in each table
-  ### Count number of missing values
-
-  ### Expressions
-  num_cols    <- listNames %>% map(~ .x %>% fun_nCol(a=listTypes, b=dataList)) %>% unlist
-  num_rows    <- listNames %>% map(~ .x %>% fun_nRow(a=listTypes, b=dataList)) %>% unlist
-  unique_rows <- listNames %>% map(~ .x %>% fun_nUnq(a=listTypes, b=dataList)) %>% unlist
-  cols_wAllNA <- listNames %>% map(~ .x %>% fun_nNna(a=listTypes, b=dataList)) %>% unlist
-
-  ### Add to df_info
-  df_info   <- df_info %>% mutate(num_cols    = num_cols)
-  df_info   <- df_info %>% mutate(num_rows    = num_rows)
-  df_info   <- df_info %>% mutate(unique_rows = unique_rows)
-  df_info   <- df_info %>% mutate(cols_wAllNA = cols_wAllNA)
-
-  ###### Check Tests ######
[…]
