### 2024-02-21 d47b21e
Update utils_save_report_objects.R
Updated spacer for appendix plots
