### 2024-02-21 73e952a
Update DESCRIPTION
R doesn't like strings in the package version
