### 2024-02-21 d27fed4
Update DESCRIPTION
Edited contributors list
