### 2024-02-21 bb111e4
Update FrEDI.R
Edited FrEDI dependencies in documentation
