### 2024-02-21 7485582
Update run_fredi.R
Changed version number to reflect text in DESCRIPTION
