+++ FrEDI/R/run_fredi.R
@@ -6,7 +6,7 @@
 #' @description
 #' This function allows users to project annual average climate change impacts through 2090 (2010-2090) for available sectors (see [FrEDI::get_sectorInfo()]), with the option to extend results to 2300 (2010-2300). Users may specify custom temperature, U.S. population, and GDP scenarios. The output is an R data frame object containing annual average impacts, by year, for each sector, variant, impact type, region, state, and model.
 #'
-#' As of FrEDI Version 4.0, [FrEDI::run_fredi()] calculates impacts at the state-level for the following sectors: **ATS Temperature-Related Mortality**, **Asphalt Roads**, **Climate-Driven Changes in Air Quality**, **Electricity Transmission and Distribution**, **Labor**, **Suicide**, **Transportation Impacts from High Tide Flooding**, **Urban Drainage**, **Wildfire**, **Wind Damage**. Eventually, all sectors will be converted to state-level. Sectors that have only region-level impacts will have values in the `"state"` and `"postal"` columns of the outputs data frame set to `"N/A"`.
+#' As of FrEDI Version 4.0.1, [FrEDI::run_fredi()] calculates impacts at the state-level for the following sectors: **ATS Temperature-Related Mortality**, **Asphalt Roads**, **Climate-Driven Changes in Air Quality**, **Electricity Transmission and Distribution**, **Labor**, **Suicide**, **Transportation Impacts from High Tide Flooding**, **Urban Drainage**, **Wildfire**, **Wind Damage**. Eventually, all sectors will be converted to state-level. Sectors that have only region-level impacts will have values in the `"state"` and `"postal"` columns of the outputs data frame set to `"N/A"`.
 #'
 #' @param inputsList=NULL A list of named elements named elements (`names(inputsList)= c("tempInput", "slrInput", "gdpInput", "popInput")`), each containing data frames of custom temperature, global mean sea level rise (GMSL), gross domestic product (GDP), and/or population scenarios, respectively, over a continuous period in the range 2010 to 2300. Temperature and sea level rise inputs should start in 2000 or earlier. Values for population and GDP scenarios can start in 2010 or earlier. Values for each scenario type must be within reasonable ranges. For more information, see [FrEDI::import_inputs()].
 #'
