### 2024-02-21 8041c72
Update utils_save_report_objects.R
Changed spacer multiplier for appendix plots
