### 2024-02-21 4fada0f
Added legend guides
Added legend guides to report plots
