### 2024-02-21 c270fd4
Updated NA tests
Fixed NA tests code
