### 2024-02-21 5868150
Updated legend parameters
