### 2024-02-21 ce4ed01
Updated report plot options
