### 2024-02-21 258a131
Update configTests_utils.R
Fixed reference to function
