### 2024-02-21 8193af8
Update utils_plot_DOW_byImpactTypes.R
Fixed ggarrange statement
