### 2024-02-21 36823e2
Update utils_plot_DOW_byImpactTypes.R
Fixed ggarrange statement
