+++ FrEDI/testing/utils_plot_DOW_byImpactTypes.R
@@ -267,11 +267,11 @@ plot_DOW_byImpactTypes <- function(
     ### Add Plot Title & Y Title
     # plotList_i <- list(spacer1=spacer0, plot=plotGrid_i, legend=legend0)
     # plotList_i <- list(spacer1=spacer0, plot=plotGrid_i, legend=grobLgd0)
-    plotList_i <- list(spacer1=spacer0, plot=plotGrid_i, spacer2=spacer0, legend=grobLgd0)
+    plotList_i <- list(spacer1=spacer0, plot=plotGrid_i, spacer2=spacer0, legend=grobLgd0, spacer3=spacer0)
     # plotGrid_i <- ggarrange(plotlist=plotList_i, nrow=2, ncol=1, common.legend=T, legend="none", heights=c(0.1, n_impTypes, 0.25))
-    plotGrid_i <- ggarrange(plotlist=plotList_i, nrow=4, ncol=1, legend="none", heights=c(0.01, n_impTypes, 0.01, 0.2))
+    plotGrid_i <- ggarrange(plotlist=plotList_i, nrow=5, ncol=1, legend="none", heights=c(0.01, n_impTypes, 0.05, 0.2, 0.01))
     title0_i   <- sector0
-    grobTit_i  <- text_grob(title0_i, color="black", size=14, face="bold", hjust=0.5)
+    grobTit_i  <- text_grob(title0_i, color="black", size=12, face="bold", hjust=0.5)
     plotYTit_i <- text_grob(yTitle, color = "black", rot  = 90)
     plotGrid_i <- plotGrid_i |> annotate_figure(top=grobTit_i, left=plotYTit_i)
     # "got here4..." |> print()
