### 2024-02-21 0bbb5ec
Update plot opts
