### 2024-02-21 2e45b6a
Updated plot opts
