### 2024-02-21 ea83179
Changed legend title positions to the left side for appearance reasons
