### 2024-02-22 6af0abc
Changed palette for report figures
Changed palette in report figures to a colorblind friendly palette (developed using https://davidmathlogic.com/colorblind/)
