### 2024-02-22 81caffa
Edited spacing for appendix figures
Edited spacing for appendix figures (smaller now that we're using dashed lines instead of point shapes)
