### 2024-02-22 aebc71f
Updated figure code
Updated colors with more distinct options. Edited spacing around legends. Removed shapes from SLR plots
