+++ FrEDI/testing/utils_create_report_figures.R
@@ -1,6 +1,7 @@
 ### Function to get manual colors for regions, states in a region, or models
 fun_manual_colors <- function(x=1){
-  colorVals <- c("D81B60", "FF792D", "4E94E4", "735EA0", "67032F", "49D4F5", "D082CD", "68A796")
+  # colorVals <- c("D81B60", "FF792D", "4E94E4", "735EA0", "67032F", "49D4F5", "D082CD", "68A796")
+  colorVals <- c("D81B60", "B38ADA", "FB9358", "2EAD8A", "49D4F5", "6B7682", "893620", "2C5EB7")
   colorVals <- "#" |> paste0(colorVals)
   return(colorVals)
 }
