### 2024-02-22 473457a
Plotting function edits
Changed color palette for plots. Updated spacing.
