+++ FrEDI/testing/utils_plot_DOW_byImpactTypes.R
@@ -268,8 +268,13 @@ plot_DOW_byImpactTypes <- function(
     h_spacer1  <- 0.05
     h_spacer2  <- 0.05
     h_plots    <- n_impTypes * (1 + 0.07)
-    h_legend   <- 0.75 / n_impTypes
-    # h_legend   <- case_when(n_variants = 1, , 0.5)
+    # h_legend   <- 0.75 / n_impTypes
+    # h_legend   <- case_when(
+    #   n_impTypes = 1 ~ 0.5,
+    #   n_impTypes > 3 ,
+    #   0.5
+    #   )
+    h_legend   <- 0.5
     h_spacers  <- c(h_spacer1, h_plots, h_spacer1, h_legend, h_spacer1/4)
     ### Add Plot Title & Y Title
     # plotList_i <- list(spacer1=spacer0, plot=plotGrid_i, legend=legend0)
