+++ FrEDI/install/00_install.R
@@ -7,7 +7,7 @@
 install_fredi <- function(
   installFrom = NULL, ### Select a number, one or two
   package     = "FrEDI",
-  version     = "3.3.7",
+  version     = "4.0.1",
   libPath     = .libPaths()[1] ### Location of library path
 ){
   ###### Load Packages ######
@@ -24,7 +24,7 @@ install_fredi <- function(
   ###### Package Info ######
   package_name      <- package ### Package name
   package_version   <- version ### Package version number
-  package_file      <- package_name %>% paste0("_", package_version, ".", "tar.gz")
+  package_file      <- package_name |> paste0("_", package_version, ".", "tar.gz")
 
   ###### Install from GitHub ######
   if(installFrom == "1"){
@@ -43,40 +43,40 @@ install_fredi <- function(
   ###### Install from Source ######
   else if(installFrom == "2"){
     ### Otherwise, install from source
-    "Copy tar.gz file to your R library before running this script..." %>% message
-    "Check location of your R library by typing '.libPaths()[1]' into the command line..." %>% message
+    "Copy tar.gz file to your R library before running this script..." |> message()
+    "Check location of your R library by typing '.libPaths()[1]' into the command line..." |> message()
 
     ###### Package Location
     package_location  <- libPath; #package_location
-    package_path      <- package_location %>% file.path(package_name)
-    package_filePath  <- package_location %>% file.path(package_file)
+    package_path      <- package_location |> file.path(package_name)
+    package_filePath  <- package_location |> file.path(package_file)
     # package_location %>% list.files
 
     ###### Check Files
     ### Check existance of package file
-    package_fileExists <- package_filePath %>% file.exists
+    package_fileExists <- package_filePath |> file.exists()
     ### Return without doing anything if package doesn't exist
     if(!package_fileExists){
-      paste0("Warning: ", package_file, " not found in ", package_location, "...") %>% message
-      paste0("\t", "Add", package_file, " to ", package_location, " and then rerun this script.", "\n") %>% message
-      paste0("Exiting...") %>% message
+      paste0("Warning: ", package_file, " not found in ", package_location, "...") |> message()
+      paste0("\t", "Add", package_file, " to ", package_location, " and then rerun this script.", "\n") |> message()
+      paste0("Exiting...") |> message()
       return()
     } else{
       ### If the package exists, try to install the package
       ### First check whether package already installed
-      package_exists     <- package_path %>% dir.exists
+      package_exists     <- package_path |> dir.exists()
       ###### Remove Previous Library
       ### If package is already installed, remove the previous install
       ### Then check that it was removed and install the package
       if(package_exists){
-        paste0("Warning: ", package, " already exists!") %>% message
+        paste0("Warning: ", package, " already exists!") |> message()
         removePrompt <- paste0("Remove existing install? (y/n)")
-        removeInput  <- readline(prompt = removePrompt) %>% tolower
+        removeInput  <- readline(prompt = removePrompt) |> tolower()
         if(removeInput=="y"){
           unlink(package_path, recursive = T)
           install.packages(package_filePath, repos=NULL, type="source", lib=package_location)
         } else{
-          paste0("Exiting without installing...") %>% message
+          paste0("Exiting without installing...") |> message()
           return()
         }
       } else{
@@ -87,8 +87,8 @@ install_fredi <- function(
   }
   ###### Exit without installing ######
   else{
-    paste0("Warning: No option selected.", "\n") %>% message
-    paste0("Exiting without installing...") %>% message
+    paste0("Warning: No option selected.", "\n") |> message()
+    paste0("Exiting without installing...") |> message()
     return()
   }
 }
