### 2024-02-22 2a3e501
Update utils_create_report_figures.R
Edited plot colors
