### 2024-02-22 67f7fae
Update utils_create_report_figures.R
Updated a single color in the report plot palettes
