+++ FrEDI/testing/utils_create_report_figures.R
@@ -2,7 +2,7 @@
 fun_manual_colors <- function(x=1){
   # colorVals <- c("D81B60", "FF792D", "4E94E4", "735EA0", "67032F", "49D4F5", "D082CD", "68A796")
   # colorVals <- c("D81B60", "B38ADA", "FB9358", "2EAD8A", "49D4F5", "6B7682", "893620", "2C5EB7")
-  colorVals <- c("FF5291", "E1C4FD", "7D2E19", "4BD886", "F97E39", "A952EC", "639BFF", "386D79")
+  colorVals <- c("FF5291", "D8B8F7", "7D2E19", "4BD886", "F97E39", "A952EC", "639BFF", "386D79")
   colorVals <- "#" |> paste0(colorVals)
   return(colorVals)
 }
