+++ FrEDI/R/FrEDI.R
@@ -4,22 +4,26 @@
 #'
 #' [FrEDI] is an R package being developed by the U.S. Environmental Protection Agency (EPA). The functions and data provided by this package can be used to estimate climate change impacts for the contiguous United States (CONUS) using the Framework for Evaluating Damages and Impacts (FrEDI), developed as part of EPA's [Climate Change Impacts and Risk Analysis](https://epa.gov/cira/) (CIRA) project. [FrEDI] contains R code that implement FrEDI and allow users to project impacts from climate change and sea level rise for a selected set of sectors.
 #'
+#' For help getting started with [FrEDI], visit <https://usepa.github.io/FrEDI/articles/manual.html>.
+#'
+#' For additional package documentation, see <https://usepa.github.io/FrEDI/index.html>.
+#'
 #' For more information on the Framework and the CIRA project, visit <https://epa.gov/cira/>, especially <https://epa.gov/cira/FrEDI/>.
 #'
 #' @section Overview of Functions:
-#' The function [FrEDI::run_fredi()] provided in this package is the primary function implementing FrEDI The main inputs to [FrEDI::run_fredi()] are climate scenarios (temperature in degrees Celsius, global mean sea level rise in centimeters) and socioeconomic scenarios (gross domestic product, regional population). [FrEDI::run_fredi()] projects climate impacts for these scenarios using the temperature binning framework. Users can also specify the levels at which to aggregate the climate impacts.
+#' The function [FrEDI::run_fredi()] provided in this package is the primary function implementing `FrEDI` The main inputs to [FrEDI::run_fredi()] are climate scenarios (temperature in degrees Celsius, global mean sea level rise in centimeters) and socioeconomic scenarios (gross domestic product, regional population). [FrEDI::run_fredi()] projects climate impacts for these scenarios using the temperature binning framework. Users can also specify the levels at which to aggregate the climate impacts.
 #'
-#' [FrEDI] also contains functions to assist in the pre-processing of input scenarios and the post-processing of outputs. For instance, the pre-processing function [FrEDI::import_inputs()] helps users to import scenarios from user-specified comma-separated value (CSV) files, while the post-processing function [FrEDI::get_plots()] generates a couple of different types of images (creates heatmaps and ribbon plots of climate impacts over time) from the outputs of [FrEDI::run_fredi()].
+#' [FrEDI] also contains functions to assist in the pre-processing of input scenarios and the post-processing of outputs.
 #'
-#' Additional helper functions provided by [FrEDI] include [FrEDI::convertTemps()], [FrEDI::temps2slr()], [FrEDI::get_sectorInfo()], and [FrEDI::aggregate_impacts()].
+#' * Pre-processing functions include [FrEDI::get_sectorInfo()], [FrEDI::import_inputs()], [FrEDI::convertTemps()], [FrEDI::temps2slr()].
+#'      * [FrEDI::get_sectorInfo()] allows users to access a list of sectors within `FrEDI` and related sector information.
+#'      * [FrEDI::import_inputs()] helps users in importing custom scenarios from user-specified comma-separated value (CSV) files.
+#'      * [FrEDI::convertTemps()] helps users to convert between global mean temperature and temperatures for the contiguous United States (CONUS) (both in degrees Celsius).
+#'      * [FrEDI::temps2slr()] helps users to estimate global mean sea level rise (GMSL, in centimeters) from global mean temperature in degrees Celsius.
 #'
[…]
