### 2024-02-23 cbfb3a6
Updated documentation for package and data
