+++ FrEDI/man/FrEDI-package.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/FrEDI.R
+% Please edit documentation in R/FrEDI-package.R
 \docType{package}
 \name{FrEDI-package}
 \alias{FrEDI}
@@ -10,23 +10,30 @@ FrEDI: The Framework for Evaluating Damages and Impacts}
 \link{FrEDI} is an R package being developed by the U.S. Environmental Protection Agency (EPA). The functions and data provided by this package can be used to estimate climate change impacts for the contiguous United States (CONUS) using the Framework for Evaluating Damages and Impacts (FrEDI), developed as part of EPA's \href{https://epa.gov/cira/}{Climate Change Impacts and Risk Analysis} (CIRA) project. \link{FrEDI} contains R code that implement FrEDI and allow users to project impacts from climate change and sea level rise for a selected set of sectors.
 }
 \details{
+For help getting started with \link{FrEDI}, visit \url{https://usepa.github.io/FrEDI/articles/manual.html}.
+
+For additional package documentation, see \url{https://usepa.github.io/FrEDI/index.html}.
+
 For more information on the Framework and the CIRA project, visit \url{https://epa.gov/cira/}, especially \url{https://epa.gov/cira/FrEDI/}.
 }
 \section{Overview of Functions}{
 
-The function \code{\link[=run_fredi]{run_fredi()}} provided in this package is the primary function implementing FrEDI The main inputs to \code{\link[=run_fredi]{run_fredi()}} are climate scenarios (temperature in degrees Celsius, global mean sea level rise in centimeters) and socioeconomic scenarios (gross domestic product, regional population). \code{\link[=run_fredi]{run_fredi()}} projects climate impacts for these scenarios using the temperature binning framework. Users can also specify the levels at which to aggregate the climate impacts.
+The function \code{\link[=run_fredi]{run_fredi()}} provided in this package is the primary function implementing \code{FrEDI} The main inputs to \code{\link[=run_fredi]{run_fredi()}} are climate scenarios (temperature in degrees Celsius, global mean sea level rise in centimeters) and socioeconomic scenarios (gross domestic product, regional population). \code{\link[=run_fredi]{run_fredi()}} projects climate impacts for these scenarios using the temperature binning framework. Users can also specify the levels at which to aggregate the climate impacts.
 
-\link{FrEDI} also contains functions to assist in the pre-processing of input scenarios and the post-processing of outputs. For instance, the pre-processing function \code{\link[=import_inputs]{import_inputs()}} helps users to import scenarios from user-specified comma-separated value (CSV) files, while the post-processing function \code{\link[=get_plots]{get_plots()}} generates a couple of different types of images (creates heatmaps and ribbon plots of climate impacts over time) from the outputs of \code{\link[=run_fredi]{run_fredi()}}.
+\link{FrEDI} also contains functions to assist in the pre-processing of input scenarios and the post-processing of outputs.
+\itemize{
+\item Pre-processing functions include \code{\link[=get_sectorInfo]{get_sectorInfo()}}, \code{\link[=import_inputs]{import_inputs()}}, \code{\link[=convertTemps]{convertTemps()}}, \code{\link[=temps2slr]{temps2slr()}}.
+\itemize{
+\item \code{\link[=get_sectorInfo]{get_sectorInfo()}} allows users to access a list of sectors within \code{FrEDI} and related sector information.
+\item \code{\link[=import_inputs]{import_inputs()}} helps users in importing custom scenarios from user-specified comma-separated value (CSV) files.
[…]
