### 2024-02-23 cda565a
Update run_fredi.R
Added information to FrEDI about whether values are defaults or not. Updated documentation for `outputList` accordingly
