+++ FrEDI/R/run_fredi.R
@@ -22,6 +22,8 @@
 #'
 #' @param outputList=FALSE A ` TRUE/FALSE` value indicating whether to output results as a data frame object (`outputList = FALSE`, default) or to return a list of objects (`outputList = TRUE`) that includes information about model provenance (including input arguments and input scenarios) along with the data frame of results.
 #'
+#' @param allCols=FALSE A `TRUE/FALSE` value indicating whether to include intermediate column values in results (e.g., physical and economic multipliers). Used in testing. Note that aggregation levels must be set to `aggLevels="none"` to properly return the intermediate columns. Defaults to `allCols=FALSE`).
+#'
 #' @param silent=TRUE A `TRUE/FALSE` value indicating the level of messaging desired by the user (default=`TRUE`).
 #'
 #'
@@ -97,11 +99,14 @@
 #'
 #'
 #' @return
-#' If `outputList=FALSE`, the output of [FrEDI::run_fredi()] is a data frame object containing annual average impacts over the analysis period, for each sector, variant, model (GCM or SLR scenario), impact type, impact year, and region. If `outputList=TRUE`, [FrEDI::run_fredi()] returns a list object containing the following:
+#' If `outputList=FALSE`, the output of [FrEDI::run_fredi()] is a dataframe object (described above) containing annual average impacts over the analysis period, for each sector, variant, impact type, impact year, region, state, and model (GCM name for temperature-driven sectors and "Interpolation" for SLR-driven sectors).
 #'
-#' * __`arguments`__, containing a list with values for the arguments passed to [FrEDI::run_fredi()] (with the exception of scenarios passed to the `inputsList` argument, which are provided in the `driverScenarios` list element).
-#' * __`driverScenarios`__, a list object containing elements with the driver scenarios for temperature, SLR, population, and GDP used in the model
-#' * `results`, containing a data frame of annual average impacts.
+#' If `outputList=TRUE`, [FrEDI::run_fredi()] returns a list object containing the following:
+#'
+#' * __`statusList`__. A list with values for the arguments passed to [FrEDI::run_fredi()] (including defaults if unspecified).
+#' * __`argsList`__. A list with elements named after [FrEDI::run_fredi()] arguments, containing the values of the arguments passed to [FrEDI::run_fredi()] (or default values if unspecified).
+#' * __`scenarios`__. A list with named elements `"temp"`, `"slr"`, `"gdp"`, and `"pop"` -- each containing the scenarios for temperature, SLR, GDP, and population as used by the model in calculating impacts.
+#' * __`results`__. Containing a data frame of annual impacts (i.e., the same data frame returned if `outputList=FALSE`).
 #'
 #'
 #'
@@ -186,21 +191,58 @@ run_fredi <- function(
 ){
   ###### Set up the environment ######
   ### Level of messaging (default is to message the user)
-  silent    <- silent |> is.null() |> ifelse(T, silent)
   msgUser   <- !silent
   ### Uncomment for allCols
   doPrimary <- F  ### whether to filter to primary impacts
   ### Model years and NPD (FrEDI past 2090)
   refYear   <- 2090
   npdYear   <- 2300
-  maxYear   <- maxYear  |> is.null() |> ifelse(refYear, maxYear)
[…]
