### 2024-02-23 02c5910
Update run_fredi.R
Small edits
