+++ FrEDI/testing/configTests.R
@@ -23,18 +23,18 @@ dataInfo_test <- function(
   listNames <- dataList  |> names()
   lenList   <- dataList  |> length()
   lenName   <- listNames |> length()
-  # c(lenList, lenName, lenList==lenName) %>% print
+  # c(lenList, lenName, lenList==lenName) |> print()
 
   ###### List Object Types ######
   ### Get info on object types and add names
   ### Simplify list of types
   # cTypes    <- c("data.frame", "list", "character", "numeric")
-  listTypes <- listNames %>% map(~ (dataList[[.]] %>% class()))
+  listTypes <- listNames |> map(~ (dataList[[.]] |> class()))
   ### Add names back to list
   listTypes <- listTypes |> set_names(listNames)
-  # listTypes[1] %>% print
+  # listTypes[1] |> print
   ### Simplify types
-  listTypes <- listNames %>% map(~ case_when(
+  listTypes <- listNames |> map(~ case_when(
     ("data.frame" %in% listTypes[[.]]) ~ "data.frame",
     ("list"       %in% listTypes[[.]]) ~ "list",
     ("character"  %in% listTypes[[.]]) ~ "character",
@@ -42,14 +42,14 @@ dataInfo_test <- function(
     TRUE ~ "N/A"
   ))
   ### Add names back to list
-  listTypes <- listTypes %>% (function(x){names(x) <- listNames; return(x)})
-  # c(length(listTypes), names(listTypes) %>% length) %>% print
+  listTypes <- listTypes |> (function(x){names(x) <- listNames; return(x)})()
+  # c(length(listTypes), names(listTypes) |> length) |> print()
 
   ###### Initial Table Info ######
   ### Initialize table of info...make methods specific to class
   ### Get class of table object
   df_info   <- tibble(table = listNames)
-  df_info   <- df_info %>% mutate(itemClass = listTypes %>% unlist)
+  df_info   <- df_info |> mutate(itemClass = listTypes |> unlist())
 
   ### Count number of columns in each table
   ### Count number of rows in each table
@@ -57,65 +57,65 @@ dataInfo_test <- function(
   ### Count number of missing values
 
   ### Expressions
-  num_cols    <- listNames %>% map(~ .x %>% fun_nCol(a=listTypes, b=dataList)) %>% unlist
-  num_rows    <- listNames %>% map(~ .x %>% fun_nRow(a=listTypes, b=dataList)) %>% unlist
-  unique_rows <- listNames %>% map(~ .x %>% fun_nUnq(a=listTypes, b=dataList)) %>% unlist
-  cols_wAllNA <- listNames %>% map(~ .x %>% fun_nNna(a=listTypes, b=dataList)) %>% unlist
+  num_cols    <- listNames |> map(~ .x |> fun_nCol(a=listTypes, b=dataList)) |> unlist()
+  num_rows    <- listNames |> map(~ .x |> fun_nRow(a=listTypes, b=dataList)) |> unlist()
+  unique_rows <- listNames |> map(~ .x |> fun_nUnq(a=listTypes, b=dataList)) |> unlist()
+  cols_wAllNA <- listNames |> map(~ .x |> fun_nNna(a=listTypes, b=dataList)) |> unlist()
 
   ### Add to df_info
-  df_info   <- df_info %>% mutate(num_cols    = num_cols)
-  df_info   <- df_info %>% mutate(num_rows    = num_rows)
-  df_info   <- df_info %>% mutate(unique_rows = unique_rows)
-  df_info   <- df_info %>% mutate(cols_wAllNA = cols_wAllNA)
+  df_info   <- df_info |> mutate(num_cols    = num_cols)
+  df_info   <- df_info |> mutate(num_rows    = num_rows)
[…]
