### 2024-02-23 2e8519c
Update build_fredi.yml
Modified command to update default results and documentation but not build the package.
