### 2024-02-23 a74c3c2
Updated contributors info
Added heaston@indecon.com to contributors list
