### 2024-02-23 de5abbf
Update run_fredi.R
Fixed elasticity and aggLevels
