+++ FrEDI/R/run_fredi.R
@@ -201,7 +201,7 @@ run_fredi <- function(
   do_npd    <- maxYear0 > refYear
 
 
-  ###### Return List ######
+  ###### ** Return List ######
   ### Initialize list to return
   returnList <- list() ### List to return args, scenarios, and statuses
   argsList   <- list() ### List of arguments
@@ -214,14 +214,12 @@ run_fredi <- function(
 
   ### Initialize status list
   ### Add statuses:
-  ### - inputsList items assessed further below
-  aggLevels0 <- c("national", "modelaverage", "impactyear", "impacttype")
-  aggLevels  <- aggLevels |> tolower()
+  ### - inputsList items and aggLevels assessed further below
   ### Add to list
   if(outputList){
     statusList[["inputsList"]] <- inputsList
     statusList[["sectorList"]] <- (!(sectorList |> is.null())) |> ifelse("Default", "Custom")
-    statusList[["aggLevels" ]] <- ("all" %in% aggLevels | (aggLevels %in% aggLevels0) |> all()) |> ifelse("Default", "Custom")
+    statusList[["aggLevels" ]] <- "placeholder"
     statusList[["elasticity"]] <- (elasticity == 1) |> ifelse("Default", "Custom")
     statusList[["maxYear"   ]] <- (maxYear == refYear & !thru2300) |> ifelse("Default", "Custom")
     statusList[["thru2300"  ]] <- (!thru2300  ) |> ifelse("Default", "Custom")
@@ -235,6 +233,8 @@ run_fredi <- function(
   ### - inputsList items, sectorList, and aggLevels assessed further below
   if(outputList){
     argsList[["inputsList"]] <- inputsList
+    argsList[["sectorList"]] <- "placeholder"
+    argsList[["aggLevels" ]] <- "placeholder"
     argsList[["elasticity"]] <- elasticity
     argsList[["maxYear"   ]] <- maxYear0
     argsList[["thru2300"  ]] <- thru2300
@@ -244,7 +244,7 @@ run_fredi <- function(
   } ### End if(outputList)
 
 
-  ###### Create file paths ######
+  ###### ** Load Objects ######
   ### Assign data objects to objects in this namespace
   for(list_i in rDataList){
     ### Assign list elements to name space
@@ -260,26 +260,45 @@ run_fredi <- function(
   ### Assign FrEDI config
   for(name_i in fredi_config |> names()){assign(name_i, fredi_config[[name_i]]); rm(name_i)}
 
+
   ### Years
   maxYear    <- maxYear0
   list_years <- minYear:maxYear
   # maxYear    |> print(); list_years |> max() |> print()
 
-  ###### Aggregation level  ######
+
+  ###### ** Aggregation Levels  ######
   ### Types of summarization to do: default
   ### Aggregation levels
-  aggList1   <- aggList0  |> c("all", "none") |> tolower()
+  aggList0   <- aggList0  |> tolower()
+  aggLevels  <- aggLevels |> tolower()
+  ### Update status list
+  statusList[["aggLevels" ]] <- ("all" %in% aggLevels | (aggLevels %in% aggList0) |> all()) |> ifelse("Default", "Custom")
+  # aggList0   <- c("national", "modelaverage", "impactyear", "impacttype")
+  aggList1   <- aggList0  |> c("all", "none")
   aggLevels  <- aggLevels |> (function(y, z=aggList1){y[y %in% z]})()
   ### If none specified, no aggregation (only SLR interpolation)
   ### Otherwise, aggregation depends on length of agg levels
   if     ("none" %in% aggLevels) {aggLevels <- c()}
   else if("all"  %in% aggLevels) {aggLevels <- aggList0}
[…]
