### 2024-02-23 04a9697
Update build_fredi.yml
Removed lines of code related to TAR files (deprecated)
