### 2024-02-26 d3b2808
Updated package documentation with requested changes
Updated FrEDI-package (.R and .Rd) files with requested changes for merge.
