+++ FrEDI/man/FrEDI-package.Rd
@@ -18,7 +18,7 @@ For more information on the Framework and the CIRA project, visit \url{https://e
 }
 \section{Overview of Functions}{
 
-The function \code{\link[=run_fredi]{run_fredi()}} provided in this package is the primary function implementing \code{FrEDI} The main inputs to \code{\link[=run_fredi]{run_fredi()}} are climate scenarios (temperature in degrees Celsius, global mean sea level rise in centimeters) and socioeconomic scenarios (gross domestic product, regional population). \code{\link[=run_fredi]{run_fredi()}} projects climate impacts for these scenarios using the temperature binning framework. Users can also specify the levels at which to aggregate the climate impacts.
+The function \code{\link[=run_fredi]{run_fredi()}} provided in this package is the primary function implementing the \href{https://epa.gov/cira/FrEDI/}{\code{FrEDI}} developed by the U.S. EPA for projecting annual climate impacts. The main inputs to \code{\link[=run_fredi]{run_fredi()}} are climate scenarios (temperature in degrees Celsius, global mean sea level rise in centimeters) and socioeconomic scenarios (U.S. gross domestic product, state population).
 
 \link{FrEDI} also contains functions to assist in the pre-processing of input scenarios and the post-processing of outputs.
 \itemize{
@@ -33,7 +33,7 @@ The function \code{\link[=run_fredi]{run_fredi()}} provided in this package is t
 
 \code{\link[=aggregate_impacts]{aggregate_impacts()}} is a post-processing helper function that helps users to aggregate and summarize the outputs of \link{FrEDI}. \code{\link[=aggregate_impacts]{aggregate_impacts()}} can be used to calculate national totals, model averages, sum over impact types, and interpolate between multiple impact years (note that \code{\link[=run_fredi]{run_fredi()}} will automatically run \code{\link[=aggregate_impacts]{aggregate_impacts()}} before returning outputs if the \code{aggLevels} argument is not \code{"none"}).
 
-Versions 2.3.0 and above include the \code{FrEDI} Social Vulnerability (SV) module for estimating impacts on socially vulnerable populations for select sectors. \code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}} allows users to access a list of sectors within the FrEDI SV module and related sector information. The function \code{\link[=run_fredi_sv]{run_fredi_sv()}} is the main function for the \code{FrEDI} SV module. \code{\link[=run_fredi_sv]{run_fredi_sv()}} is designed to calculate impacts for a single sector at a time for a custom population scenario or one or more custom temperature or sea level rise scenarios. When running \code{\link[=run_fredi_sv]{run_fredi_sv()}}, users have the option to save results to an Excel file that provides basic data visualization for the module outputs. For more information on the data underlying the \code{FrEDI} SV module, visit \url{https://www.epa.gov/cira/social-vulnerability-report/}.
+Versions 2.3.0 and above include the \code{FrEDI} Social Vulnerability (SV) module for estimating impacts on socially vulnerable populations for select sectors. \code{\link[=get_sv_sectorInfo]{get_sv_sectorInfo()}} allows users to access a list of sectors within the FrEDI SV module and related sector information. The function \code{\link[=run_fredi_sv]{run_fredi_sv()}} is the main function for the \code{FrEDI} SV module. \code{\link[=run_fredi_sv]{run_fredi_sv()}} is designed to calculate impacts for a single sector at a time for a custom population scenario or one or more custom temperature or sea level rise scenarios. For more information on the data underlying the \code{FrEDI} SV module, see \url{https://www.epa.gov/cira/social-vulnerability-report/}.
 
 @section Overview of Package Contents:
 \link{FrEDI} consists of files in the following directories:
