### 2024-03-08 b05481f
Update data.R
Fixed wording around temperature scenarios.
