+++ .github/workflows/test_fredi.yml
@@ -6,7 +6,7 @@
 ### For uploading artifacts:
 ###     "path:" is the output path where Pandoc will write the compiled PDF.
 ###      Note, this should be the same directory as the input paper.md
-name: Test FrEDI Package
+name: 3. Test FrEDI Package
 
 # on: [workflow_dispatch]
 on:
