### 2024-05-21 deb4bf2
Updated create_DoW_results workflow
Updated create_DoW_results workflow to try to get it to work on GitHub actions
