+++ FrEDI/testing/utils_create_report_figures.R
@@ -67,7 +67,7 @@ sum_with_na <- function(
 
   ### Multiply column
   df0[[col0]] <- df0[[col0]] * df0[["is_NA"]]
-  # df0    <- df0 |> rename_at(.vars=c("yCol"), ~c(col0))
+  # df0    <- df0 |> rename_at(c("yCol"), ~c(col0))
   ### Drop columns
   if(drop0){df0 <- df0 |> select(-c("is_NA"))}
   ### Return
@@ -263,7 +263,7 @@ agg_fredi_scenario <- function(
   group0  <- group0 |> c("model_type", "model")
   group0  <- group0 |> c("sectorprimary", "includeaggregate")
   group0  <- group0 |> c(drop0)
-  df0     <- df0 |> FrEDI::aggregate_impacts(aggLevels = aggLevels, groupByCols = group0)
+  df0     <- df0 |> FrEDI::aggregate_impacts(aggLevels=aggLevels, groupByCols=group0)
   # df0     <- df0 |> FrEDI::aggregate_impacts(aggLevels = aggLevels)
   ### Return
   return(df0)
@@ -278,75 +278,113 @@ run_scenario <- function(
     sectors   = FrEDI::get_sectorInfo(), ### Which sectors
     scenCols  = c("scenario", "year", "temp_C_conus", "temp_C_global", "slr_cm"),
     joinCols  = c("year"),
-    aggLevels = c("modelaverage", "national")
+    aggLevels = c("modelaverage", "national"),
+    save      = FALSE,
+    return    = TRUE,
+    outPath   = "." |> file.path("report_figures")
 ){
+  ### Values & conditions
+  agg0       <- !("none" %in% aggLevels)
+  scenario0  <- scenario
+  rm(scenario)
+  # agg0 |> print()
+
+  ### File names and info
+  fType0 <- "rda"
+  fName0 <- "integer_results_byType" |> paste0("_", scenario0)
+  # fName0 <- fName0 |> paste0(".", fType0)
+
   ### Filter to scenario
-  scenario0 <- scenario
-  df_x0     <- df0 |> filter(scenario==scenario0)
-  rm("df0")
+  cScenarios <- df0 |> pull(scenario) |> unique()
+  nScenarios <- cScenarios |> length()
+  df0        <- df0 |> filter(scenario==scenario0)
+  # df0 |> nrow() |> print()
+
+  ### Message user
+  "\n" |> paste0("Running scenario ", (cScenarios == scenario0) |> which(), "/" , nScenarios, "...") |> message()
+
   ### Run FrEDI
-  if(fredi){
-    df_x0 <- df_x0 |> run_fredi_scenario(
-      sectors   = sectors,
-      scenCols  = scenCols,
-      joinCols  = joinCols
-    ) ### End run_fredi_scenario
-  } ### End if(fredi)
-  # "got here1" |> print(); df_x0 |> glimpse()
+  if(fredi) df0 <- df0 |> run_fredi_scenario(sectors=sectors, scenCols=scenCols, joinCols=joinCols)
+  # "got here1" |> print(); df0 |> glimpse()
 
   ### Aggregate FrEDI
-  agg0      <- !("none" %in% aggLevels)
-  # agg0 |> print()
-  if(agg0){
-    # "got here1" |> print()
-    df_x0 <- df_x0 |> agg_fredi_scenario(
-      scenCols  = scenCols,
-      joinCols  = joinCols,
-      aggLevels = aggLevels
-    ) ### End run_fredi_scenario
-  } ### End if(agg0)
-  # "got here2" |> print(); df_x0 |> glimpse()
[…]
