+++ FrEDI/scripts/create_DoW_results.R
@@ -3,6 +3,7 @@
 ###### Load Packages ######
 require(tidyverse)
 require(ggpubr)
+# require(arrow)
 # require(cowplot)
 # require(FrEDI)
 
@@ -11,13 +12,14 @@ create_DoW_results <- function(
     sectors  = FrEDI::get_sectorInfo(), ### Which sectors
     gcmYears = c(2090),   ### Which years to report on for GCM sectors
     slrYears = c(2050, 2090), ### Which years to report on for SLR sectors
-    byState  = TRUE,       ### Whether values are by state or just by region
+    byState  = TRUE,      ### Whether values are by state or just by region
+    # totals   = FALSE,     ### Whether to do totals
     silent   = TRUE,      ### Degree of messaging
     testing  = FALSE,     ### Whether to print out extra diagnostic values
     aggOnly  = TRUE,      ### Whether to only include sectors for which "includeaggregate==1" in Fig 7 plots
     loadCode = "project", ### Whether to load code as source or devtools
     fpath    = "." ,      ### Path to main FrEDI directory to load code from if loadCode == "project" or loadCode == "package"
-    saveFile = FALSE,     ### Save file
+    saveFile = TRUE,      ### Save file
     outPath  = "." |> file.path("report_figures"),  ### Path to save results if saveFile == TRUE
     img_dev  = "pdf",     ### Image device if saveFile == TRUE
     return   = TRUE       ### Whether to return list object
@@ -31,7 +33,7 @@ create_DoW_results <- function(
   # aggOnly  = TRUE    ### Whether to only include sectors for which "includeaggregate==1" in Fig 7 plots
   # loadCode = "project" ### Whether to load code as source or devtools
   # fpath    = "."     ### Path to main FrEDI directory to load code from if loadCode == "project" or loadCode == "package"
-  # saveFile = FALSE   ### Save file
+  # saveFile = TRUE   ### Save file
   # outPath  = "." |> file.path("report_figures")  ### Path to save results if saveFile == TRUE
   # img_dev  = "pdf"   ### Image device if saveFile == TRUE
   # return   = TRUE    ### Whether to return list object
@@ -53,15 +55,32 @@ create_DoW_results <- function(
   codePath        <- projectPath |> file.path("R")
   # projectPath |> list.files() |> print()
   # codePath |> list.files() |> print()
+
   ### Output Paths
-  dowResultsPath  <- outPath |> file.path("DoW")
-  fig7ResultsPath <- outPath |> file.path("fig7")
-  appxResultsPath <- outPath |> file.path("appendix_figures")
+  dowPath  <- outPath |> file.path("DoW")
+  fig7Path <- outPath |> file.path("fig7")
+  appxPath <- outPath |> file.path("appendix_figures")
   # ### Check and create paths
   # outPath |> check_and_create_path()
-  # dowResultsPath  |> check_and_create_path()
-  # fig7ResultsPath |> check_and_create_path()
-  # appxResultsPath |> check_and_create_path()
+  # dowPath  |> check_and_create_path()
+  # fig7Path |> check_and_create_path()
+  # appxPath |> check_and_create_path()
+
+  ### Output file names
+  rda_byType    <- "integer_results_byType"
+  rda_totals    <- "integer_results_totals"
+
+  ### Plots by type
[…]
