### 2024-05-24 afc5907
Edits to report figures action
Edits to try to get report figures action to run successfully
