+++ FrEDI/scripts/create_DoW_results.R
@@ -13,7 +13,7 @@ create_DoW_results <- function(
     gcmYears = c(2090),   ### Which years to report on for GCM sectors
     slrYears = c(2050, 2090), ### Which years to report on for SLR sectors
     byState  = TRUE,      ### Whether values are by state or just by region
-    # totals   = FALSE,     ### Whether to do totals
+    totals   = FALSE,     ### Whether to do totals
     silent   = TRUE,      ### Degree of messaging
     testing  = FALSE,     ### Whether to print out extra diagnostic values
     aggOnly  = TRUE,      ### Whether to only include sectors for which "includeaggregate==1" in Fig 7 plots
@@ -39,27 +39,27 @@ create_DoW_results <- function(
   # return   = TRUE    ### Whether to return list object
   ###### Initial values ######
   ### Messaging
-  do_msg          <- !silent
+  do_msg        <- !silent
   ### Initialize Return List
-  return0         <- return; rm(return)
-  resultsList     <- list()
+  return0       <- return; rm(return)
+  resultsList   <- list()
   ### How to load code
-  loadProject     <- "project" %in% (loadCode |> tolower())
-  loadPackage     <- "package" %in% (loadCode |> tolower())
-  loadSource      <- !loadProject & !loadPackage
+  loadProject   <- "project" %in% (loadCode |> tolower())
+  loadPackage   <- "package" %in% (loadCode |> tolower())
+  loadSource    <- !loadProject & !loadPackage
 
   ###### Set Up Environment ######
   ###### ** Set Paths ######
-  # projectPath      <- getwd()     |> file.path("FrEDI")
-  projectPath     <- fpath; rm(fpath)
-  codePath        <- projectPath |> file.path("R")
+  # projectPath    <- getwd()     |> file.path("FrEDI")
+  projectPath   <- fpath; rm(fpath)
+  codePath      <- projectPath |> file.path("R")
   # projectPath |> list.files() |> print()
   # codePath |> list.files() |> print()
 
   ### Output Paths
-  dowPath  <- outPath |> file.path("DoW")
-  fig7Path <- outPath |> file.path("fig7")
-  appxPath <- outPath |> file.path("appendix_figures")
+  dowPath       <- outPath |> file.path("DoW")
+  fig7Path      <- outPath |> file.path("fig7")
+  appxPath      <- outPath |> file.path("appendix_figures")
   # ### Check and create paths
   # outPath |> check_and_create_path()
   # dowPath  |> check_and_create_path()
@@ -172,18 +172,26 @@ create_DoW_results <- function(
   ###### ** Inputs List ######
   ### Message
   if(testing|do_msg) "Creating tibble of integer scenarios..." |> print()
-  ### Load scenario inputs
-  inputs_df_int  <- list(
-    x = df_scenarios[["temp_C"  ]],
-    y = df_scenarios[["tempType"]],
-    z = df_scenarios[["prefix"  ]]
-  )
+  # ### Load scenario inputs
+  # inputs_df_int  <- list(
+  #   x = df_scenarios[["temp_C"  ]],
+  #   y = df_scenarios[["tempType"]],
+  #   z = df_scenarios[["prefix"  ]]
+  # )
+  # ### Create constant temp scenarios
+  # inputs_df_int  <- inputs_df_int |> pmap(function(x, y, z){
+  #   create_constant_temp_scenario(
+  #     temp0   = x,
+  #     type0   = y,
+  #     prefix0 = z ### Prefix for scenario
+  #   )
+  # }) |> bind_rows()
   ### Create constant temp scenarios
[…]
