+++ FrEDI/testing/utils_create_report_figures.R
@@ -115,45 +115,45 @@ format_values <- function(
 ### Run CONUS scenarios
 create_constant_temp_scenario <- function(
     temp0,
-    type0   = "conus",
-    prefix0 = "Other_Integer" ### Prefix for scenario
+    type0 = "conus",
+    scen0 = "Other_Integer" |> paste(temp0 |> round(1), sep="_") ### Prefix for scenario
 ){
   ### Temperature Type
   isConus <- "conus" %in% (type0 |> tolower())
   ### Format scenario label
-  # pre0  <- (type0=="conus") |> ifelse("Other_Integer", "preI_global")
-  pre0  <- prefix0
-  lab0  <- temp0 |> round(1)
-  scen0 <- pre0  |> paste(lab0, sep="_")
+  # pre0    <- prefix0
+  # lab0    <- temp0 |> round(1)
+  # scen0   <- pre0  |> paste(lab0, sep="_")
+  lab0    <- scen0 |> str_extract("(\\d+)(\\.)(\\d+)")
   ### Get annual values 1995 - 2010: starting with zero in 1995
-  xIn0  <- c(1995, 2010)
-  yIn0  <- c(0, temp0)
-  xOut0 <- seq(xIn0[1], xIn0[2])
-  df0   <- approx(x = xIn0, y = yIn0, xout=xOut0) |>
+  xIn0    <- c(1995, 2010)
+  yIn0    <- c(0, temp0)
+  xOut0   <- seq(xIn0[1], xIn0[2])
+  df0     <- approx(x=xIn0, y=yIn0, xout=xOut0) |>
     as_tibble() |>
     rename(year=x, temp_C=y)
   ### Extend values
-  df1   <- tibble(year = seq(2011, 2090, by=1))
-  df1   <- df1 |> mutate(temp_C  = temp0)
-  df0   <- df0 |> rbind(df1)
-  rm("df1")
+  df1     <- tibble(year = seq(2011, 2090, by=1))
+  df1     <- df1 |> mutate(temp_C = temp0)
+  df0     <- df0 |> rbind(df1)
+  rm(df1)
 
   ### Get other temp types and rename
   if(isConus){
-    df0   <- df0 |> mutate(temp_C_global = temp_C |> FrEDI::convertTemps(from="conus"))
-    df0   <- df0 |> mutate(temp_C_conus  = temp_C)
+    df0 <- df0 |> mutate(temp_C_global = temp_C |> FrEDI::convertTemps(from="conus"))
+    df0 <- df0 |> mutate(temp_C_conus  = temp_C)
   } ### End if(isConus)
   else       {
-    df0   <- df0 |> mutate(temp_C_conus  = temp_C |> FrEDI::convertTemps(from="global"))
-    df0   <- df0 |> mutate(temp_C_global = temp_C)
+    df0 <- df0 |> mutate(temp_C_conus  = temp_C |> FrEDI::convertTemps(from="global"))
+    df0 <- df0 |> mutate(temp_C_global = temp_C)
   } ### End else(isConus)
   ### Drop temp_C
-  df0   <- df0 |> select(-c("temp_C"))
+  df0     <- df0 |> select(-c("temp_C"))
   ### Get SLR
-  ySlr0 <- FrEDI::temps2slr(temps = df0[["temp_C_global"]], years = df0[["year"]])
-  df0   <- df0 |> left_join(ySlr0, by="year")
-  df0   <- df0 |> mutate(temp_lab = lab0)
-  df0   <- df0 |> mutate(scenario = scen0)
+  ySlr0   <- FrEDI::temps2slr(temps = df0[["temp_C_global"]], years = df0[["year"]])
+  df0     <- df0 |> left_join(ySlr0, by="year")
[…]

[…]
