### 2024-05-24 9e74c84
Update test_fredi.yml
Trying to fix yaml error
