+++ .github/workflows/test_fredi.yml
@@ -47,7 +47,7 @@ jobs:
 
       - name: Send input status
         run: |
-          echo "${{ inputs.run_tests }} ${{ inputs.make_figures }}"
+          echo "${{ inputs.run_tests }} ${{ inputs.make_figures }} ${{ inputs.figH_results }}"
 
       - name: Setup R
         uses: r-lib/actions/setup-r@v2
@@ -76,6 +76,8 @@ jobs:
               type    = "source"
             )'
 
+    name: Run tests and make figures
+    steps:
       - name: Set paths and load FrEDI
         run: |
           Rscript -e '
