### 2024-05-24 c1c8ff2
Update test_fredi.yml
Edits to fix actions
