+++ .github/workflows/test_fredi.yml
@@ -47,7 +47,7 @@ jobs:
 
       - name: Send input status
         run: |
-          echo "${{ inputs.run_tests }} ${{ inputs.make_figures }} ${{ inputs.figH_results }}"
+          echo "${{ inputs.run_tests }} ${{ inputs.figH_results }} ${{ inputs.make_figures }}"
 
       - name: Setup R
         uses: r-lib/actions/setup-r@v2
@@ -76,8 +76,6 @@ jobs:
               type    = "source"
             )'
 
-    name: Run tests and make figures
-    steps:
       - name: Set paths and load FrEDI
         run: |
           Rscript -e '
@@ -110,9 +108,9 @@ jobs:
           exists2  <- oPath2 |> dir.exists()
 
           ### Check if path exists and, if not, create it
-          if(!exists0) oPath0 |> dir.create(recursive = TRUE)
-          if(!exists1) oPath1 |> dir.create(recursive = TRUE)
-          if(!exists2) oPath2 |> dir.create(recursive = TRUE)
+          if(!exists0){ oPath0 |> dir.create(recursive = TRUE)}
+          if(!exists1){ oPath1 |> dir.create(recursive = TRUE)}
+          if(!exists2){ oPath2 |> dir.create(recursive = TRUE)}
           '
 
       - name: Run tests
@@ -150,18 +148,18 @@ jobs:
             ./FrEDI/data_tests/general/defaultScenarioTotals.rda
             ./FrEDI/data_tests/general/testResults_fredi_general.xlsx
 
-       - name: Get FigH results
-         run: |
-           Rscript -e '
-           ### Create report results
-           if(do_figH){
-            ### Figure H results
-            results0 <- run_fredi(aggLevels=c("modelaverage", "national", "impactYear"))
-            results0 <- results0 |> filter(year %in% seq(2010, 2090, by=5))
-            save0    <- results0 |> write.csv(file=oPath2 |> file.path("FigH_results.csv"), row.names=F) |> try()
+      - name: Get FigH results
+        run: |
+          Rscript -e '
+          ### Create report results
+          if(do_figH){
+          ### Figure H results
+          results0 <- run_fredi(aggLevels=c("modelaverage", "national", "impactYear"))
+          results0 <- results0 |> filter(year %in% seq(2010, 2090, by=5))
+          save0    <- results0 |> write.csv(file=oPath2 |> file.path("FigH_results.csv"), row.names=F) |> try()
 
-            ### Remove reslts
-            rm(results0, save0)
+          ### Remove reslts
+          rm(results0, save0)
           }
           '
 
