### 2024-05-24 4c0c2b3
Update test_fredi.yml
Trying to fix action
