### 2024-05-24 af9d760
Update test_fredi.yml
More fixes to test workflow
