### 2024-05-24 7cd426e
Update test_fredi.yml
More fixes
