### 2024-05-24 7c3e7d4
Edited workflow
Edited workflow to better use github resources
