+++ FrEDI/scripts/create_DoW_results.R
@@ -204,24 +204,29 @@ create_DoW_results <- function(
   ###### Run scenarios to get results by type
   ### Run scenarios in FrEDI. Get model averages and national totals
   if(testing|do_msg) "Running integer scenarios..." |> message()
+  aggLvls0       <- c("modelaverage", "national") |> c("impactyear", "impacttype")
+  if(totals) aggLvls0 <- aggLvls0 |> c()
   df_int_byType  <- inputs_df_int |> run_scenarios(
     col0      = "scenario",
     fredi     = TRUE,
     sectors   = sectors,
-    aggLevels = c("modelaverage", "national"),
+    # aggLevels = c("modelaverage", "national"),
+    aggLevels = aggLvls0,
     scenCols  = c("scenario", "year", "temp_C_conus", "temp_C_global", "slr_cm"),
     joinCols  = c("year"),
     return    = TRUE,
     save      = FALSE,
     outPath   = dowPath
   )
+  rm(aggLvls0)
 
   ### Glimpse results
   # if(return0) resultsList[["df_int_byType"]] <- df_int_byType
   if(testing) df_int_byType |> glimpse()
   ### Save results
   if(do_msg & saveFile) paste0("Saving integer scenario results by type...") |> message()
-  if(saveFile) df_int_byType |> save_data(fpath=dowPath, fname=rda_byType, ftype="rda")
+  if(saveFile & (!totals)) df_int_byType |> save_data(fpath=dowPath, fname=rda_byType, ftype="rda")
+  if(saveFile &   totals ) df_int_byType |> save_data(fpath=dowPath, fname=rda_totals, ftype="rda")
   # for(scenario_i in cScenarios){
   #   ### Run scenario
   #   df_i <- scenario_i |> run_scenario(
@@ -240,153 +245,158 @@ create_DoW_results <- function(
 
   ###### ** Appendix Figs: DoW By Type ######
   # codePath  |> loadCustomFunctions()
-  if(testing|do_msg) "Summarizing GCM results by sector, impact type, degree of warming (DOW)..." |> message()
-  sum_gcm_byType <- df_int_byType |> sum_impacts_byDoW_years(
-    scenarios   = c_scen_con,
-    bySector    = TRUE,
-    sumCol      = "annual_impacts",
-    impactYears = c("NA", "2010", "2090"),
-    models      = c("GCM"),
-    adjVal      = 1/10**9, ### Factor to multiply by
-    adjCol      = "impact_billions",
-    silent      = TRUE
-  )
-  ### Glimpse
-  # if(return0) resultsList[["sum_gcm_byType"]] <- sum_gcm_byType
-  if(testing) sum_gcm_byType |> glimpse()
-
-  ### Save summary table
-  if(do_msg & saveFile) paste0("Saving summary of GCM results by sector, impact type, degree of warming...") |> message()
-  if(saveFile) sum_gcm_byType |> save_data(fpath=appxPath, fname=csv_appx_gcm, ftype="csv", row.names=F)
-
-
-  ### Create Plots
-  # codePath  |> loadCustomFunctions()
-  if(testing|do_msg) "Plotting GCM results by sector, impact type, degree of warming (DOW)..." |> message()
-  plots_gcm_byType <- sum_gcm_byType |>
-    # filter(sector %in% c_sectorNames[c(10)]) |>
-    filter(!(sector %in% c("Roads"))) |>
-    plot_DoW_by_sector(
-      models  = c("GCM"),
-      yCol    = "annual_impacts"
+  if(!totals){
+    if(testing|do_msg) "Summarizing GCM results by sector, impact type, degree of warming (DOW)..." |> message()
[…]
