### 2024-05-24 0d3e65f
Update create_DoW_results.R
Fixed error in calculating by impact (appendix) results
