### 2024-05-24 7342a6f
Update test_fredi.yml
Fixed code for uploading artifacts for report figures
