### 2024-06-09 55202ee
Add population downscaling functionality to import_inputs
Have not yet updated documentation. If new sysdata is configured and transferred over from the new branch of the same name on FrEDI_Data, this current iteration should work for "state", "regional", and "conus" values of the new popArea input assuming column names and values are as expected (see comments in code here). To get "national" functionality working, the sysdata will also need to include a df with cols "year" and "nation_to_conus" (ratio of conus/full US) saved at rDataList[["stateData"]][["data"]][["df_conusRatios"]] (or can up
[…]
