### 2024-06-11 785c7f7
Updated documentation for state-level
Updated documentation for FrEDI functions that indicate that `run_fredi` would output results at either the region or state-level. Now, only refers to the state-level.
