+++ FrEDI/R/import_inputs.R
@@ -110,7 +110,7 @@ import_inputs <- function(
   ### Get objects from FrEDI name space
   ### Get input scenario info: co_inputScenarioInfo
   ### Get state info: co_states
-  fListNames <- c("co_inputScenarioInfo", "co_states", "df_popRatios")
+  fListNames <- c("co_inputScenarioInfo", "co_states")
   sListNames <- c("df_popRatios")
   for(name_i in fListNames){name_i |> assign(rDataList[["frediData"]][["data"]][[name_i]]); rm(name_i)}
   for(name_i in sListNames){name_i |> assign(rDataList[["stateData"]][["data"]][[name_i]]); rm(name_i)}
@@ -148,34 +148,15 @@ import_inputs <- function(
 
 
   ###### Other Values ######
-  ### Create a list with expected name of column containing values
-  valCols    <- list()
-  valCols[["temp"]] <- c("temp_C")
-  valCols[["slr" ]] <- c("slr_cm")
-  valCols[["gdp" ]] <- c("gdp_usd")
-  valCols[["pop" ]] <- list()
-  valCols[["pop" ]][["national"]] <- c("national_pop")
-  valCols[["pop" ]][["conus"   ]] <- c("area_pop")
-  valCols[["pop" ]][["regional"]] <- c("reg_pop")
-  valCols[["pop" ]][["state"   ]] <- c("state_pop")
-  valCols[["pop" ]] <- valCols[["pop"]][[popArea]]
-
-  ### Create a list with expected name of columns used for unique ids
-  idCols     <- list()
-  idCols[["temp"]] <- c()
-  idCols[["slr" ]] <- c()
-  idCols[["gdp" ]] <- c()
-  idCols[["pop" ]] <- list()
-  idCols[["pop" ]][["national"]] <- c()
-  idCols[["pop" ]][["conus"   ]] <- c()
-  idCols[["pop" ]][["regional"]] <- c("region")
-  idCols[["pop" ]][["state"   ]] <- c("state")
-  idCols[["pop" ]] <- idCols[["pop"]][[popArea]]
+  ### Get list with expected name of column containing values
+  ### Get list with expected name of columns used for unique ids
+  valCols    <- get_import_inputs_valCols(popArea=popArea)
+  idCols     <- get_import_inputs_idCols (popArea=popArea)
+  # valCols |> unlist() |> print(); idCols  |> unlist() |> print()
 
   ###### Initialize Input File Names List ######
   ### Initialize list of file names and
-  fList      <- list()
-  for(name_i in inNames){fList[[name_i]] <- parse(text=name_i |> paste0("file")) |> eval(); rm(name_i)}
+  fList      <- inNames |> map(function(x){parse(text=x |> paste0("file")) |> eval()}) |> set_names(inNames)
 
   ###### Initialize Inputs List ######
   ### Initialize a list for loaded data
@@ -189,216 +170,43 @@ import_inputs <- function(
 
   ### If some file names supplied, iterate over list, trying to read in file names
   inputsList <- list(
-    name_i = inputNames,
-    file_i = inputsList
-  ) |> pmap(function(name_i, file_i){
-    ### Messages
-    msg_i1 <- paste0("User specified ", name_i |> paste0("file"), "...")
-    msg_i2 <- paste0("File does not exist! Returning a null data frame for", name_i |> paste0("Input", "."))
-    msg_i3 <- paste0("Importing data from ", file_i, "...")
-
-    ### Initialize data frame as a null value
-    df_i   <- NULL
-
-    ### Check if a file name was supplied
-    ### - If no file supplied, return NULL
-    ### - Otherwise, continue
-    has_i  <- file_i |> length()
-    if(!has_i) return(df_i)
-
-    ### If a file name was supplied, message user
[…]
