+++ FrEDI/R/utils_import_inputs.R
@@ -1,10 +1,65 @@
+###### get_msg_prefix ######
+### Function to get message prefix
+get_msg_prefix <- function(
+    level  = 0,
+    prefix = ""
+){
+  msgP    <- "\t"   |> rep(level) |> paste(collapse="")
+  msgP    <- prefix |> paste0(msgP)
+  return(msgP)
+}
+
+###### get_import_inputs_valCols ######
+### 2024.06.18: eventually move to FrEDI Data
+### Info about which column contains input values
+get_import_inputs_valCols <- function(
+    popArea = "state" ### One of: c("state", "regional", "area", "national")
+){
+  ### Initialize list
+  list0    <- list()
+  list0[["temp"]] <- c("temp_C")
+  list0[["slr" ]] <- c("slr_cm")
+  list0[["gdp" ]] <- c("gdp_usd")
+  list0[["pop" ]] <- list()
+  list0[["pop" ]][["national"]] <- c("national_pop")
+  list0[["pop" ]][["conus"   ]] <- c("area_pop")
+  list0[["pop" ]][["regional"]] <- c("reg_pop")
+  list0[["pop" ]][["state"   ]] <- c("state_pop")
+  ### Update population value column based on pop area
+  list0[["pop" ]] <- list0[["pop"]][[popArea]]
+  ### Return
+  return(list0)
+}
+
+###### get_import_inputs_valCols ######
+### 2024.06.18: eventually move to FrEDI Data
+### Info about which column contains info on region
+get_import_inputs_idCols <- function(
+    popArea = "state" ### One of: c("state", "regional", "area", "national")
+){
+  ### Initialize list
+  list0    <- list()
+  list0[["temp"]] <- c("year")
+  list0[["slr" ]] <- c("year")
+  list0[["gdp" ]] <- c("year")
+  list0[["pop" ]] <- list()
+  list0[["pop" ]][["national"]] <- c("year")
+  list0[["pop" ]][["conus"   ]] <- c("year")
+  list0[["pop" ]][["regional"]] <- c("region", "year")
+  list0[["pop" ]][["state"   ]] <- c("state", "year")
+  ### Update population value column based on pop area
+  list0[["pop" ]] <- list0[["pop"]][[popArea]]
+  ### Return
+  return(list0)
+}
+
 ###### fun_tryInput ######
 ### Created 2021.02.08. Last updated 2022.05.09
 ### This function attempts to load a user-specified input file
 fun_tryInput <- function(
-  filename = NULL,
-  silent   = FALSE, ### Whether to message
-  msg0     = ""     ### Message prefix
+    filename = NULL,
+    silent   = FALSE, ### Whether to message
+    msg0     = ""     ### Message prefix
 ){
   ###### Messaging ######
   msgUser <- !silent
@@ -52,26 +107,82 @@ fun_tryInput <- function(
 }
 
 
-###### rename_inputs ######
-### Created 2021.02.08. Last updated 2021.02.08
-rename_inputs <- function(
-  data,
-  new_names
+###### run_fun_tryInput ######
+### Function to iterate over a list of file names
+run_fun_tryInput <- function(
+    inputName = "temp",      ### Type of input; one of: c("temp", "slr", "gdp", "pop")
+    fileName  = "temps.csv", ### File path and name of CSV to be imported
+    msgLevel  = 1
 ){
-  ### Get the length of the new names
-  data_names   <- data |> names()
-  num_names    <- new_names |> length()
-  num_dataCols <- data |> ncol()
+  ### Messages
+  msgN   <- "\n"
+  msg0_i <- get_msg_prefix(level=msgLevel)
+  msg1_i <- get_msg_prefix(level=msgLevel + 1)
[…]
