### 2024-06-25 055a5ee
Update run_fredi.R
Changed default max year to 2100,  added checks to return if missing scenario values in and after max year
