### 2024-06-26 d37a9ef
Update run_fredi_sv.R
Updated FrEDI SV to run to 2100
