### 2024-06-26 d409205
Removed SV population scenario
Removed SV-specific population scenario from FrEDI data
