### 2024-06-26 f0dbd79
Update run_fredi.R
Added lines to make sure that the population inputs would have the correct columns. Also fixed a type in the documentation
