+++ FrEDI/R/utils_sv.R
@@ -103,56 +103,118 @@ get_sv_sectorInfo <- function(
 ### Created 2022.02.14. Last updated 2022.02.14
 ### This function attempts to load a user-specified input file
 ### Use this function to calculate tract level impacts
+# calc_countyPop <- function(
+    #     regPop,  ### Dataframe of population projection
+#     funList, ### Dataframe of population projections
+#     years = seq(2010, 2100, by=5)
+#
+# ){
+#   c_regions <- regPop$region |> unique()
+#   ### Iterate over regions
+#   x_popProj <- c_regions |> map(function(region_i){
+#     ### Subset population projection to a specific region
+#     ### Get states in the region
+#     ### Get unique years
+#     df_i     <- regPop |> filter(region==region_i)
+#     states_i <- funList[[region_i]] |> names()
+#     years_i  <- df_i$year |> unique() |> sort()
+#
+#     ### Iterate over states
+#     regionPop_i <- states_i |> map(function(state_j){
+#       ### Function for state j
+#       fun_j   <- funList[[region_i]][[state_j]]$state2region
+#
+#       # state_j |> print()
+#       df_j <- tibble(x = years_i, y = fun_j(years_i))
+#       df_j <- df_j |> rename(year  = x, ratioState2RegionPop = y)
+#       df_j <- df_j |> mutate(state = state_j, region  = region_i)
+#
+#       ### Get list of counties in the state
+#       geoids_j      <- funList[[region_i]][[state_j]]$county2state |> names()
+#       statePop_j    <- geoids_j |> map(function(geoid_k){
+#         fun_k <- funList[[region_i]][[state_j]]$county2state[[geoid_k]]
+#         if(!is.null(fun_k)) {y_k <- fun_k(years_i)} else{y_k <- NA}
+#         df_k  <- tibble(year = years_i, ratioCounty2StatePop = y_k)
+#         df_k  <- df_k |> mutate(state = state_j, geoid10 = geoid_k)
+#         return(df_k)
+#       }) |> bind_rows()
+#       ### Join with state population
+#       df_j <- df_j |> left_join(statePop_j, by = c("state", "year"))
+#       return(df_j)
+#     }) |> bind_rows()
+#     ### Join with regional population
+#     df_i <- df_i |> left_join(regionPop_i, by = c("region", "year"))
+#     df_i <- df_i |> mutate(state_pop  = region_pop * ratioState2RegionPop)
+#     df_i <- df_i |> mutate(county_pop = state_pop  * ratioCounty2StatePop)
+#     ### Return
+#     return(df_i)
+#   }) |> bind_rows()
+#   ### Return
+#   # x_popProj <- x_popProj |> as.data.frame()
+#   return(x_popProj)
+# }
 calc_countyPop <- function(
-    regPop,  ### Dataframe of population projection
-    funList, ### Dataframe of population projections
-    years = seq(2010, 2090, by=5)
-
+    fun0 = NULL, ### Function from list svPopList[["popProjList"]]
+    df0,  ### Tibble with population projections
+    xCol0 = "year",      ### X column in df0
+    yCol0 = "state_pop"  ### Y column in df0
 ){
-  c_regions <- regPop$region |> unique()
-  ### Iterate over regions
-  x_popProj <- c_regions |> map(function(region_i){
-    ### Subset population projection to a specific region
-    ### Get states in the region
-    ### Get unique years
-    df_i     <- regPop |> filter(region==region_i)
[…]
