### 2024-06-26 7dea46b
Improved efficiency of SV module
Edited code to improve efficiency of SV module. Mainly changed map functions to for loops. Also removed impact list from environment earlier in the process
