+++ FrEDI/R/run_fredi_sv.R
@@ -172,16 +172,22 @@ run_fredi_sv <- function(
   c_sector        <- sector
   paste0("Running FrEDI SV for sector '", c_sector, "':") |> message()
   ### Sector info
-  which_sector    <- (svSectorInfo$sector == c_sector) |> which() #; which_sector |> print()
-  df_sectorInfo   <- svSectorInfo[which_sector,]
-  c_variants      <- df_sectorInfo[["variant_abbr" ]][which_sector]
-  c_variantLabels <- df_sectorInfo[["variant_label"]][which_sector]
-  rm(which_sector)
+  # which_sector    <- (svSectorInfo$sector == c_sector) |> which() #; which_sector |> print()
+  # df_sectorInfo   <- svSectorInfo[which_sector,]
+  # c_variants      <- df_sectorInfo[["variant_abbr" ]][which_sector]
+  # c_variantLabels <- df_sectorInfo[["variant_label"]][which_sector]
+  # rm(which_sector)
+  df_sectorInfo   <- svSectorInfo  |> filter(sector == c_sector)
+  c_variants      <- df_sectorInfo |> pull(variant_abbr)
+  c_variantLabels <- df_sectorInfo |> pull(variant_label)
+
   ###### Invalid Sectors ######
-  which_sector    <- (sectorInfo$sector == c_sector) |> which() #; which_sector |> print()
-  c_popWtCol      <- sectorInfo[["popWeightCol"]][which_sector] |> tolower()
-  c_modelType     <- sectorInfo[["modelType"   ]][which_sector] |> tolower()
-  rm(which_sector)
+  # which_sector    <- (sectorInfo$sector == c_sector) |> which() #; which_sector |> print()
+  # c_popWtCol      <- sectorInfo[["popWeightCol"]][which_sector] |> tolower()
+  # c_modelType     <- sectorInfo[["modelType"   ]][which_sector] |> tolower()
+  # rm(which_sector)
+  c_popWtCol      <- sectorInfo |> filter(sector == c_sector) |> pull(popWeightCol) |> tolower()
+  c_modelType     <- sectorInfo |> filter(sector == c_sector) |> pull(modelType) |> tolower()
   df_validGroups  <- svDemoInfo |> get_validGroups(df1 = svValidTypes, col0 = c_popWtCol)
   # return(df_validGroups)
 
@@ -639,65 +645,183 @@ run_fredi_sv <- function(
   msg2 |> message("Calculating county population from regional population...")
   df_popProj <- pop_df |> get_countyPop(
     years   = yearsBy5,
-    xCol0   = "year",       ### X column in df0
-    yCol0   = "state_pop",  ### Y column in df0
-    years   = yearsBy5
+    xCol0   = "year",     ### X column in df0
+    yCol0   = "state_pop" ### Y column in df0
   ) ### End get_countyPop
 
   ###### Calculate Impacts ######
+  # ### Iterate over adaptations/variants
+  # df_results  <- df_sectorInfo |> nrow() |> seq_len() |> map(function(
+  #   row_i,
+  #   info_x      = df_sectorInfo,
+  #   scenarios_x = c_scenarios
+  # ){
+  #   ### Which SV data to use
+  #   svName_i       <- (c_sector=="Coastal Properties") |> ifelse("svDataCoastal", "svData");
+  #   # scenarios_x |> print(); # svName_i |> print()
+  #
+  #   ### Sector info
+  #   info_i         <- info_x[row_i,]
+  #   sectorAbbr_i   <- info_i$impactList_fileExt[1]
+  #   varLabel_i <- info_i$variant_label[1]
+  #   varAbbr_i  <- info_i$variant_abbr[1]
+  #   weightsCol_i   <- info_i$popWeightCol[1]
+  #   # info_i |> print()
+  #
[…]
