### 2024-06-26 978ad20
Update utils_sv.R
Deleted deprecated code. Updated `get_sv_sectorInfo`
