+++ FrEDI/R/utils_sv.R
@@ -49,113 +49,58 @@
 #' @md
 #'
 ###
-###
 get_sv_sectorInfo <- function(
     description = F,
     gcmOnly     = F,
     slrOnly     = F
 ){
-  if(description |> is.null()){description <- F}
-  if(gcmOnly     |> is.null()){gcmOnly     <- F}
-  if(slrOnly     |> is.null()){slrOnly     <- F}
-  #svDataList <- load(file = "C:/Users/wmaddock/Documents/GitHub/FrEDI/createSystemData/data/sv/svDataList.rda")
-  # co_sectorsRef$sector_label
-  cData <- c("sectorInfo", "svSectorInfo")
-  for(item_i in cData){assign(item_i, svDataList[[item_i]])}
-  # print(sectorInfo)
-  select0 <- c("sector", "modelType", "impactUnit")
-  select1 <- c("sector", "modelType", "driverUnit", "variants", "impactUnit")
+  ### Get objects
+  cData        <- c("sectorInfo", "svSectorInfo")
+  for(item_i in cData){item_i |> assign(svDataList[[item_i]])}
 
-  ### - Select appropriate columns
-  ### - Format modelType as uppercase
-  ### - Add variant label
-  ### - Add driver value
-  sectorInfo <- sectorInfo |> select(all_of(select0))
-  sectorInfo <- sectorInfo |> mutate(modelType  = modelType |> toupper())
-  sectorInfo <- sectorInfo |> mutate(
-    variants = sector |> map(function(sector_i){
-      variants_i <- (svSectorInfo |> filter(sector==sector_i))$variant_label |> paste(collapse=", ")
-      return(variants_i)
-    }) |> unlist())
-  sectorInfo <- sectorInfo |> mutate(driverUnit=ifelse(modelType=="GCM", "degrees Celsius", "cm"))
+  ### Select appropriate columns
+  ### Format modelType as uppercase
+  select0      <- c("sector", "impactUnit")
+  sectorInfo   <- sectorInfo |> select(all_of(select0))
+  rm(select0)
+
   ### Select and arrange
-  sectorInfo <- sectorInfo |> select(all_of(select1))
-  sectorInfo <- sectorInfo |> arrange_at(vars("sector"))
+  group0       <- c("sector", "modelType", "driverUnit")
+  rename0      <- c("driverUnit_label", "variant_label")
+  renameTo     <- c("driverUnit", "variants")
+  svSectorInfo <- svSectorInfo |> rename_at(c(rename0), ~renameTo)
+  svSectorInfo <- svSectorInfo |> mutate(modelType = modelType |> toupper())
+  svSectorInfo <- svSectorInfo |>
+    group_by_at(c(group0)) |>
+    summarize(variants = variants |> paste(collapse=", "), .groups="drop")
+
+  ### Join with sector info
+  join0      <- "sector"
+  move0      <- c("modelType", "driverUnit")
+  sectorInfo <- sectorInfo |> left_join(svSectorInfo, by=c(join0))
+  sectorInfo <- sectorInfo |> relocate(all_of(move0), .after=all_of(join0))
+  sectorInfo <- sectorInfo |> arrange_at(vars(join0))
+
   ### GCM or SLR
   doFilter   <- (gcmOnly | slrOnly)
-  if(doFilter){
-    cFilter    <- gcmOnly
-    gcm_string <- "GCM"
-    sectorInfo <- sectorInfo |> mutate(is_gcm =  modelType==gcm_string)
-    sectorInfo <- sectorInfo |> filter(is_gcm == cFilter)
-    sectorInfo <- sectorInfo |> select(-c("is_gcm"))
-  }
[…]
