### 2024-06-26 e81bbc0
Update run_fredi_sv.R
Deleted deprecated lines. Fixed messaging. Changed loading and removing of impacts list.
