+++ FrEDI/R/run_fredi_sv.R
@@ -102,8 +102,6 @@ run_fredi_sv <- function(
     popInput    = NULL,
     silent      = TRUE,  ### Whether to message the user
     .testing    = FALSE
-    # addDate     = FALSE, ### Whether to add the date to the file name
-    # libPath     = .libPaths()[1]
 ){
   ###### Set up the environment ######
   pkgPath     <- NULL
@@ -119,11 +117,10 @@ run_fredi_sv <- function(
 
   ### Assign config files
   fredi_config |> list2env(envir = environment())
-  # for(name_i in names(fredi_config)){ name_i |> assign(fredi_config[[name_i]])}
 
   ### Group types
   c_svGroupTypes <- svDataList$c_svGroupTypes
-  # minYear <- 2010
+  minYear  <- 2010
   maxYear  <- 2100
   yearsBy5 <- minYear |> seq(maxYear, by=5)
 
@@ -145,25 +142,22 @@ run_fredi_sv <- function(
 
 
   ###### Sector Info ######
-  ### Objects <-
   sectorInfo    <- svDataList$sectorInfo
   svSectorInfo  <- svDataList$svSectorInfo
   svDemoInfo    <- svDataList$svDemoInfo
   svValidTypes  <- svDataList$svValidTypes
   co_formatting <- svDataList$co_formatting
   co_formatting <- svDataList$co_formatting
-  # co_modelTypes <- rDataList$co_modelTypes
+
   ### Sector names
   hasSector <- !(sector |> is.null())
   if(!hasSector){
     sector_msg1   <- paste0("Please select a sector: ") |> print()
-    sector_msg2   <- 1:nrow(sectorInfo) |> lapply(function(i){
+    sector_msg2   <- sectorInfo |> nrow() |> seq_len() |> map(function(i){
       sector_i    <- sectorInfo$sector[i]
       msg_i       <- paste0(i, ". ", sector_i)
       msg_i |> print()
-      # return(msg_i)
-    }) #|> unlist() |> paste(collapse="")
-    # sector_msg3  <- sector_msg1 |> paste0(sector_msg2)
+    }) ### End map(function(i)
     sector_msg3   <- "Enter a number:"
     sector_input  <- readline(prompt = sector_msg3) |> as.numeric()
     sector        <- sectorInfo$sector[sector_input]
@@ -171,25 +165,14 @@ run_fredi_sv <- function(
   } ### End if(hasSector)
   c_sector        <- sector
   paste0("Running FrEDI SV for sector '", c_sector, "':") |> message()
+
   ### Sector info
-  # which_sector    <- (svSectorInfo$sector == c_sector) |> which() #; which_sector |> print()
-  # df_sectorInfo   <- svSectorInfo[which_sector,]
-  # c_variants      <- df_sectorInfo[["variant_abbr" ]][which_sector]
-  # c_variantLabels <- df_sectorInfo[["variant_label"]][which_sector]
-  # rm(which_sector)
   df_sectorInfo   <- svSectorInfo  |> filter(sector == c_sector)
   c_variants      <- df_sectorInfo |> pull(variant_abbr)
   c_variantLabels <- df_sectorInfo |> pull(variant_label)
-
-  ###### Invalid Sectors ######
-  # which_sector    <- (sectorInfo$sector == c_sector) |> which() #; which_sector |> print()
-  # c_popWtCol      <- sectorInfo[["popWeightCol"]][which_sector] |> tolower()
-  # c_modelType     <- sectorInfo[["modelType"   ]][which_sector] |> tolower()
-  # rm(which_sector)
   c_popWtCol      <- sectorInfo |> filter(sector == c_sector) |> pull(popWeightCol) |> tolower()
[…]
