### 2024-06-27 674e1c2
Update temps2slr.R
Edited temps2slr to use the series as the basis for calculating years (was calculating values out to 2300, even when only values out to 2100 supplied). Also changed approximation rule to 1...years outside of range will produce NAs
