+++ FrEDI/R/temps2slr.R
@@ -65,8 +65,10 @@
 ###
 
 temps2slr <- function(
-  temps, years
+    temps, years
 ){
+  ###### Messages ######
+  msg1    <- "\t"
   ###### Kopp Equations ######
   ### dh/dt = a*(T(t) - Te(t)) + c(t)
   ### dTe(t)/dt = (T(t) - Te(t))/tau1
@@ -88,74 +90,49 @@ temps2slr <- function(
   ### Assign reference year from config file (max_year)
   fredi_config        <- "fredi_config" |> get_frediDataObj("frediData")
   temps2slr_constants <- fredi_config[["temps2slr"]]
-  names_slrConstants  <- temps2slr_constants |> names()
-  for(name_i in names_slrConstants){assign(name_i, temps2slr_constants[[name_i]])}
+  temps2slr_constants |> list2env(envir = environment())
+
   #### Reference year is 2000
-  co_modelTypes <- rDataList[["frediData"]][["data"]][["co_modelTypes"]] |> filter(modelType_id == "slr")
-  ref_year0     <- co_modelTypes$modelRefYear[1]
-  # year0         <- ref_year0
-  eqtemp_offset <- 0.62
+  co_modelTypes <- "co_modelTypes" |> get_frediDataObj("frediData") |> filter(modelType_id == "slr")
+  ref_year0     <- co_modelTypes |> pull(modelRefYear)
 
   ###### Other constants ######
+  eqtemp_offset <- 0.62
   mm2cm         <- 0.1 ### Number of centimeters per millimeter
 
-  # We used HadCrUT4 to determine the appropriate temperature offset between the actual temperature and the equilibrium temperature in 2000.
-  # Met Office Hadley Centre observations datasets. https://www.metoffice.gov.uk/hadobs/hadcrut4/data/current/download.html.
-
-  # ###### Standardize data #####
-  # ### Filter to years of interest 2000-2090
-  # # max_year     <- 2090
-  # new_years    <- seq(year0, max_year)
-  # num_x        <- new_years |> length()
-  # ind_x        <- 1:num_x
-
-  # ### Initialize Data
-  # df_x0 <- tibble(year = years, temp_C = temps) |>
-  #   filter(year >= year0) |>
-  #   filter(year <= max_year)
-  #
-  #
-  #   # pull out 2000 data
-  # temp_C0     <- (df_x0 |> filter(year == 2000))$temp_C[1]
-  #
-  # ###To-do exit gracefully within tempbin()
-  # if(is.na(temp_C0)) {
-  #   message("Error in 'temps2slr()':")
-  #   message("\t", "Enter a temperature series starting in 2000.")
-  #   message("Exiting...")
-  #   return()
-  # }
-
+  ### We used HadCrUT4 to determine the appropriate temperature offset between the actual temperature and the equilibrium temperature in 2000.
+  ### Met Office Hadley Centre observations datasets. https://www.metoffice.gov.uk/hadobs/hadcrut4/data/current/download.html.
   ###### Initialize Data ######
   ### Filter NA values and make sure values are numeric
-  df_x0 <- tibble(year = years, temp_C = temps) |>
-    mutate_at(c("year", "temp_C"), as.character) |>
-    mutate_at(c("year", "temp_C"), as.numeric) |>
-    filter(!is.na(year) & !is.na(temp_C)) |>
-    arrange_at(.vars=c("year"))
+  mutate0 <- c("year", "temp_C")
[…]
