### 2024-06-27 3b2b566
Update run_fredi_sv.R
Fixed issues with warning messages and checks for non-missing input values after the maximum year.
