+++ FrEDI/R/run_fredi_sv.R
@@ -112,8 +112,9 @@ run_fredi_sv <- function(
   ### Get FrEDI data objects
   fredi_config  <- "fredi_config"  |> get_frediDataObj("frediData")
   co_modelTypes <- "co_modelTypes" |> get_frediDataObj("frediData")
+  co_states     <- "co_states"     |> get_frediDataObj("frediData")
+  temp_default  <- "temp_default"  |> get_frediDataObj("frediData")
   pop_default   <- "pop_default"   |> get_frediDataObj("stateData")
-  co_states     <- "co_states"     |> get_frediDataObj("stateData")
 
   ### Assign config files
   fredi_config |> list2env(envir = environment())
@@ -338,45 +339,13 @@ run_fredi_sv <- function(
       msg2 |> message("Exiting...")
       return()
     } ### End if(!("data.frame" %in% class_popInput))
+
     ### Check for popInput columns
     ### Info about popInputs
     popInputCols    <-  popInput |> names()
     popCols_inInput <- (popCols %in% popInputCols)
-    if(popCols_inInput |> all()){
-      msg2 |> message("All population scenario columns present in `popInput`...")
-      ### Filter to non-missing data
-      popInput     <- popInput |> filter_all(all_vars(!(. |> is.na())))
-
-      ### Check input Population values: no repeating years
-      if(popInput |> nrow()){
-        df_dups       <- popInput |>
-          group_by_at(c("year", "region", "state")) |>
-          summarize(n=n(), .groups="drop")
-        checkIssues  <- any(df_dups$n > 1)
-        rm(df_dups)
-      } else{ ### if(popInput |> nrow())
-        checkIssues <- TRUE
-      } ### End else(popInput |> nrow())
-      ### If there are issues with years:
-      if(checkIssues){
-        msg2 |> message("Error: duplicate years present in `popInput`!")
-        msg2 |> message("Exiting...")
-        return()
-      } ### End if(checkIssues)
-      rm(checkIssues)
-
-      ### Check input Population values: population >= 0
-      checkIssues <- (popInput$reg_pop < 0)
-      anyIssues   <- checkIssues |> any()
-      if(anyIssues){
-        msg2 |> message("Error: Values for 'reg_pop' in `popInput` must be greater than zero!")
-        msg2 |> message("Exiting...")
-        return()
-      } ### End if(checkIssues)
-      has_popInput <- TRUE
-      rm(checkIssues, anyIssues)
-    } ### End if(all(popCols_inInput))
-    else{
+    allPopCols      <- popCols_inInput |> all()
+    if(!allPopCols) {
       ### Exit and message the user
       # msg1 |> message("Using default regional population scenario", "...", "\n")
       # has_popInput    <- FALSE
@@ -384,9 +353,47 @@ run_fredi_sv <- function(
       msg3 |> message("'", paste(popCols[!popCols_inInput], collapse="', '"),"'...", "\n")
       msg2 |> message("Exiting...")
       return()
-    } ### End else(all(popCols_inInput))
+    } ### End if(!allPopCols)
+
+    ### Otherwise message user
+    msg2 |> message("All population scenario columns present in `popInput`...")
+    ### Filter to non-missing data
+    popInput     <- popInput |> filter_all(all_vars(!(. |> is.na())))
+    popInput     <- popInput |> unique()
+
+    ### Check input Population values: no repeating rows
+    if(popInput |> nrow()){
+      df_dups       <- popInput |>
[…]
