### 2024-06-27 b22d5f1
Update temps2slr.R
Fixed issue with ref year
