### 2024-06-27 555d4e3
Update run_fredi_sv.R
Removed deprecated lines
