### 2024-06-27 53c0743
Update utils_sv.R
Added additional step to check impact functions to avoid an error when the impact function was NA
