### 2024-06-27 61b767e
Update run_fredi_sv.R
- Corrected issue with earlier file name edit
- Relocated columns, added sector name to outputs
