### 2024-06-27 805de30
Quick fix to calc_tractScaledImpacts
