+++ FrEDI/man/run_fredi_sv.Rd
@@ -17,13 +17,19 @@ run_fredi_sv(
 
 \item{driverInput}{A data frame of up to four custom scenarios for drivers (temperature or global mean sea level rise). \code{driverInput} requires a data frame with columns of \code{"year"} and \code{"scenario"}. The data frame must also include a third column: \code{"temp_C"} for temperature-driven sectors (containing temperature values in degrees Celsius of warming for the contiguous U.S.) or \code{"slr_cm"} for sea level rise (SLR)-driven sectors (containing values for global mean sea level rise in centimeters). Run \code{get_sv_sectorInfo(gcmOnly=TRUE)} to see temperature-driven sectors in the SV module and \code{get_sv_sectorInfo(slrOnly=TRUE)} to see SLR-driven scenarios. Users can also pass a data frame with all four columns (\code{"year"}, \code{"scenario"}, \code{"temp_C"}, and \code{"slr_cm"}), in which case \code{\link[=run_fredi_sv]{run_fredi_sv()}} determines whether to use the \code{"temp_C"} or \code{"slr_cm"} column as the driver trajectory based on the specified sector. Driver inputs for all scenarios should start in the year 2000 or earlier. All scenarios must include at least two non-missing values  (especially values before or at 2000 and at or after 2090). If any required columns are missing, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the default temperature or sea level rise scenario from \code{\link[=run_fredi]{run_fredi()}}. If the data frame passed to \code{driverInput} has more than four unique scenarios, \code{\link[=run_fredi_sv]{run_fredi_sv()}} will only run the first four scenarios.}
 
-\item{popInput}{A data frame containing regional population trajectories for each of the seven regions for the contiguous U.S. (\code{"Midwest"}, \code{"Northeast"}, \code{"Northern Plains"}, \code{"Northwest"}, \code{"Southeast"}, \code{"Southern Plains"}, \code{"Southwest"}) as defined by the \href{https://scenarios.globalchange.gov/regions_nca4}{National Climate Assessment (NCA)}. The data frame passed to \code{popInput} should have columns \code{"year"}, \code{"region"}, and \code{"reg_pop"}, which respectively contain values for year, NCA region name, and regional population. \code{popInput} only accepts a data frame with a single scenario; \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses the same population scenario for any and all driver scenarios in the data frame passed to \code{driverInput}. If \code{popInput=NULL} (default), \code{\link[=run_fredi_sv]{run_fredi_sv()}} will use the default regional population trajectories.}
+\item{popInput}{The input population scenario requires state-level population values. Population values must be greater than or equal to zero.
+\itemize{
+\item The population scenario must be a data frame object with five columns with names \code{"year"}, \code{"region"}, \code{"state"}, \code{"postal"}, and \code{"state_pop"} containing the year, the NCA region name, the state name, the postal code abbreviation (e.g., "ME" for "Maine") for the state, and the state population, respectively.
+\item \code{popInput} only accepts a data frame with a single scenario; \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses the same population scenario for any and all driver scenarios in the data frame passed to \code{driverInput}.
+\item Population inputs must have at least one non-missing value in 2010 or earlier and at least one non-missing value in or after the final analysis year (2100).
+\item If the user does not specify an input scenario for population (i.e., \code{popInput = NULL}, \code{\link[=run_fredi_sv]{run_fredi_sv()}} uses a default population scenario.
+}}
 
[…]
