+++ FrEDI/man/FrEDI-package.Rd
@@ -7,10 +7,10 @@
 \title{README
 FrEDI: The Framework for Evaluating Damages and Impacts}
 \description{
-\link{FrEDI} is an R package being developed by the U.S. Environmental Protection Agency (EPA). The functions and data provided by this package can be used to estimate climate change impacts for the contiguous United States (CONUS) using the Framework for Evaluating Damages and Impacts (FrEDI), developed as part of EPA's \href{https://epa.gov/cira/}{Climate Change Impacts and Risk Analysis} (CIRA) project. \link{FrEDI} contains R code that implement FrEDI and allow users to project impacts from climate change and sea level rise for a selected set of sectors.
+FrEDI is an R package being developed by the U.S. Environmental Protection Agency (EPA). The functions and data provided by this package can be used to estimate climate change impacts for the contiguous United States (CONUS) using the \href{https://epa.gov/cira/FrEDI/}{Framework for Evaluating Damages and Impacts (FrEDI)}, developed as part of EPA's \href{https://epa.gov/cira/}{Climate Change Impacts and Risk Analysis} (CIRA) project. FrEDI contains R code that implement FrEDI and allow users to project impacts from climate change and sea level rise for a selected set of sectors.
 }
 \details{
-For help getting started with \link{FrEDI}, visit \url{https://usepa.github.io/FrEDI/articles/manual.html}.
+For help getting started with FrEDI, visit \url{https://usepa.github.io/FrEDI/articles/manual.html}.
 
 For additional package documentation, see \url{https://usepa.github.io/FrEDI/index.html}.
 
@@ -18,9 +18,10 @@ For more information on the Framework and the CIRA project, visit \url{https://e
 }
 \section{Overview of Functions}{
 
-The function \code{\link[=run_fredi]{run_fredi()}} provided in this package is the primary function implementing the \href{https://epa.gov/cira/FrEDI/}{\code{FrEDI}} developed by the U.S. EPA for projecting annual climate impacts. The main inputs to \code{\link[=run_fredi]{run_fredi()}} are climate scenarios (temperature in degrees Celsius, global mean sea level rise in centimeters) and socioeconomic scenarios (U.S. gross domestic product, state population).
 
-\link{FrEDI} also contains functions to assist in the pre-processing of input scenarios and the post-processing of outputs.
+The function \code{\link[=run_fredi]{run_fredi()}} provided in this package is the primary function implementing the \href{https://epa.gov/cira/FrEDI/}{Framework for Evaluating Damages and Impacts (FrEDI)}, developed by the U.S. EPA for projecting annual climate impacts. The main inputs to \code{\link[=run_fredi]{run_fredi()}} are climate scenarios (temperature in degrees Celsius, global mean sea level rise in centimeters) and socioeconomic scenarios (U.S. gross domestic product, state population).
+
+FrEDI also contains functions to assist in the pre-processing of input scenarios and the post-processing of outputs.
 \itemize{
 \item Pre-processing functions include \code{\link[=get_sectorInfo]{get_sectorInfo()}}, \code{\link[=import_inputs]{import_inputs()}}, \code{\link[=convertTemps]{convertTemps()}}, \code{\link[=temps2slr]{temps2slr()}}.
 \itemize{
@@ -31,12 +32,15 @@ The function \code{\link[=run_fredi]{run_fredi()}} provided in this package is t
 }
 }
 
[…]
