+++ FrEDI/man/gcamScenarios.Rd
@@ -25,7 +25,7 @@ A dataframe containing six driver scenarios that can be passed as inputs to \cod
 \details{
 The scenarios in this dataframe were created using \href{https://jgcri.github.io/hector/}{Hector}, an open-source, reduced-form global carbon-cycle climate model (Hartin et al., 2015) to model temperatures associated with emissions scenarios from the Global Change Analysis Model v5.3 (GCAM). The Global Change Analysis Model v5.3 (\href{https://gcims.pnnl.gov/modeling/gcam-global-change-analysis-model}{GCAM}) is an open source model that represents the linkages between energy, water, land, climate and economic systems (Calvin et al., 2019).
 
-Note that these temperature scenarios have been converted from the original global temperatures to CONUS temperatures using the \link{convertTemps} (with argument \code{from="global"}). The scenarios can therefore be passed directly to \code{\link[=run_fredi]{run_fredi()}} or \code{\link[=run_fredi_sv]{run_fredi_sv()}}. *Please note that the \code{gcamScenarios} should be subset to any of the individual scenarios specified in the \code{scenario} column before passing to \code{\link[=run_fredi]{run_fredi()}} (e.g., \code{gcamScenarios |> dplyr::filter(scenario=="ECS_3.0_ref_0")} for the reference scenario).
+These temperature scenarios have the original global temperatures. Users must convert to CONUS temperatures using the \link{convertTemps} (with argument \code{from="global"}) before passing the scenarios to \code{\link[=run_fredi]{run_fredi()}} or \code{\link[=run_fredi_sv]{run_fredi_sv()}}. *Please note that the \code{gcamScenarios} should be subset to any of the individual scenarios specified in the \code{scenario} column before passing to \code{\link[=run_fredi]{run_fredi()}} (e.g., \code{gcamScenarios |> dplyr::filter(scenario=="ECS_3.0_ref_0")} for the reference scenario).
 
 Calvin, K., Patel, P., Clarke, L., et al. 2019. GCAM v5.1: representing the linkages between energy, water, land, climate, and economic systems, Geosci. Model Dev., 12:677–698. https://doi.org/10.5194/gmd-12-677-2019.
 
