+++ FrEDI/DESCRIPTION
@@ -44,7 +44,7 @@ Depends: R (>= 4.2.0),
     ggpubr,
     openxlsx
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.1
+RoxygenNote: 7.3.2
 LazyData: TRUE
 Suggests: knitr, rmarkdown, prettydoc, devtools, kableExtra
 VignetteBuilder: knitr

[…]
