### 2024-07-02 6edd2b1
Updated documentation
Updated package documentation to correct some things related to the new population data
