+++ FrEDI/R/FrEDI-package.R
@@ -41,12 +41,12 @@
 #' * __help__ and __html__. Contain documentation for functions available to the user, including function descriptions, lists of arguments and outputs, and examples. See `"html/00Index.html"` or the individual R help functions for more information about individual functions.
 #' * __Meta__. Contains RDS files (ending in `".rds"`) with information about the package contents.
 #' * __extdata__. __extdata/scenarios__ Contains four CSV files for users to test the function for importing data. For more information on importing scenarios for use with [FrEDI::run_fredi()], refer to documentation for the function [FrEDI::import_inputs()].
-#'      * `"GCAM_scenario.csv"` contains a set of temperature scenarios that can be used with FrEDI, including the default temperature scenario used by both [FrEDI::run_fredi()] and [FrEDI::run_fredi_sv()] (labeled `"description="Hector_GCAM_v5.3_ECS_3.0_REF"` in `"slr_from_GCAM.csv"` and `"ECS_3.0_ref_0"` in the [FrEDI::gcamScenarios] dataset)). Also see the documentation for the [FrEDI::gcamScenarios] dataset for more information.
-#'      * `"State ICLUS Population.csv"` and `"Region ICLUS Population.csv"` contain the default population scenarios used by [FrEDI::run_fredi()] and [FrEDI::run_fredi_sv()], respectively (see [FrEDI::popScenario] and [FrEDI::popScenario_sv]).
+#'      * `"GCAM_scenario.csv"` contains a set of temperature scenarios that can be used with FrEDI, including the default temperature scenario used by both [FrEDI::run_fredi()] and [FrEDI::run_fredi_sv()]. Also see documentation for the [FrEDI::gcamScenarios] dataset for more information.
+#'      * `"State ICLUS Population.csv"` contains the default state population scenario used by [FrEDI::run_fredi()] (see [FrEDI::popScenario] and [FrEDI::popScenario_sv]).
 #'      * `"slr_from_GCAM.csv"` contains global mean sea level rise heights in centimeters (created from the reference temperature scenario).
 #'  * __extdata/sv__ Contains files used by the `FrEDI` SV module to calculate impacts.
 #'
-#' The `FrEDI` package contains a loadable dataset with default results `defaultResults`, which contains annual impacts produced by [FrEDI::run_fredi()] for the with the default options and default scenarios (i.e., default temperature, GDP, and state population trajectories). Other loadable datasets provided by FrEDI are a set of driver scenarios ([FrEDI::gcamScenarios]), a state population scenario ([FrEDI::popScenario]) for use with [FrEDI::run_fredi()] or [FrEDI::import_inputs()], and a regional population scenario (`popScenario_sv`) for use with [FrEDI::run_fredi_sv()]. These data can be loaded into the workspace using the `data()` function.
+#' The `FrEDI` package contains a loadable dataset with default results `defaultResults`, which contains annual impacts produced by [FrEDI::run_fredi()] for the with the default options and default scenarios (i.e., default temperature, GDP, and state population trajectories). Other loadable datasets provided by FrEDI are a set of driver scenarios ([FrEDI::gcamScenarios]) and a state population scenario ([FrEDI::popScenario]) for use with [FrEDI::run_fredi()] or [FrEDI::run_fredi_sv()], which can be loaded into the workspace using the `data()` function (e.g., `data(gcamScenarios)`).
 #'
 #' Typical use will involve `library(FrEDI)` or `require(FrEDI)`.
 #'
