### 2024-07-02 048a2de
Update run_fredi_sv.R
Added another check to delete impactsList if it exists
