### 2024-07-02 8ec98f8
Changed output of FrEDI SV from region to state level
- Changed output of FrEDI SV from region to state level
- Cleaned up SV functions
