### 2024-07-03 c97f746
Changed SV outputs back to region level
Converted SV results back to the region level, since outputs may not be as accurate as small spatial resolutions.
