### 2024-07-03 20a24a2
Updated input workflow
Streamlined input workflow. Tested defaults, but not inputs
