+++ FrEDI/R/utils.R
@@ -1,41 +1,80 @@
 ###### Helpers ######
 ### Get column as vector
-get_vector <- function(
-    data, column = NULL
-){
-  ### Select column and get values as vector
-  col0  <- column |> is.null() |> ifelse(c(), column)
-  vals0 <- data[[column]] |> as.vector()
-  ### Return
-  return(vals0)
-}
+# get_vector <- function(
+#     data, column = NULL
+# ){
+#   ### Select column and get values as vector
+#   col0  <- column |> is.null() |> ifelse(c(), column)
+#   vals0 <- data[[column]] |> as.vector()
+#   ### Return
+#   return(vals0)
+# }
 
-### Get unique values from a dataframe column
-get_uniqueValues <- function(
-    data, column = NULL, sort=TRUE
-){
-  ### Select column and get values as vector
-  vals0 <- data  |> get_vector(column)
-  vals0 <- vals0 |> unique()
-  # vals0 |> print()
-  ### Sort
-  if(sort){vals0 <- vals0 |> sort()}
-  ### Return
-  return(vals0)
-}
+# ### Get unique values from a dataframe column
+# get_uniqueValues <- function(
+#     data, column = NULL, sort=TRUE
+# ){
+#   ### Select column and get values as vector
+#   vals0 <- data  |> get_vector(column)
+#   vals0 <- vals0 |> unique()
+#   # vals0 |> print()
+#   ### Sort
+#   if(sort){vals0 <- vals0 |> sort()}
+#   ### Return
+#   return(vals0)
+# }
 
 ### Get message prefix
-get_msgPrefix <- function(level=1){
-  ### Other vals
-  mnl0   <- "\n" ### Message new line
-  msg0   <- "\t" ### Message indent level 0
-  mcom   <- ", " ### Comma for collapsing lists
-  mqu0   <- "\'" ### Message quote
-  mqu1   <- mqu0 |> paste0(mcom, mqu0, collapse="")
-  mend0  <- "..."
-  msg_x  <- msg0
-  return(msg_x)
+get_msgPrefix <- function(level=1, newline=FALSE){
+  ### Message new line
+  msgN0  <- newline |> ifelse("\n", "")
+  ### Message indents
+  msgX0  <- "\t" |> rep(level)
+  ### Message prefix
+  msg0   <- msgN0 |> paste0(msgX0)
+  ### Return
+  return(msg0)
 }
+
+### Determine whether something is default or custom
+get_returnListStatus <- function(cond0=TRUE){
+  cond0 |> ifelse("Default", "Custom")
+}
+
+### Function to make it easier to subset a vector
+get_matches <- function(
+    x, ### Vector to subset
+    y, ### Vector to check x against
+    matches = TRUE,    ### Whether to subset to matches (TRUE) or non-matches (FALSE)
+    type    = "values" ### c("values", "matches", "which")
+){
+  ### Check which type to return
+  type      <- type |> tolower()
+  doValues  <- "values"  %in% type
+  doMatches <- "matches" %in% type
+  doWhich   <- "which"   %in% type
+
+  ### Get matches
+  set0      <- x %in% y
+
+  ### Get anti-match if matches = FALSE
+  if(!matches) set0 <- !set0
+
+  ### Get which values to return
+  which0    <- set0 |> which()
+
+  ### Get subset
+  x0        <- x[set0]
+
+  ### If doValues, return values
+  ### If doMatches, return matches
+  ### If doWhich, return which
+  if     (doValues) return(x0)
+  else if(doWhich ) return(which0)
+  else              return(set0)
+  ### End function
+}
+
 
[…]

[…]
