+++ FrEDI/R/run_fredi_modular.R
@@ -284,15 +284,18 @@ run_fredi_modular <- function(
 
 
   ###### ** Sectors List ######
-  ### Sector names & labels
-  sectorIds    <- co_sectors[["sector_id"   ]]
-  sectorLbls   <- co_sectors[["sector_label"]]
   ### Initialize sector list if the sectors list is null
-  if(sectorList |> is.null()) sectorList <- sectorIds
+  if(sectorList |> is.null()) sectorList <- co_sectors |> pull(sector_label)
+  ### Sector names & labels
+  dfSectors    <- co_sectors |> filter((sector_label |> tolower()) %in% (sectorList |> tolower()))
+  sectorIds    <- dfSectors  |> pull(sector_id)
+  sectorLbls   <- dfSectors  |> pull(sector_label)
+
   ### Check sectors
+  # sectorList |> print(); sectorLbls |> print()
   naSectors0   <- sectorList |> get_matches(y=sectorLbls, matches=F, type="values")
   ### Message the user
-  if(na_sectors |> length()){
+  if(naSectors0 |> length()){
     naSectors0  <- "\"" |> paste0(paste(naSectors0, collapse= "\", \""), "\"")
     msgSectors0 <- "\"" |> paste0(paste(sectorLbls, collapse= "\", \""), "\"")
     1 |> get_msgPrefix(newline=T) |> paste0("Warning! Error in `sectorList`:") |> message()
@@ -314,14 +317,14 @@ run_fredi_modular <- function(
   df_inputInfo <- co_inputScenarioInfo
   df_inputInfo <- df_inputInfo |> mutate(ref_year = c(1995, 2000, 2010, 2010))
   df_inputInfo <- df_inputInfo |> mutate(min_year = c(2000, 2000, 2010, 2010))
-  df_inputInfo <- df_inputInfo |> mutate(max_year = maxYear |> rep(inNames |> length()))
+  df_inputInfo <- df_inputInfo |> mutate(max_year = maxYear |> rep(df_inputInfo |> nrow()))
 
   ### Input defaults
   inputDefs    <- list()
-  inputsDefs[["temp"]] <- co_defaultTemps |> rename_at(c("temp_C_conus"), ~"temp_C") |> select(c("year", "temp_C"))
-  inputsDefs[["slr" ]] <- slr_default
-  inputsDefs[["gdp" ]] <- gdp_default
-  inputsDefs[["pop" ]] <- pop_default
+  inputDefs[["temp"]] <- co_defaultTemps |> rename_at(c("temp_C_conus"), ~"temp_C") |> select(c("year", "temp_C"))
+  inputDefs[["slr" ]] <- slr_default
+  inputDefs[["gdp" ]] <- gdp_default
+  inputDefs[["pop" ]] <- pop_default
 
   ### Input info
   # inNames      <- c("temp", "slr", "gdp", "pop")
@@ -347,14 +350,13 @@ run_fredi_modular <- function(
   })()
 
   ### Make sure all inputs are present
-  inputsList0  <- inputsList
-  for(name_i in inNames) {inputsList[[name_i]] <- inputsList0[[name_i]]; rm_name_i}
-  rm(inputsList0)
+  inputsList   <- inNames |> map(function(name_i){inputsList[[name_i]]}) |> set_names(inNames)
 
   ### Check whether inputs are present
   hasInputs    <- inNames     |> map(function(name_i){!(inputsList[[name_i]] |> is.null())}) |> set_names(inNames)
   whichInputs  <- hasInputs   |> unlist()
   hasAnyInputs <- whichInputs |> length()
+  # whichInputs |> print()
 
   ### Create logicals and initialize inputs list
   if(hasAnyInputs) {
[…]
