### 2024-07-03 08924d4
More fixes to inputs workflow
