### 2024-07-08 0dbc33d
Edits to package documentation
