+++ FrEDI/man/import_inputs.Rd
@@ -18,18 +18,18 @@ import_inputs(
 
 \item{slrfile=NULL}{A character string indicating the location of a CSV file containing a custom sea level rise scenario. The first column in the CSV should be named \code{"year"} and contain years associated with the sea level rise estimates; second column should be named "\verb{slr_cm"} and contain values for global mean sea level rise (GMSL), in centimeters, above a baseline year of 2000. The SLR scenario must start in 2000 or earlier and end at or after the maximum model run year (e.g., as specified by the \code{maxYear} argument to \code{\link[=run_fredi]{run_fredi()}}).}
 
-\item{gdpfile=NULL}{A character string indicating the location of a CSV file containing a custom scenario for U.S. gross domestic product (GDP). The first column in the CSV should be named \code{"year"} and contains years associated with the GDP estimates; the second column (named \code{"gdp_usd"}) should contain values for U.S. GDP in 2015 U.S. dollars  (2015$ USD). The GDP scenario must start in 2010 or earlier and end at or after the maximum model run year (e.g., as specified by the \code{maxYear} argument to \code{\link[=run_fredi]{run_fredi()}}).}
+\item{gdpfile=NULL}{A character string indicating the location of a CSV file containing a custom scenario for U.S. gross domestic product (GDP). The first column in the CSV should be named \code{"year"} and contains years associated with the GDP estimates; the second column (named \code{"gdp_usd"}) should contain values for U.S. GDP in 2015 U.S. dollars  (2015$ USD). The GDP scenario must start in 2010 or earlier and end at or after the maximum model run year (e.g., as specified by the \code{maxYear} argument to \code{\link[=run_fredi]{run_fredi()}}). Values for GDP in column \code{"gdp_usd"} must be greater than or equal to zero.}
 
-\item{popfile=NULL}{A character string indicating the location of a CSV file containing a custom population scenario for states and NCA regions. The first column should be named \code{"year"} and contain years in the interval 2010 to 2300. The second column should be called \code{"pop"} and contain population values. Whether a third column is required depends on the geographical scale of the data, as specified by the \code{popArea} argument.
+\item{popfile=NULL}{A character string indicating the location of a CSV file containing a custom population scenario for states and NCA regions. The first column should be named \code{"year"} and contain years in the interval 2010 to 2300. The second column should be called \code{"pop"} and contain population values. Whether a third column is required depends on the geographical scale of the data, as specified by the \code{popArea} argument (for more information, see the \code{popArea} argument, below.
 \itemize{
-\item If \code{popArea = "state"}, the third column should be called \code{"state"} and contain the name of the state associated with each population estimate in a given year. If \code{popArea = "regional"}, the third column should be called \code{"region"} and contain the NCA Region label associated with each population estimate in a given year. If \code{popArea = "national"} or \code{popArea = "area"}, only the \code{year} and \code{pop} columns are required.
-\item If \code{popArea = "state"}, the file must contain estimates for all CONUS states (all states except Alaska and Hawaii) must be present. If \code{popArea = "regional"}, the file must contain estimates for all seven NCA regions.
-\item The population scenario must start in 2010 or earlier and end at or after the maximum model run year (e.g., as specified by the \code{maxYear} argument to \code{\link[=run_fredi]{run_fredi()}}).
[…]
