+++ FrEDI/man/gcamScenarios.Rd
@@ -3,12 +3,12 @@
 \docType{data}
 \name{gcamScenarios}
 \alias{gcamScenarios}
-\title{Six driver scenarios that can be passed as inputs to \code{\link[=run_fredi_sv]{run_fredi_sv()}}}
+\title{Six driver scenarios that can be passed as inputs to \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}}}
 \format{
 A data frame with 546 rows and 4 columns:
 \describe{
 \item{year}{Year}
-\item{temp_C}{CONUS temperature (in degrees Celsius) for the associated year and scenario}
+\item{temp_C}{Global temperature (in degrees Celsius) for the associated year and scenario}
 \item{slr_cm}{Global Mean Sea Level Rise (in centimeters) for the associated year and scenario}
 \item{scenario}{Associated scenario identifier}
 }
@@ -20,12 +20,21 @@ A data frame with 546 rows and 4 columns:
 gcamScenarios
 }
 \description{
-A dataframe containing six driver scenarios that can be passed as inputs to \code{\link[=run_fredi_sv]{run_fredi_sv()}}.
+A dataframe containing six driver scenarios that can be passed as inputs to \code{\link[=run_fredi]{run_fredi()}} and \code{\link[=run_fredi_sv]{run_fredi_sv()}}. This data frame has four columns -- \code{year}, \code{temp_C}, \code{slr_cm}, and \code{scenario} -- respectively containing:
 }
 \details{
-The scenarios in this dataframe were created using \href{https://jgcri.github.io/hector/}{Hector}, an open-source, reduced-form global carbon-cycle climate model (Hartin et al., 2015) to model temperatures associated with emissions scenarios from the Global Change Analysis Model v5.3 (GCAM). The Global Change Analysis Model v5.3 (\href{https://gcims.pnnl.gov/modeling/gcam-global-change-analysis-model}{GCAM}) is an open source model that represents the linkages between energy, water, land, climate and economic systems (Calvin et al., 2019).
+\describe{
+\item{year}{The Common Era (CE) year associated with the observation.}
+\item{temp_C}{Global temperature (i.e., degrees of warming above the baseline year of 1995) in degrees Celsius, for the associated year and scenario.}
+\item{slr_cm}{Global Mean Sea Level Rise (i.e., SLR) in centimeters for the associated year and scenario.}
+\item{scenario}{Associated scenario identifier.}
+}
 
-These temperature scenarios have the original global temperatures. Users must convert to CONUS temperatures using the \link{convertTemps} (with argument \code{from="global"}) before passing the scenarios to \code{\link[=run_fredi]{run_fredi()}} or \code{\link[=run_fredi_sv]{run_fredi_sv()}}. *Please note that the \code{gcamScenarios} should be subset to any of the individual scenarios specified in the \code{scenario} column before passing to \code{\link[=run_fredi]{run_fredi()}} (e.g., \code{gcamScenarios |> dplyr::filter(scenario=="ECS_3.0_ref_0")} for the reference scenario).
[…]

[…]
