### 2024-07-08 30ad5ef
Update FrEDI/R/run_fredi_sv.R
Accepting suggesting

Co-authored-by: Erin McDuffie <51747880+emcduffie@users.noreply.github.com>
