+++ FrEDI/vignettes/articles/References.Rmd
@@ -41,9 +41,9 @@ Sarofim, M.C., Martinich, J., Neumann, J.E. et al. A temperature binning approac
 **Technical Documentation**
 <br>
 
-FrEDI Technical Documentation (last updated July 2023): https://www.epa.gov/system/files/documents/2023-08/Technical%20Documentation%20on%20the%20Framework%20for%20Evaluating%20Damages%20and%20Impacts_MainText_2023.07.pdf 
+FrEDI Technical Documentation: https://www.epa.gov/system/files/documents/2023-08/Technical%20Documentation%20on%20the%20Framework%20for%20Evaluating%20Damages%20and%20Impacts_MainText_2023.07.pdf 
 
-Technical Documentation Appendix (last updated July 2023): https://www.epa.gov/system/files/documents/2023-08/Technical%20Documentation%20on%20the%20Framework%20for%20Evaluating%20Damages%20and%20Impacts_Appendices_2023.07.pdf 
+Technical Documentation Appendix: https://www.epa.gov/system/files/documents/2023-08/Technical%20Documentation%20on%20the%20Framework%20for%20Evaluating%20Damages%20and%20Impacts_Appendices_2023.07.pdf 
 
 <br>
 
