+++ FrEDI/R/import_inputs.R
@@ -214,10 +214,18 @@ import_inputs <- function(
   hasPop     <- df_pop |> length()
   if(hasPop) df_pop <- df_pop |> calc_import_pop(popArea=popArea)
   if(!(df_pop |> is.null())) {
+    ### Rename state_pop columne
+    doRename <- "state_pop" %in% (data1 |> names())
+    if(doRename) {
+      rename0  <- c("state_pop")
+      renameTo <- c("pop")
+      df_pop   <- df_pop |> rename_at(c(rename0), ~renameTo)
+    } ### End if(doRename)
+    ### Update in output
     inputsList[["pop"]] <- df_pop
   } else {
     inputsList["pop"] <- list(NULL)
-  }
+  } ### End if(!(df_pop |> is.null()))
 
   ###### Rename List ######
   inputsList <- inputsList |> set_names(outNames)
