+++ FrEDI/R/run_fredi_modular.R
@@ -191,7 +191,7 @@ run_fredi_modular <- function(
   ### Level of messaging (default is to message the user)
   msgUser   <- !silent
   ### Uncomment for allCols
-  doPrimary <- F  ### whether to filter to primary impacts
+  # doPrimary <- F  ### whether to filter to primary impacts
   ### Model years and NPD (FrEDI past 2100)
   refYear   <- 2100
   npdYear   <- 2300
@@ -242,7 +242,10 @@ run_fredi_modular <- function(
   ###### ** Load Objects ######
   ### Assign data objects to objects in this namespace
   ### Assign FrEDI config
-  for(list_i in rDataList) list_i[["data"]] |> list2env(envir = environment())
+  # for(list_i in rDataList) list_i[["data"]] |> list2env(envir = environment())
+  fredi_config <- "fredi_config" |> get_frediDataObj("frediData")
+  # fredi_config |> print()
+  # for(name_i in fredi_config) {name_i |> assign(fredi_config[[name_i]]); rm(name_i)}
   fredi_config |> list2env(envir = environment())
 
   ###### ** Years   ######
@@ -251,7 +254,7 @@ run_fredi_modular <- function(
 
   ###### ** Elasticity ######
   ### Message user about elasticity
-  has_elasticity <- elasticity |> is.numeric()
+  has_elasticity <- elasticity     |> is.numeric()
   elasticity     <- has_elasticity |> ifelse(elasticity, elasticity0)
   if(!has_elasticity){
     paste0("\t", "Incorrect value type provided for argument 'elasticity'...") |> message()
@@ -285,6 +288,7 @@ run_fredi_modular <- function(
 
   ###### ** Sectors List ######
   ### Initialize sector list if the sectors list is null
+  co_sectors   <- "co_sectors" |> get_frediDataObj("frediData")
   if(sectorList |> is.null()) sectorList <- co_sectors |> pull(sector_label)
   ### Sector names & labels
   dfSectors    <- co_sectors |> filter((sector_label |> tolower()) %in% (sectorList |> tolower()))
@@ -296,8 +300,8 @@ run_fredi_modular <- function(
   naSectors0   <- sectorList |> get_matches(y=sectorLbls, matches=F, type="values")
   ### Message the user
   if(naSectors0 |> length()){
-    naSectors0  <- "\"" |> paste0(paste(naSectors0, collapse= "\", \""), "\"")
-    msgSectors0 <- "\"" |> paste0(paste(sectorLbls, collapse= "\", \""), "\"")
+    naSectors0  <- "\"" |> paste0(naSectors0 |> paste(collapse= "\", \""), "\"")
+    msgSectors0 <- "\"" |> paste0(sectorLbls |> paste(collapse= "\", \""), "\"")
     1 |> get_msgPrefix(newline=T) |> paste0("Warning! Error in `sectorList`:") |> message()
     2 |> get_msgPrefix() |> paste0("Impacts are not available for sectors: ", naSectors0) |> message()
     2 |> get_msgPrefix(newline=T) |> paste0("Available sectors: ", msgSectors0) |> message()
@@ -306,28 +310,27 @@ run_fredi_modular <- function(
   ### Update in list
   if(outputList) {argsList[["sectorList"]] <- sectorLbls}
 
-  ### Filter to sectors
-  co_slrScalars  <- co_slrScalars  |> filter(sector %in% sectorIds)
-  df_sectorsInfo <- df_sectorsInfo |> filter(sector %in% sectorIds)
 
 
-
-  ###### ** Inputs ######
+  ###### Scenarios ######
+  ###### ** Check Inputs ######
[…]
