+++ FrEDI/R/utils.R
@@ -1056,21 +1056,19 @@ get_slrScaledImpacts <- function(
   df_ext0     <- "slrExtremes"   |> get_frediDataObj("stateData")
   df_imp0     <- "slrImpacts"    |> get_frediDataObj("stateData")
   co_modTypes <- "co_modelTypes" |> get_frediDataObj("frediData")
-  # co_models   <- "co_models"     |> get_frediDataObj("frediData")
 
   ###### Values #######
   slr0        <- "slr"
-  # co_modTypes |> glimpse()
   co_modTypes <- co_modTypes |> rename(modelType = modelType_id)
   slrMax0     <- (co_modTypes |> filter(modelType==slr0))[["modelMaxOutput"]][1]
 
   ###### Format Data #######
   ### Filter to appropriate driver values
   df1         <- df1  |> filter(modelType |> tolower() == slr0) #|> rename(model_type=modelType)
-  # df1 |> glimpse()
 
   ### Filter to appropriate years
   maxYear0    <- df1[["year"]] |> max()
+  maxYear0 |> print()
   df_ext0     <- df_ext0 |> filter(year <= maxYear0)
   df_imp0     <- df_imp0 |> filter(year <= maxYear0)
 
@@ -1086,12 +1084,8 @@ get_slrScaledImpacts <- function(
   drop0       <- join0 |> (function(x){x[!(x %in% c("year"))]})()
   df0         <- df0 |> select(all_of(select0))
   ### Join
-  # df0 |> glimpse(); df_ext0 |> glimpse()
-  # df0 |> group_by_at(c(join0)) |> summarize(n=n(),groups="keep") |> filter(n>1) |> glimpse()
-  # df_ext0 |> group_by_at(c(join0)) |> summarize(n=n(),groups="keep") |> filter(n>1) |> glimpse()
   df_ext0     <- df_ext0 |> left_join(df0, by=c(join0))
   df_imp0     <- df_imp0 |> left_join(df0, by=c(join0))
-  # "got here1" |> print(); df_ext0 |> glimpse(); df_imp0 |> glimpse()
   ### Drop
   df_ext0     <- df_ext0 |> select(-all_of(drop0))
   df_imp0     <- df_imp0 |> select(-all_of(drop0))
@@ -1112,8 +1106,6 @@ get_slrScaledImpacts <- function(
   df_slr0     <- df_slr0 |> relocate(all_of(select0))
   rm(df_ext0, df_imp0);
   rm(join0, select0, drop0)
-  # "got here1" |> print(); df_max0 |> glimpse(); df_slr0 |> glimpse()
-  # rm(select0, select1)
 
   ###### Scaled Impacts >= Max #######
   ### Filter to appropriate years
@@ -1121,11 +1113,10 @@ get_slrScaledImpacts <- function(
   maxYrs0     <- df_max0 |> pull(year) |> unique()
   nrow_max    <- maxYrs0 |> length()
 
-  ## Calculate scaled impacts for values > slrMax0
+  ### Calculate scaled impacts for values > slrMax0
   if(nrow_max){
-    df_max0  <- df_max0   |> mutate(deltaDriver    = modelUnitValue    - driverValue_ref)
-    df_max0  <- df_max0   |> mutate(scaled_impacts = impacts_intercept + impacts_slope * deltaDriver)
-    # df_max0 |> filter(deltaDriver < 0) |> nrow() |> print()
+    df_max0  <- df_max0 |> mutate(deltaDriver    = modelUnitValue    - driverValue_ref)
+    df_max0  <- df_max0 |> mutate(scaled_impacts = impacts_intercept + impacts_slope * deltaDriver)
   } else{
     df_max0  <- df_max0 |> mutate(scaled_impacts = NA)
   } ### End if(nrow_max)
[…]
