### 2024-07-09 ccb5984
Added flexibility to column names
Added flexibility to column names: Now, if the driver column is not found, the import_inputs helper function will try to string match the input type against column names. If it finds a match:

- If a temperature input, it will look for matches to strings "conus" and "global". If conus matches, it will use that column and set tempType to "conus". Else if global matches, it will use that column and set tempType to "global". Otherwise, it will use the first matched column.
- If a population input, it will look for matches to strings "state", "region", "conus", an
[…]
