+++ FrEDI/R/utils_import_inputs.R
@@ -144,12 +144,9 @@ run_fun_tryInput <- function(
   } ### End if(!has_i)
 
   ### Message user, then check if file import was successful
-  msg2_i |> paste0(msg_i4) |> message()
-  if(!has_i) {
-    return(df_i)
-  } else{
-    df_i   <- try_i[["fileInput"]]
-  } ### End if(!has_i)
+  msg1_i |> paste0(msg_i4) |> message()
+  if(!has_i) return(df_i)
+  else       df_i <- try_i[["fileInput"]]
 
   ### Return df_i
   return(df_i)
@@ -158,7 +155,7 @@ run_fun_tryInput <- function(
 
 ###### rename_inputs ######
 # rename_inputs <- function(
-#     data,
+    #     data,
 #     new_names
 # ){
 #   ### Get the length of the new names
@@ -219,10 +216,10 @@ check_pop_regions <- function(
 ){
   ###### Load Data from FrEDI ######
   ### Get objects from FrEDI name space
-  ### Get input scenario info: co_inputScenarioInfo
+  ### Get input scenario info: co_info
   ### Get state info: co_states
   fListNames <- c("co_states")
-  sListNames <- c("df_popRatios")
+  sListNames <- c("df_ratios")
   for(name_i in fListNames){name_i |> assign(rDataList[["frediData"]][["data"]][[name_i]]); rm(name_i)}
   for(name_i in sListNames){name_i |> assign(rDataList[["stateData"]][["data"]][[name_i]]); rm(name_i)}
 
@@ -297,6 +294,158 @@ check_pop_regions <- function(
 
 
 
+###### check_valCols ######
+### Check value columns
+check_valCols <- function(
+    inputDf,   ### Data frame of values
+    inputName, ### Input name c("temp", "slr", "gdp", "pop")
+    valCol   , ### Value column
+    tempType = "conus",  ### One of: c("conus", "global")
+    popArea  = "state", ### One of: c("state", "regional", "conus", "national")
+    msgLevel = 2
+) {
+  ###### Messages ######
+  msgN      <- "\n"
+  msg0_i    <- get_msg_prefix(level=msgLevel)
+  msg1_i    <- get_msg_prefix(level=msgLevel + 1)
+  msg2_i    <- get_msg_prefix(level=msgLevel + 2)
+  msg3_i    <- get_msg_prefix(level=msgLevel + 3)
+
+  ###### Conditionals ######
+  doTemp    <- "temp" %in% inputName
+  doPop     <- "pop"  %in% inputName
+
+  ###### Data Columns ######
+  ### Get value column and id column
+  if(valCol |> is.null()) valCol <- get_import_inputs_valCols(popArea=popArea)[[inputName]]
+  # if(idCol  |> is.null()) idCol  <- get_import_inputs_idCols (popArea=popArea)[[inputName]]
+
+  ###### Check Values ######
+  namesDF   <- inputDf |> names()
+  checkVCol <- valCol %in% namesDF
+  vColMatch <- namesDF[checkVCol]
+  nMatches  <- vColMatch |> length()
+  ### Check if there are any string matches, and get number of matches
+  if(!nMatches) {
+    ### Messages
+    msg_vcol0 <- paste0("Column ", "\"", valCol, "\"", " not found in ", inputName, "file data!")
+    msg_vcol1 <- paste0("Looking for columns with matches to the string ", "\"", inputName, "\"", "...")
+    msg1_i |> paste0(msg_vcol0) |> message()
+    # msg2_i |> paste0(msg_vcol1) |> message()
+    ### Look for matches
+    checkVCol <- namesDF |> tolower() |> str_detect(inputName |> tolower())
+    vColMatch <- namesDF[checkVCol]
+    nMatches  <- vColMatch |> length()
+    matches2  <- nMatches > 1
[…]
