### 2024-07-09 fe83049
Improved messaging
Made small improvements to messaging
