+++ FrEDI/R/utils.R
@@ -337,7 +337,7 @@ format_inputScenarios <- function(
       # "got here" |> print()
       ### - Get new info
       valCol0 <- infoSlr |> pull(valueCol)
-      yrRef0  <- infoSlr |> pull(ref_year)
+      # yrRef0  <- infoSlr |> pull(ref_year)
       ### - First, calculate global temps
       df0     <- df0 |> mutate(temp_C = temp_C |> convertTemps(from="conus"))
       ### - Then, zero out again
@@ -1045,7 +1045,7 @@ get_gcmScaledImpacts <- function(
 get_slrScaledImpacts <- function(
     df0, ### Initial results for SLR sectors
     df1  ### Driver data frames
-    ){
+){
   ####### By State
   ### By state
   byState     <- "state" %in% (df0 |> names())
@@ -1059,31 +1059,35 @@ get_slrScaledImpacts <- function(
 
   ###### Values #######
   slr0        <- "slr"
-  co_modTypes <- co_modTypes |> rename(modelType = modelType_id)
-  slrMax0     <- (co_modTypes |> filter(modelType==slr0))[["modelMaxOutput"]][1]
+  # co_modTypes <- co_modTypes |> rename(modelType = modelType_id)
+  # slrMax0     <- (co_modTypes |> filter(modelType==slr0))[["modelMaxOutput"]][1]
+  # slrMax0     <- df1 |> pull(modelMaxOutput) |> unique()
 
   ###### Format Data #######
   ### Filter to appropriate driver values
-  df1         <- df1  |> filter(modelType |> tolower() == slr0) #|> rename(model_type=modelType)
+  df1         <- df1 |> filter(modelType |> tolower() == slr0) #|> rename(model_type=modelType)
+  slrMax0     <- df1 |> pull(modelMaxOutput) |> unique()
 
   ### Filter to appropriate years
-  maxYear0    <- df1[["year"]] |> max()
-  maxYear0 |> print()
+  maxYear0    <- df1 |> pull(year) |> max()
   df_ext0     <- df_ext0 |> filter(year <= maxYear0)
   df_imp0     <- df_imp0 |> filter(year <= maxYear0)
 
   ### Filter to appropriate sectors
-  sectors0    <- df0[["sector"]] |> unique()
+  sectors0    <- df0 |> pull(sector) |> unique()
   df_ext0     <- df_ext0 |> filter(sector %in% sectors0)
   df_imp0     <- df_imp0 |> filter(sector %in% sectors0)
 
   ###### Get scenario_id #######
   ### Join with df0 to get scenario_id
-  join0       <- c("sector", "variant", "impactType", "impactYear", "region") |> c(stateCols0) |> c("byState") |> c("year")
+  # join0       <- c("sector", "variant", "impactType", "impactYear", "region") |> c(stateCols0) |> c("byState") |> c("year")
+  join0       <- df0   |> names() |> get_matches(df_ext0 |> names(), matches=TRUE) |> get_matches(c("scenario_id"), matches=FALSE)
   select0     <- join0 |> c("scenario_id")
-  drop0       <- join0 |> (function(x){x[!(x %in% c("year"))]})()
-  df0         <- df0 |> select(all_of(select0))
+  # drop0       <- join0 |> (function(x){x[!(x %in% c("year"))]})()
+  drop0       <- join0 |> get_matches(c("year"), matches=FALSE)
+  df0         <- df0   |> select(all_of(select0))
   ### Join
+  # return(list(df_ext0=df_ext0, df0=df0))
   df_ext0     <- df_ext0 |> left_join(df0, by=c(join0))
[…]
