### 2024-07-09 c9cc6f5
Delete run_fredi.R
Deleted run_fredi.R to update with a new workflow (after testing new workflow)
