### 2024-07-09 025276c
Renamed run_fredi_modular.R
Renamed `run_fredi_modular` to `run_fredi` after modularizing workflow for inputs and calculating impacts. May still need to modularize formatting of column names, etc.
