### 2024-07-10 0e1063c
Updated documentation & default results
