+++ FrEDI/man/run_fredi.Rd
@@ -56,7 +56,7 @@ This function allows users to project annual average climate change impacts thro
 
 Users can specify an optional list of custom scenarios with \code{inputsList} (for more information on the format of inputs, see \code{\link[=import_inputs]{import_inputs()}}). The function \code{\link[=import_inputs]{import_inputs()}} can be used to importing custom scenarios from CSV files. \code{\link[=import_inputs]{import_inputs()}} returns a list with elements \code{tempInput}, \code{slrInput}, \code{gdpInput}, and \code{popInput}, with each containing a data frame with a custom scenario for temperature, GMSL, GDP, and state-level population, respectively. If a user imports scenarios using \code{\link[=import_inputs]{import_inputs()}}, they can pass the outputs of \code{\link[=import_inputs]{import_inputs()}} directly to the \code{\link[=run_fredi]{run_fredi()}} argument \code{inputsList}. Note that the documentation for \code{\link[=import_inputs]{import_inputs()}} can also provide additional guidance and specification on the formats for each scenario type.
 
-If \code{inputsList = NULL}, \code{\link[=run_fredi]{run_fredi()}} uses defaults for temperature, SLR, GDP, and population. Otherwise, \code{\link[=run_fredi]{run_fredi()}} looks for a list object passed to the argument \code{inputsList}. Within that list, \code{\link[=run_fredi]{run_fredi()}} looks for list elements \code{tempInput}, \code{slrInput}, \code{gdpInput}, and \code{popInput} containing data frames with custom scenarios for temperature, GMSL, GDP, and regional population, respectively. \code{\link[=run_fredi]{run_fredi()}} will default back to the default scenarios for any list elements that empty or \code{NULL} (in other words, running \code{run_fredi(inputsList = list())} returns the same outputs as running \code{\link[=run_fredi]{run_fredi()}}).
+If \code{inputsList = NULL}, \code{\link[=run_fredi]{run_fredi()}} uses defaults for temperature, SLR, GDP, and population. Otherwise, \code{\link[=run_fredi]{run_fredi()}} looks for a list object passed to the argument \code{inputsList}. Within that list, \code{\link[=run_fredi]{run_fredi()}} looks for list elements \code{tempInput}, \code{slrInput}, \code{gdpInput}, and \code{popInput} containing data frames with custom scenarios for temperature, GMSL, GDP, and state population, respectively. \code{\link[=run_fredi]{run_fredi()}} will default back to the default scenarios for any list elements that empty or \code{NULL} (in other words, running \code{run_fredi(inputsList = list())} returns the same outputs as running \code{\link[=run_fredi]{run_fredi()}}).
 \itemize{
 \item \strong{Temperature Inputs.} The input temperature scenario requires CONUS temperatures in degrees Celsius relative to 1995 (degrees of warming relative to the baseline year--i.e., the central year of the 1986-2005 baseline). CONUS temperature values must be greater than or equal to zero degrees Celsius.
 \itemize{
@@ -79,7 +79,7 @@ If \code{inputsList = NULL}, \code{\link[=run_fredi]{run_fredi()}} uses defaults
 }
 \item \strong{Population Inputs.} The input population scenario requires state-level population values. Population values must be greater than or equal to zero.
 \itemize{
-\item \code{popInput} requires a data frame object with five columns with names \code{"year"}, \code{"region"}, \code{"state"}, \code{"postal"}, and \code{"state_pop"} containing the year, the NCA region name, the state name, the postal code abbreviation for the state, and the state population, respectively.
[…]
