+++ .github/workflows/build_fredi.yml
@@ -8,7 +8,32 @@
 ###      Note, this should be the same directory as the input paper.md
 name: 4. Build FrEDI Package
 
-on: [workflow_dispatch]
+# on: [workflow_dispatch]
+on:
+  workflow_dispatch:
+    inputs:
+      update_docs:
+        type: choice
+        description: Update documentation?
+        required: true
+        options:
+        - no
+        - yes
+      update_results:
+        type: choice
+        description: Update default results?
+        required: true
+        options:
+        - no
+        - yes
+      update_scenarios:
+        type: multi-choice
+        description: Update scenarios?
+        required: true
+        options:
+        - Temps & SLR
+        - Population
+        - GDP
 
 jobs:
   compile_data:
@@ -34,6 +59,69 @@ jobs:
             any::ggpubr
             any::openxlsx
             any::devtools
+
+      - name: Update Scenarios
+        run: |
+          Rscript -e '
+            ###### Conditionals ######
+            ### Which scenarios to update
+            do_temp  <- "Temps & SLR" %in% "${{ inputs.update_scenarios}}"
+            do_gdp   <- "GDP"         %in% "${{ inputs.update_scenarios}}"
+            do_pop   <- "Pop"         %in% "${{ inputs.update_scenarios}}"
+
+
+            ###### File Paths ######
+            ### - Main repo path, FrEDI project path
+            ### - Data path, scenario csv path
+            rPath0   <- "."
+            pPath0   <- rPath0 |> file.path("FrEDI")
+            dPath0   <- rPath0 |> file.path("data" )
+            cPath0   <- rPath0 |> file.path("inst", "extdata", "scenarios")
+
+
+            ###### Update Scenarios ######
+            ### Temperature & CSV
+            if(do_temp) {
+              gcamScenarios <- "scenarios_gcam" |> get_frediDataObj("frediData")
+              gcamScenarios <- gcamScenarios |> mutate(temp_C_conus = temp_C_global |> convertTemps(from="global"))
+              gcamScenarios <- gcamScenarios |> (function(df0){
+                scenarios0 <- gcamScenarios |> pull(scenario) |> unique()
+                dfSlr0     <- scenarios0 |> map(function(x0, df0=df0){
+                  df0 <- df0 |> filter(scenario == x0)
+                  df0 <- temps2slr(temps = df0 |> pull(temp_C_global), years = df0 |> pull(year))
+                  return(df0)
+                }) |> bind_rows(.id = "scenario")
+                df0        <- df0 |> left_join(dfSlr0, by=c("year"))
+                return(df0)
+              })()
+              arrange0      <- c("scenario", "year", "temp_C_conus", "temp_C_global", "slr_cm")
+              gcamScenarios <- gcamScenarios |> arrange_at(c(arrnage0))
+              fTemps        <- "Hector_GCAM_v5.3_scenarios"
+              dTemps        <- dPath0 |> paste0(fTemps, ".rda")
+              cTemps        <- cPath0 |> paste0(fTemps, ".csv")
+              gcamScenarios |> save(file=dTemps)
+              gcamScenarios |> write.csv(filename=cTemps, row.names=F)
+
+            } ### End if(do_temp)
+            pPath0 |> devtools::load_all()
+            defaultResults <- run_fredi()
+            save(defaultResults, file=oPath0)
+
+
+            ###### Update Documentation ######
+            ###### - Build Manual
+            ###### - Add and build vignettes
+            ###### - Generate Documentation
+            roxygen2::roxygenise(pPath0)
+            devtools::document(pkg = pPath0)
+            # devtools::build_manual(pkg = pPath0)
+            # devtools::build_vignettes(pkg = pPath0)
+
[…]
