### 2024-07-10 85a9bb4
Update run_fredi.R
- Dropped additional columns from outputs.
- Using `impactYear_label` column instead of `impactYear_id`
