### 2024-07-10 c7444d1
Update utils.R
Corrected issue with ref year (2000) for SLR heights (previously calculation was erroneously using the temperature ref year of 1995)
