### 2024-08-01 2849d26
Update utils.R
Modified `update_popScalars ` function to work with new, modular FrEDI Data procedures
