### 2024-08-05 743205a
Update temps2slr.R
Updated temps2slr with new rDataList calls
