+++ FrEDI/R/temps2slr.R
@@ -88,9 +88,11 @@ temps2slr <- function(
   ###### Initial Values ######
   ### Reference year 2000 and equilibrium temperature offset for 2000
   ### Assign reference year from config file (max_year)
-  fredi_config        <- "fredi_config" |> get_frediDataObj("frediData")
+  # fredi_config        <- "fredi_config" |> get_frediDataObj("frediData")
+  fredi_config        <- rDataList[["fredi_config"]]
   temps2slr_constants <- fredi_config[["temps2slr"]]
-  temps2slr_constants |> list2env(envir = environment())
+  # temps2slr_constants |> list2env(envir = environment())
+  for(name_i in fredi_config |> names()) {name_i |> assign(fredi_config[[name_i]]); rm(name_i)}
 
   #### Reference year is 2000
   co_modelTypes <- "co_modelTypes" |> get_frediDataObj("frediData") |> filter(modelType_id == "slr")
