### 2024-08-05 b77c86c
Update get_sectorInfo.R
Edited for new data list organization
