### 2024-08-06 9dfad63
Fixes
Fixes to aggregate_impacts an utils to get them working with new workflow
