+++ FrEDI/R/utils.R
@@ -78,37 +78,48 @@ get_matches <- function(
 
 ### This function makes it easier to get data objects from the sysdata.rda file
 get_frediDataObj <- function(
-    x        = NULL,       ### Object name
+    x        = NULL,        ### Object name
     listSub  = "frediData", ### Sublist name
     listall  = FALSE,
     listName = "rDataList",
     pkg      = "FrEDI",
-    lib.loc  = .libPaths()[1] ### Library path to look for packages
+    lib.loc  = .libPaths()[1], ### Library path to look for packages
+    msg0     = "\t"
 ){
   ### Messaging
-  msg0    <- "\t"
+  msg1    <- msg0 |> paste0("\t")
   ### Check if list name exists
   exists0 <- listName |> exists()
+  # exists0 |> print()
   ### If the listname exists in the name space, parse it
   ### Otherwise, grab it from a package name space
-  if(exists0){new_x <- parse(text=listName) |> eval()}
-  else       {
+  if(exists0) {
+    new_x <- parse(text=listName) |> eval()
+    # new_x |> names() |> print()
+  } else      {
     ### Check if package & list name
     pkgList0    <- lib.loc |> installed.packages()
     pkgExists0  <- pkg %in% pkgList0
-    if(!pkgExists0){
+    if(!pkgExists0) {
       msg0 |> paste0("Package doesn't exist...") |> message()
       msg0 |> paste0("Exiting...") |> message()
       return()
-    } ### End if(!pkgExists0)
-    else           {new_x <- getFromNamespace(listName, ns=pkg)}
+    } else          {
+      new_x <- getFromNamespace(listName, ns=pkg)
+    } ### End ### End if(!pkgExists0)
   } ### End else(exists0)
 
   ### Whether to list all items in data object or not
   # if(listall) {return_x <- new_x |> names()}
   # else        {return_x <- new_x[[x]]}
-  if(listall) {return_x <- new_x |> names()}
-  else        {return_x <- new_x[[listSub]][["data"]][[x]]}
+  if(listall) {
+    return_x <- new_x |> names()
+  } else      {
+    # return_x <- new_x[[listSub]][["data"]][[x]]
+    # new_x |> names() |> print()
+    # new_x[[listSub]] |> names() |> print()
+    return_x <- new_x[[listSub]][[x]]
+  } ### End if(listall)
   ### Return
   return(return_x)
 } ### End get_frediDataObj
@@ -123,7 +134,7 @@ interpolate_annual <- function(
     column    = NULL, ### Column to create results for
     rule      = NULL, ### for interpolation,
     method    = NULL, ### method for interpolation; default=linear
-    region    = "National.Total", ### Region if "addRegion"
+    region    = "NationalTotal", ### Region if "addRegion"
     byState   = FALSE ### If breakdown by state
 ){
   ###### Data Info ######
@@ -246,7 +257,7 @@ interpolate_annual <- function(
 ### Function to standardize the creation of the scenario_id
 get_scenario_id <- function(
     data_x,
-    include=c("region_dot", "model_dot") ### Character vector of column names to include
+    include=c("region", "model") ### Character vector of column names to include
 ){
   ### Other vals
   mnl0   <- "\n" ### Message new line
@@ -306,7 +317,8 @@ format_inputScenarios <- function(
   yrRef0  <- info1  |> pull(ref_year)
 
   ### Format data
-  select0 <- idCols0 |> c(valCol0) |> c("year")
[…]
