+++ FrEDI/R/aggregate_impacts.R
@@ -59,24 +59,22 @@ aggregate_impacts <- function(
     silent      = TRUE
 ){
   ###### Defaults ######
-  ### Not used currently; preserving it in messaging logicals for the future
+  # ### Not used currently; preserving it in messaging logicals for the future
   msgUser      <- !silent
   msg0         <- function(lvl0=1){c("\t") |> rep(lvl0) |> paste(collapse = c(""))}
   yearCol0     <- c("year")
-  # summaryCols  <- columns; rm(columns)
-  ####### By State  ######
-  byState      <- c("state") %in% (data |> names())
-  if(byState){stateCols0 <- c("state", "postal")} else{stateCols0 <- c()}
-  popCol0      <- byState |> ifelse("state_pop", "reg_pop")
-  # byState |> print()
+
+  ####### State Cols  ######
+  stateCols0   <- c("state", "postal")
+  popCol0      <- c("pop")
 
   ###### Format Data ######
   ### Ungroup
-  data         <- data |> ungroup() #; names(data) |> print()
+  data         <- data |> ungroup()
   # data |> glimpse()
+
   ###### Years Info ######
   ### Years in data
-  # c_npdRefYear <- 2090
   c_dataYears  <- data[[yearCol0]] |> unique()
 
   ###### Aggregation Levels  ######
@@ -125,7 +123,7 @@ aggregate_impacts <- function(
 
   ###### - Formate SLR Info ######
   co_slrs     <- co_models |> filter(modelType=="slr")
-  co_slrs     <- co_slrs   |> mutate_at(c("model_dot", "model_label"), as.character)
+  co_slrs     <- co_slrs   |> mutate_at(c("model_id", "model_label"), as.character)
 
   ###### - Format Sector Info ######
   rename0     <- c("sector_id", "sector_label", "modelType")
@@ -322,17 +320,12 @@ aggregate_impacts <- function(
   driverCols    <- driverCols |> (function(y){y[(y %in% names0)]})()
   rm(names0)
 
-  # ### Add state_pop column
-  # if(aggNational){
-  #   groupByCols <- groupByCols |> c(regPopCols) |> unique()
-  #   summaryCols <- summaryCols |> (function(y){y[!(y %in% groupByCols)]})()
-  # } ### End if(aggNational)
-
   ### List of standardized columns
   standardCols  <- c(groupByCols, baseCols, regPopCols, natPopCols) |> unique()
   standardCols  <- standardCols |> c(driverCols, summaryCols) |> unique()
   scenarioCols  <- standardCols |> (function(y){y[!(y %in% c(groupByCols, yearCol0, summaryCols))]})()
-  data          <- data |> select(any_of(standardCols))#; names(data) |> print
+  data          <- data |> select(any_of(standardCols))
+  # data |> names() |> print
 
   ###### Base Scenario Info  ######
   ### Some values are the same for all runs and regions...separate those values
@@ -345,14 +338,12 @@ aggregate_impacts <- function(
     group_by_at(c(regPopCols)) |>
     summarize(n=n(), .groups="keep") |> ungroup() |>
     select(-c("n"))
-  # baseCols |> print(); baseScenario |> glimpse()
-  # regPopCols |> print(); regionalPop |> glimpse()
   ### Create national population scenario from the base scenario
   nationalPop  <- baseScenario |>
     mutate(region = "National Total") |>
     select(all_of(natPopCols)) |>
     rename_at(c("national_pop"), ~popCol0)
-  if(byState){nationalPop <- nationalPop |> mutate(state="All", postal="US")}
+  nationalPop <- nationalPop |> mutate(state="All", postal="US")
   ### Driver Scenario
   driverScenario <- data |>
[…]
