### 2024-08-06 2a1d670
Fixed region
Fixed region values and standardized scaled impacts. Updated utilities to handle new data
