+++ FrEDI/R/utils.R
@@ -778,9 +778,9 @@ extend_slrScalars <- function(
   df_adj0    <- df_phys0 |> filter(year == refYear0)
   df_adj0    <- df_adj0  |> rename_at(c(rename0), ~rename1)
   df_adj0    <- df_adj0  |> select(all_of(select0))
-  df0        <- df0      |> mutate(physAdjName = physScalarName)
+  df0        <- df0 |> mutate(physAdjName = physScalarName)
   # df0 |> glimpse(); df_adj0 |> glimpse()
-  df0        <- df0      |> left_join(df_adj0, by=c(join0))
+  df0        <- df0 |> left_join(df_adj0, by=c(join0))
   ### Drop values
   rm(rename0, rename1, select0, join0, df_phys0, df_adj0)
 
@@ -791,8 +791,8 @@ extend_slrScalars <- function(
 
   ### Drop columns & join
   drop0      <- c("c2", "refYear")
-  df0        <- df0    |> filter(year > refYear)
-  df0        <- df0    |> select(-all_of(drop0))
+  df0        <- df0 |> filter(year > refYear)
+  df0        <- df0 |> select(-all_of(drop0))
   rm(drop0)
 
   ###### Return ######
@@ -806,117 +806,147 @@ extend_slrScalars <- function(
 ### Function "match_scalarValues" replaces "get_popWts", "get_physMultipliers", and "get_econScalars"
 match_scalarValues <- function(
     df0,       ### Initial results dataframe
-    sectors    = "co_sectors" |> get_frediDataObj("frediData") |> pull(sector_id), ### Vector of sectors
-    scalars    = "df_scalars" |> get_frediDataObj("stateData"),                    ### Scalars dataframe
+    # scalars    = "df_scalars" |> get_frediDataObj("stateData"),
     scalarType ### Type of scalar (one of: c("damageAdj", "econScalar", "physAdj", "phsScalar"))
 ){
-  ### State columns
-  # byState <- "state" %in% (df0 |> colnames())
-  # if(byState){stateCols0 <- c("state", "postal")} else{stateCols0 <- c()}
-  stateCols0       <- c("state", "postal")
+  ###### State columns ######
+  stateCols0  <- c("state", "postal")
+
+  ###### FrEDI Data ######
+  ### Get data from FrEDI
+  ### Rename column
+  scalars     <- "df_scalars" |> get_frediDataObj("stateData") ### Scalars dataframe
+
+  ### Scalar columns to rename
+  renameAt0   <- "scalarName"
+  renameTo0   <- scalarType |> paste0(c("Name"))
+  scalarName  <- renameTo0
+  scalars     <- scalars |> rename_at(c(renameAt0), ~renameTo0)
+
+  ### Filter to years
+  years0      <- df0     |> pull(year) |> get_years_fromData()
+  minYr0      <- years0  |> min()
+  maxYr0      <- years0  |> max()
+  scalars     <- scalars |> filter(year >= minYr0, year <= maxYr0)
 
   ###### Filter to Scalar Type ######
   # scalarType |> print()
-  scalarType0      <- scalarType; rm(scalarType)
-  scalars          <- scalars |> filter(scalarType==scalarType0)
-  scalars          <- scalars |> select(-c("scalarType"))
+  ### Values
+  scalarType0 <- scalarType
+  scalars0    <- df0     |> pull(all_of(scalarName)) |> unique()
+  scalars     <- scalars |> filter(scalarType %in% scalarType0)
+  scalars     <- scalars |> filter(scalarName %in% scalars)
+  scalars     <- scalars |> select(-c("scalarType"))
+  rm(scalarType)
   # "got here1" |> print(); scalars |> glimpse()
-
[…]
