+++ FrEDI/R/run_fredi.R
@@ -187,17 +187,29 @@ run_fredi <- function(
     allCols    = FALSE, ### Whether to include additional columns in output
     silent     = TRUE   ### Whether to message the user
 ){
+  ###### Load Objects ######
+  ### Assign data objects to objects in this namespace
+  ### Assign FrEDI config
+  # for(list_i in rDataList) list_i[["data"]] |> list2env(envir = environment())
+  # fredi_config <- "fredi_config" |> get_frediDataObj("fredi_config")
+  fredi_config <- rDataList[["fredi_config"]]
+  # fredi_config |> names() |> print()
+  for(name_i in fredi_config |> names()) {name_i |> assign(fredi_config[[name_i]]); rm(name_i)}
+  # fredi_config |> names() |> print()
+  # "got here" |> print()
+  # fredi_config |> list2env(envir = environment())
+
+
   ###### Set up the environment ######
   ### Level of messaging (default is to message the user)
   msgUser   <- !silent
   ### Uncomment for allCols
   # doPrimary <- F  ### whether to filter to primary impacts
   ### Model years and NPD (FrEDI past 2100)
-  refYear   <- 2100
-  npdYear   <- 2300
-  maxYear0  <- thru2300 |> ifelse(npdYear, maxYear)
-  do_npd    <- maxYear0 > refYear
-
+  minYear   <- minYear0
+  maxYear   <- thru2300 |> ifelse(npdYear0, maxYear)
+  do_npd    <- maxYear > maxYear0
+  # "got here2" |> print()
 
   ###### ** Return List ######
   ### Initialize list to return
@@ -218,7 +230,7 @@ run_fredi <- function(
     statusList[["sectorList"]] <- (!(sectorList |> is.null())) |> get_returnListStatus()
     statusList[["aggLevels" ]] <- "placeholder"
     statusList[["elasticity"]] <- (elasticity == 1) |> get_returnListStatus()
-    statusList[["maxYear"   ]] <- (maxYear == refYear & !thru2300) |> get_returnListStatus()
+    statusList[["maxYear"   ]] <- (maxYear == maxYear0 & !thru2300) |> get_returnListStatus()
     statusList[["thru2300"  ]] <- (!thru2300) |> get_returnListStatus()
     statusList[["allCols"   ]] <- (!allCols ) |> get_returnListStatus()
     statusList[["silent"    ]] <- ( silent  ) |> get_returnListStatus()
@@ -232,25 +244,17 @@ run_fredi <- function(
     argsList[["sectorList"]] <- "placeholder"
     argsList[["aggLevels" ]] <- "placeholder"
     argsList[["elasticity"]] <- elasticity
-    argsList[["maxYear"   ]] <- maxYear0
+    argsList[["maxYear"   ]] <- maxYear
     argsList[["thru2300"  ]] <- thru2300
     argsList[["allCols"   ]] <- allCols
     argsList[["silent"    ]] <- silent
   } ### End if(outputList)
 
 
-  ###### ** Load Objects ######
-  ### Assign data objects to objects in this namespace
-  ### Assign FrEDI config
-  # for(list_i in rDataList) list_i[["data"]] |> list2env(envir = environment())
-  fredi_config <- "fredi_config" |> get_frediDataObj("frediData")
-  # fredi_config |> print()
-  # for(name_i in fredi_config) {name_i |> assign(fredi_config[[name_i]]); rm(name_i)}
-  fredi_config |> list2env(envir = environment())
+
 
   ###### ** Years   ######
-  maxYear    <- maxYear0
-  list_years <- minYear:maxYear
+  # cYears0 <- minYear:maxYear
 
[…]
