+++ FrEDI/R/utils.R
@@ -800,6 +800,136 @@ extend_slrScalars <- function(
   return(df0)
 }
 
+
+get_co_sectorsInfo <- function(
+    sectors0   = NULL,  ### Sector IDs
+    addRegions = FALSE, ### Whether to include regions & states
+    addModels  = FALSE, ### Whether to include models
+    colTypes   = c("ids", "labels", "extra") ### Types of columns to include: IDs, labels, or extra. If only labels, will return labels without the "_label"
+){
+  ### Get objects from FrEDI
+  co_sectors  <- "co_sectors"     |> get_frediDataObj("frediData")
+  co_variants <- "co_variants"    |> get_frediDataObj("frediData")
+  co_impTypes <- "co_impactTypes" |> get_frediDataObj("frediData")
+  co_impYears <- "co_impactYears" |> get_frediDataObj("frediData")
+  co_regions  <- "co_regions"     |> get_frediDataObj("frediData")
+  co_states   <- "co_states"      |> get_frediDataObj("frediData")
+  co_modTypes <- "co_modelTypes"  |> get_frediDataObj("frediData")
+  co_models   <- "co_models"      |> get_frediDataObj("frediData")
+
+  ### Conditionals
+  colTypes    <- colTypes |> tolower()
+  doIds       <- "ids"    %in% colTypes
+  doLabs      <- "labels" %in% colTypes
+  doExtra     <- "extra"  %in% colTypes
+  onlyLabs    <- !doIds
+
+  ### Initialize some values as empty vectors
+  ### Adjust values in vectors depending on conditionals
+  colsReg0    <- c()
+  colsMod0    <- c()
+  if(addRegions) colsReg0 <- c("region", "state", "postal")
+  if(addModels ) colsMod0 <- c("model")
+
+  ### Column names
+  colsData0   <- c("sector", "variant", "impactType", "impactYear") |> c(colsReg0) |> c("modelType") |> c(colsMod0)
+  colsIds0    <- colsData0 |> get_matches(y=c("modelType", "state", "postal"), matches=FALSE) |> paste0("_id")
+  colsLabs0   <- colsData0 |> get_matches(y=c("modelType", "state", "postal"), matches=FALSE) |> paste0("_label")
+  colsVars    <- c("sectorprimary", "includeaggregate", "damageAdjName")
+  colsTypes   <- c("impactType_description", "physicalmeasure") |>
+    c(c("physScalar", "physAdj", "econScalar", "econMultiplier") |> paste0("Name")) |>
+    c("c0", "c1", "exp0", "year0")
+  colsMods0   <- c("maxUnitValue", "inputName") |>
+    c("model" |> paste0(c("UnitDesc", "Unit_id", "Unit_label"))) |>
+    c("model" |> paste0(c("UnitScale", "RefYear", "MaxOutput", "MaxExtrap")))
+  colsMods0   <- colsMods0 |> get_matches(y=c("model" |> paste0(c("UnitScale", "RefYear", "MaxOutput", "MaxExtrap"))), matches=F)
+  colsOth0    <- c()
+
+  ### Add additional columns
+  if(doExtra) {
+    colsOth0 <- c(colsVars, colsTypes, colsMods0)
+    if(addModels) colsOth0 <- colsOth0 |> c("maxUnitValue")
+  } ### if(doAll)
+
+
+  ### Filter data
+  hasSectors  <- sectors0 |> length()
+  if(hasSectors) co_sectors <- co_sectors |> filter(sector_id %in% sectors0)
+
+  ### Rename columns
+  renameAt0   <- c("modelType")
[…]
