### 2024-08-07 9fc6024
Fixed code for SLR sectors
Apparently fixed code for SLR sectors...still need to check against previous data
