+++ FrEDI/R/utils.R
@@ -800,7 +800,8 @@ extend_slrScalars <- function(
   return(df0)
 }
 
-
+###### get_co_sectorsInfo ######
+### This helper function helps get info about sector groups
 get_co_sectorsInfo <- function(
     sectors0   = NULL,  ### Sector IDs
     addRegions = FALSE, ### Whether to include regions & states
@@ -843,12 +844,12 @@ get_co_sectorsInfo <- function(
     c("model" |> paste0(c("UnitDesc", "Unit_id", "Unit_label"))) |>
     c("model" |> paste0(c("UnitScale", "RefYear", "MaxOutput", "MaxExtrap")))
   colsMods0   <- colsMods0 |> get_matches(y=c("model" |> paste0(c("UnitScale", "RefYear", "MaxOutput", "MaxExtrap"))), matches=F)
-  colsOth0    <- c()
 
   ### Add additional columns
+  colsOth0    <- c()
   if(doExtra) {
-    colsOth0 <- c(colsVars, colsTypes, colsMods0)
-    if(addModels) colsOth0 <- colsOth0 |> c("maxUnitValue")
+    colsOth0 <- c(colsVars, colsTypes)
+    if(addModels) colsOth0 <- colsOth0 |> c(colsMods0, "maxUnitValue")
   } ### if(doAll)
 
 
@@ -1202,16 +1203,25 @@ initialize_resultsDf <- function(
 
   ###### FrEDI Data ######
   ### Get FrEDI data objects and format data
-  df_info    <- "co_sectorsInfo" |> get_frediDataObj("frediData")  ### Tibble with info on SLR scalars
+  # df_info    <- "co_sectorsInfo" |> get_frediDataObj("frediData")  ### Tibble with info on SLR scalars
   slrScalars <- "co_slrScalars"  |> get_frediDataObj("frediData")  ### Tibble with info on SLR scalars
   df_scalars <- "df_scalars"     |> get_frediDataObj("stateData")  ### Tibble of main scalars
 
+  ### Get info
+  df_info    <- sectors |> get_co_sectorsInfo(
+    addRegions = TRUE , ### Whether to include regions & states
+    addModels  = FALSE, ### Whether to include models
+    colTypes   = c("ids", "extra") ### Types of columns to include: IDs, labels, or extra. If only labels, will return labels without the "_label"
+  ) ### End get_co_sectorsInfo()
+
   ### Format data
-  select0    <- c("sector", "variant", "impactType", "impactYear", "region") |> c(stateCols0) |> c("modelType", "model") |> paste0("_label")
+  drop0      <- c("sector", "variant", "impactType", "impactYear", "region") |> c(stateCols0) |> c("modelType") |> paste0("_label")
   # df_info |> glimpse(); select0 |> print()
   df_scalars <- df_scalars |> filter(year >= minYr0, year <= maxYr0)
   df_info    <- df_info    |> filter(sector %in% sectors)
-  df_info    <- df_info    |> select(-any_of(select0))
+  df_info    <- df_info    |> select(-any_of(drop0))
+  df_info    <- df_info    |> distinct()
+  rm(drop0)
 
   ### Model Types
   types0     <- df_info |> pull(modelType) |> unique() |> tolower()
@@ -1222,7 +1232,6 @@ initialize_resultsDf <- function(
   # df_se |> glimpse(); df_info |> glimpse(); # df_scalars |> glimpse()
   join0      <- df_info |> names() |> get_matches(df_se |> names())
   df0        <- df_info |> left_join(df_se, by=c(join0), relationship="many-to-many")
-  # df0        <- df0     |> select(-all_of(join0))
   rm(join0)
   # df0 |> glimpse(); # df0 |> dim() |> print()
   # return(df0)
[…]
