### 2024-08-07 20ff365
Fixed issue with duplicate GCM impacts
