+++ FrEDI/R/aggregate_impacts.R
@@ -67,6 +67,7 @@ aggregate_impacts <- function(
   ####### State Cols  ######
   stateCols0   <- c("state", "postal")
   popCol0      <- c("pop")
+  national0    <- c("National Total")
 
   ###### Format Data ######
   ### Ungroup
@@ -75,11 +76,10 @@ aggregate_impacts <- function(
 
   ###### Years Info ######
   ### Years in data
-  c_dataYears  <- data[[yearCol0]] |> unique()
+  years0  <- data |> pull(all_of(yearCol0)) |> unique()
 
   ###### Aggregation Levels  ######
   ### Types of summarization to do: default
-  # aggList0     <- c("national", "modelaverage", "impactyear", "impacttype", "all")
   aggList0     <- c("national", "modelaverage", "impacttype", "impactyear")
   null_aggLvls <- aggLevels |> is.null()
   aggLevels    <- aggLevels |> tolower()
@@ -102,7 +102,6 @@ aggregate_impacts <- function(
   ### Check if aggregation required
   requiresAgg <- aggLevels |> length() > 0
 
-  ###### Aggregation Level Options  ######
   ### Aggregate to impact years or national...reduces the columns in output
   aggImpYear  <- "impactyear"   %in% aggLevels
   aggImpTypes <- "impacttype"   %in% aggLevels
@@ -111,41 +110,46 @@ aggregate_impacts <- function(
 
   ### Filter data
   ### If "national" aggregation, filter out national totals
-  if(aggNational){data <- data |> filter(region!="National Total")}
   ### If modelAverage %in% aggLevels, filter out model averages
+  if(aggNational){data <- data |> filter(region!=national0)}
   if(aveModels  ){data <- data |> filter(!(model %in% c("Average", "Model Average")))}
 
   ###### Get FrEDI Data Objects ######
   ### Load info tables from sysdata.rda
-  co_models   <- "co_models"   |> get_frediDataObj("frediData")
-  co_sectors  <- "co_sectors"  |> get_frediDataObj("frediData")
-  co_variants <- "co_variants" |> get_frediDataObj("frediData")
+  # co_models   <- "co_models"     |> get_frediDataObj("frediData")
+  # co_sectors  <- "co_sectors"    |> get_frediDataObj("frediData")
+  # co_variants <- "co_variants"   |> get_frediDataObj("frediData")
+  # co_sectInfo <- "co_sectorInfo" |> get_frediDataObj("frediData")
+  co_sectInfo <- get_co_sectorsInfo(addRegions=F, addModels=F, colTypes=c("ids", "labels"))
 
-  ###### - Formate SLR Info ######
-  co_slrs     <- co_models |> filter(modelType=="slr")
-  co_slrs     <- co_slrs   |> mutate_at(c("model_id", "model_label"), as.character)
-
-  ###### - Format Sector Info ######
-  rename0     <- c("sector_id", "sector_label", "modelType")
-  rename1     <- c("sector_id", "sector", "model_type")
-  co_sectors  <- co_sectors |> select(all_of(rename0))
-  co_sectors  <- co_sectors |> rename_at(c(rename0), ~rename1)
-  co_sectors  <- co_sectors |> mutate_at(c("model_type"), toupper)
-  rm(rename0, rename1)
-
-  ###### - Format Variant Info ######
-  ### Format variants
-  rename0     <- c("sector_id", "variant_label")
-  rename1     <- c("sector_id", "variant")
-  select0     <- c(rename1) #|> c("sectorprimary", "includeaggregate")
-  co_variants <- co_variants |> rename_at(c(rename0), ~rename1)
[…]
