### 2024-08-07 3ace166
Update aggregate_impacts.R
Corrected issue with impact years
