### 2024-08-07 3f05509
Update utils.R
Corrected issue with joining scalars
