### 2024-08-07 f69fc61
Update utils_import_inputs.R
Fixed a line when check if data is present
