### 2024-08-07 cbde61e
Update run_fredi.R
Corrected issue with input list names
