### 2024-08-08 a05152f
Update aggregate_impacts.R
Deleted deprecated lines
