+++ FrEDI/R/aggregate_impacts.R
@@ -116,40 +116,7 @@ aggregate_impacts <- function(
 
   ###### Get FrEDI Data Objects ######
   ### Load info tables from sysdata.rda
-  # co_models   <- "co_models"     |> get_frediDataObj("frediData")
-  # co_sectors  <- "co_sectors"    |> get_frediDataObj("frediData")
-  # co_variants <- "co_variants"   |> get_frediDataObj("frediData")
-  # co_sectInfo <- "co_sectorInfo" |> get_frediDataObj("frediData")
   co_sectInfo <- get_co_sectorsInfo(addRegions=F, addModels=F, colTypes=c("ids", "labels"))
-
-  # ### Format sector info
-  # select0     <- c("sector", "variant", "impactType", "impactYear") |>
-  #   c("region") |> c(stateCols0) |>
-  #   c("modelType", "model")
-  # renameAt0   <- select0
-  # rename
-  #
-  # ###### - Format Sector Info ######
-  # rename0     <- c("sector_id", "sector_label", "modelType")
-  # rename1     <- c("sector_id", "sector", "model_type")
-  # co_sectors  <- co_sectors |> select(all_of(rename0))
-  # co_sectors  <- co_sectors |> rename_at(c(rename0), ~rename1)
-  # co_sectors  <- co_sectors |> mutate_at(c("model_type"), toupper)
-  # rm(rename0, rename1)
-  #
-  # ###### - Format Variant Info ######
-  # ### Format variants
-  # rename0     <- c("sector_id", "variant_label")
-  # rename1     <- c("sector_id", "variant")
-  # select0     <- c(rename1) #|> c("sectorprimary", "includeaggregate")
-  # co_variants <- co_variants |> rename_at(c(rename0), ~rename1)
-  # co_variants <- co_variants |> select(all_of(select0))
-  # rm(rename0, rename1, select0)
-  # ### Combine sector and variant info
-  # join0          <- "sector_id"
-  # co_sectorVars  <- co_sectors |> left_join(co_variants, by =join0)
-  # co_sectorVars  <- co_sectorVars |> select(-all_of(join0))
-  # rm(join0, co_variants)
 
   ###### Grouping Columns  ######
   ### Use default group by columns if none specified...otherwise, check which are present
@@ -187,7 +154,6 @@ aggregate_impacts <- function(
     ### If hasDropCols
     if(hasDropCols){
       ### Drop levels
-      # groupByCols  <- groupByCols |> (function(y){y[!(y %in% dropCols)]})()
       groupByCols  <- groupByCols |> get_matches(y=dropCols, matches=F)
     } ### End if(hasDropCols)
 
@@ -210,7 +176,6 @@ aggregate_impacts <- function(
 
   ###### Summary Columns  ######
   ### Columns to summarize
-  # if(!is.null(columns)){summaryCols <- columns}
   summaryCols0 <- c("annual_impacts")
   nullSumCols  <- columns |> is.null()
   if(nullSumCols) summaryCols <- summaryCols0
@@ -226,18 +191,19 @@ aggregate_impacts <- function(
   summaryCols <- summaryCols[isPresent0]
   rm(summaryCols0, nullSumCols, isPresent0, hasNaCols0)
 
+  ### Scalar columns
+  # scalarCols  <- c("physScalar", "physAdj", "damageAdj", "econScalar", "econAdj", "econMultiplier") |> paste("Value")
+  # scalarCols  <- scalarCols |> c("physScalar", "econScalar", "physEconScalar")
+  # scalarCols  <- c("c0", "c1", "exp0", "year0") |> c(scalarCols)
+
   ### Drop some columns from summary columns
   if(aggImpTypes){
[…]
