### 2024-08-08 2bb4228
Added lines to make sure national totals are dropped from population inputs
Added lines to make sure national totals are dropped from population inputs
