### 2024-08-13 9cafd35
Create run_fredi_methane.R
Initialized methane function
