+++ FrEDI/R/run_fredi_methane.R
@@ -1,0 +1,382 @@
+###### Documentation ######
+#' Project annual average impacts from methane, NOx, and ozone.
+#'
+#'
+#'
+#' @description
+#' This function allows users to estimate impacts from changes to methane, NOx, and ozone. Additional documentation to follow.
+
+###### run_fredi_methane ######
+### This function creates a data frame of sector impacts for default values or scenario inputs.
+### run_fredi relies on the following helper functions: "interpolate_annual", "match_scalarValues","get_econAdjValues" , "calcScalars", "interpolate_tempBin"
+run_fredi_methane <- function(
+    inputsList = list(o3 = NULL, ch4=NULL, nox=NULL, gdp=NULL, pop=NULL), ### List of inputs
+    elasticity = 1,    ### Override value for elasticity for economic values
+    maxYear    = 2100,
+    thru2300   = FALSE,
+    outputList = FALSE, ### Whether to return input arguments as well as results. [If TRUE], returns a list instead of a data frame
+    allCols    = FALSE, ### Whether to include additional columns in output
+    silent     = TRUE   ### Whether to message the user
+){
+  ###### Load Objects ######
+  ### Assign data objects to objects in this namespace
+  ### Assign FrEDI config
+  fredi_config <- rDataList[["fredi_config"]]
+  for(name_i in fredi_config |> names()) {name_i |> assign(fredi_config[[name_i]]); rm(name_i)}
+
+
+  ###### Set up the environment ######
+  ### Level of messaging (default is to message the user)
+  msgUser   <- !silent
+  ### Uncomment for allCols
+  # doPrimary <- F  ### whether to filter to primary impacts
+  ### Model years and NPD (FrEDI past 2100)
+  minYear   <- listMethane[["package"]][["coefficients"]][["minYear0"]]
+  maxYear   <- thru2300 |> ifelse(npdYear0, maxYear)
+  do_npd    <- maxYear > maxYear0
+  # "got here2" |> print()
+
+  ###### ** Return List ######
+  ### Initialize list to return
+  returnList <- list() ### List to return args, scenarios, and statuses
+  argsList   <- list() ### List of arguments
+  statusList <- list() ### List to return custom or default
+
+  ### Initialize return list: add scenarios
+  if(outputList) {returnList[["scenarios"]] <- list()}
+
+
+  ### Initialize status list
+  ### Add statuses:
+  ### - inputsList items and aggLevels assessed further below
+  ### Add to list
+  if(outputList){
+    statusList[["inputsList"]] <- inputsList
+    statusList[["elasticity"]] <- (elasticity == 1) |> get_returnListStatus()
+    statusList[["maxYear"   ]] <- (maxYear == maxYear0 & !thru2300) |> get_returnListStatus()
+    statusList[["thru2300"  ]] <- (!thru2300) |> get_returnListStatus()
+    statusList[["allCols"   ]] <- (!allCols ) |> get_returnListStatus()
+    statusList[["silent"    ]] <- ( silent  ) |> get_returnListStatus()
+  } ### End if(outputList)
+
+
+  ### Initialize arguments list
+  ### - inputsList items, sectorList, and aggLevels assessed further below
+  if(outputList){
+    argsList[["inputsList"]] <- inputsList
+    argsList[["elasticity"]] <- elasticity
+    argsList[["maxYear"   ]] <- maxYear
+    argsList[["thru2300"  ]] <- thru2300
[…]
