+++ FrEDI/man/run_fredi_sv.Rd
@@ -32,7 +32,7 @@ run_fredi_sv(
 The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing average annual physical impacts for socially vulnerable groups, at the NCA region level and five-year increments.
 }
 \description{
-\code{run_fredi_sv} allows users to project annual average climate change impacts throughout the 21st century (2010-2100) for socially vulnerable (SV) populations for available sectors. Users can run \code{\link[=run_fredi_sv]{run_fredi_sv()}} for individual sectors to generate annual physical impacts for SV populations. \code{\link[=run_fredi_sv]{run_fredi_sv()}} can be run with default population and climate (temperature and sea level rise trajectories) or provide custom trajectories. The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing annual average physical impacts at five-year increments for the period 2010 to 2100. Users have the option to write outputs to Excel files that provide additional visualization of SV outputs.
+\code{run_fredi_sv} allows users to project annual average climate change impacts throughout the 21st century (2010-2100) for socially vulnerable (SV) populations for available sectors. Users can run \code{\link[=run_fredi_sv]{run_fredi_sv()}} for individual sectors to generate annual physical impacts for SV populations. \code{\link[=run_fredi_sv]{run_fredi_sv()}} can be run with default population and climate (temperature and sea level rise trajectories) or using custom trajectories. The output of \code{\link[=run_fredi_sv]{run_fredi_sv()}} is an R data frame object containing annual average physical impacts at five-year increments for the period 2010 to 2100. The basic structure, specific methodology, and underlying data supporting FrEDI-SV are derived from EPA’s independently peer-reviewed September 2021 report, \href{https://epa.gov/cira/social-vulnerability-report}{Climate Change and Social Vulnerability in the United States: A Focus on Six Impacts}
 }
 \details{
 \code{\link[=run_fredi_sv]{run_fredi_sv()}} projects annual climate change impacts for socially vulnerable (SV) populations throughout the 21st century (2010-2100) for available sectors, using default or user-specified population, temperature, and sea level rise (SLR) trajectories. \code{\link[=run_fredi_sv]{run_fredi_sv()}} is the main function for the FrEDI Social Vulnerability (SV) module in the \link{FrEDI} R package, described elsewhere (See \url{https://epa.gov/cira/FrEDI} for more information). The SV module extends the \link{FrEDI} framework to socially vulnerable populations using data underlying a 2021 U.S. Environmental Protection Agency (EPA) report on \href{https://www.epa.gov/cira/social-vulnerability-report/}{Climate Change and Social Vulnerability in the United States}.
