### 2024-08-14 3719079
Update import_inputs.R
Updated `import_inputs` function to allow import of functions for methane
