### 2024-08-14 39b5422
Update run_fredi.R
Edited `run_fredi` to work with updates to `import_inputs` utilities
