### 2024-08-15 d1f1f4e
Update data.R
Updated documentation for GCAM scenarios and pop scenario. Added new GDP scenario.
