+++ FrEDI/R/data.R
@@ -32,42 +32,55 @@
 ###### gcamScenarios ######
 #' Six driver scenarios that can be passed as inputs to [FrEDI::run_fredi()] and [FrEDI::run_fredi_sv()]
 #'
-#' A dataframe containing six driver scenarios that can be passed as inputs to [FrEDI::run_fredi()] and [FrEDI::run_fredi_sv()]. This data frame has four columns -- `year`, `temp_C`, `slr_cm`, and `scenario` -- respectively containing:
+#' `gcamScenarios` is a data frame object containing six driver scenarios that can be passed as inputs to [FrEDI::run_fredi()] and [FrEDI::run_fredi_sv()]. This data frame has five columns -- `year`, `temp_C_global`, `temp_C_conus`, `slr_cm`, `scenario`, and `model`:
 #'
 #' \describe{
-#'   \item{year}{The Common Era (CE) year associated with the observation.}
-#'   \item{temp_C}{Global temperature (i.e., degrees of warming above the baseline year of 1995) in degrees Celsius, for the associated year and scenario.}
-#'   \item{slr_cm}{Global Mean Sea Level Rise (i.e., SLR) in centimeters for the associated year and scenario.}
-#'   \item{scenario}{Associated scenario identifier.}
+#'   \item{year         }{The Common Era (CE) year associated with the observation.}
+#'   \item{temp_C_global}{Global temperature (i.e., degrees of warming above the baseline year of 1995), in degrees Celsius, for the associated year and scenario.}
+#'   \item{temp_C_conus }{Temperatures (i.e., degrees of warming above the baseline year of 1995) for the contiguous U.S. (CONUS), in degrees Celsius, for the associated year and scenario (calculated from global temperatures using [FrEDI::convertTemps(from="global")]).}
+#'   \item{slr_cm       }{Global Mean Sea Level Rise (GMSL or SLR), in centimeters, for the associated year and scenario (calculated from global temperatures using [FrEDI::temps2slr()]).}
+#'   \item{scenario     }{Associated scenario identifier (e.g., `"ECS_3.0_REF"`).}
+#'   \item{model        }{A string (`"Hector_GCAM_v5.3"`) identifying the model (**Hector**, with **GCAM v5.3**) used in generating the global temperatures associated with each scenario.}
 #' }
 #'
-#' The scenarios in this dataframe were created using [Hector](https://jgcri.github.io/hector/), an open-source, reduced-form global carbon-cycle climate model (Hartin et al., 2015) to model temperatures associated with emissions scenarios from the Global Change Analysis Model v5.3 (GCAM). The Global Change Analysis Model v5.3 ([GCAM](https://gcims.pnnl.gov/modeling/gcam-global-change-analysis-model)) is an open source model that represents the linkages between energy, water, land, climate and economic systems (Calvin et al., 2019). Scenario identifiers in the `scenario` column of [FrEDI::gcamScenarios()] have the string `"Hector_GCAM_v5.3_ECS_"` as a prefix, followed by the average warming temperature and the suffix `"_REF"` (e.g., `"Hector_GCAM_v5.3_ECS_3.0_REF"` for the default scenario for [FrEDI::run_fredi()] and [FrEDI::run_fredi_sv()].
+#' The scenarios in this data frame were created using **Hector** with **GCAM v5.3**:
 #'
[…]
