### 2024-08-15 0b68067
Update data.R
Edited description of regions
