### 2024-08-15 4a393ff
Update run_fredi.R
Fixed issues with using default scenarios
