### 2024-08-15 32af0d8
Fixed small issues with input process
