+++ FrEDI/R/utils_import_inputs.R
@@ -341,27 +341,27 @@ check_valCols <- function(
   ###### Check Values ######
   ### Initial names and values
   valCol0   <- valCol
-  namesDF0  <- inputDf |> names()
+  namesDf0  <- inputDf |> names()
   inName0   <- inputName
   ### Convert to lowercase
   valCol    <- valCol    |> tolower()
-  namesDF   <- namesDF0  |> tolower()
+  namesDf   <- namesDf0  |> tolower()
   inName    <- inputName |> tolower()
   inputDf   <- inputDf   |> set_names(namesDf)
   ### Check if there is a match
-  checkVCol <- valCol %in% namesDF
-  vColMatch <- namesDF[checkVCol]
+  checkVCol <- valCol %in% namesDf
+  vColMatch <- namesDf[checkVCol]
   nMatches  <- vColMatch |> length()
   ### Check if there are any string matches, and get number of matches
   if(!nMatches) {
     ### Messages
-    msg_vcol0 <- paste0("Column ", "\"", valCol, "\"", " not found in ", inName0, "file data!")
+    msg_vcol0 <- paste0("Column ", "\"", valCol0, "\"", " not found in ", inName0, "file data!")
     msg_vcol1 <- paste0("Looking for columns with matches to the string ", "\"", inName0, "\"", "...")
     msg1_i |> paste0(msg_vcol0) |> message()
     # msg2_i |> paste0(msg_vcol1) |> message()
     ### Look for matches
-    checkVCol1 <- namesDF |> str_detect(inName)
-    vColMatch1 <- namesDF[checkVCol1]
+    checkVCol1 <- namesDf |> str_detect(inName)
+    vColMatch1 <- namesDf[checkVCol1]
     nMatches1  <- vColMatch1 |> length()
     matches1   <- nMatches1 > 1
     ### If matches, check if there are more than once match
@@ -377,59 +377,35 @@ check_valCols <- function(
         ###     - Otherwise, use first column
         if(doTemp) {
           ### Try to match based on the tempType
-          findTemp   <- vColMatch1 |> str_detect(tempType)
-          vColMatch2 <- vColMatch1[findTemp]
+          # tempType |> print(); vColMatch1 |> print()
+          findTemp2  <- vColMatch1 |> str_detect(tempType)
+          vColMatch2 <- vColMatch1[findTemp2]
           nMatches2  <- vColMatch2 |> length()
-          matches2   <- nMatches2 > 1
+          matches2   <- nMatches2 > 0
           ### If no matches found, message user and change type
           if(!matches2) {
-            tempType <- ("conus" %in% tempType) |> ifelse("global", "conus")
-            findTemp   <- vColMatch1 |> str_detect(tempType)
-            vColMatch2 <- vColMatch1[findTemp]
+            tempType2  <- ("conus" %in% tempType) |> ifelse("global", "conus")
+            tempType2 |> print()
+            findTemp2  <- vColMatch1 |> str_detect(tempType2)
+            vColMatch2 <- vColMatch1[findTemp2]
             nMatches2  <- vColMatch2 |> length()
-            matches2   <- nMatches2 > 1
+            matches2   <- nMatches2 > 0
             if(matches2) {
-              msg_vcol3 <- paste0("Only a match for `tempType=", tempType, "` found! Changing `tempType`...")
+              msg_vcol3 <- paste0("Only a match for `tempType=", tempType2, "` found! Changing `tempType`...")
               msg2_i |> paste0(msg_vcol1, msg_vcol3) |> message()
+              tempType  <- tempType2
[…]
