### 2024-08-15 38626ca
Fixed issue with formatting input scenarios
