### 2024-08-15 2819a12
Fixes to update import inputs
Fixed error message for import inputs, utils fixed small variable name issue in import inputs
