### 2024-08-15 dd7dd12
Update utils.R
Small change to message for when SLR scenario is created from temperature inputs
