### 2024-08-16 4d04c00
Update run_fredi.R
Added case in input process for when running GCM sectors only
