### 2024-08-16 01072e1
Update utils.R
Modified `interpolate_annual` function to allow it to work with methane module for ozone response (added new grouping for model)
