+++ FrEDI/R/utils.R
@@ -135,7 +135,8 @@ interpolate_annual <- function(
     rule      = NULL, ### for interpolation,
     method    = NULL, ### method for interpolation; default=linear
     region    = "NationalTotal", ### Region if "addRegion"
-    byState   = FALSE ### If breakdown by state
+    byState   = FALSE, ### If breakdown by state
+    byModel   = FALSE  ### If breakdown by model
 ){
   ###### Data Info ######
   ##### Other values
@@ -143,20 +144,22 @@ interpolate_annual <- function(
   rm(region)
   ##### By state
   if (byState) {stateCols0 <- c("state", "postal")} else{stateCols0 <- c()}
+  if (byModel) {modelCols0 <- c("model")} else{modelCols0 <- c()}
   ##### Columns
   dataCols  <- data |> names()
-  defCols   <- c("year", "region") |> c(stateCols0)
-  defCol0   <- dataCols[!(dataCols %in% defCols)][1]
-  column0   <- column |> is.null() |> ifelse(defCol0, column)
-  othCols   <- dataCols[!(dataCols %in% c(defCols, column0))]
+  defCols   <- c("year", "region") |> c(stateCols0) |> c(modelCols0)
+  # defCol0   <- dataCols[!(dataCols %in% defCols)][1]
+  defCol0   <- dataCols |> get_matches(y=defCols, matches=FALSE) |> first()
+  column0   <- column   |> is.null() |> ifelse(defCol0, column)
+  # othCols   <- dataCols[!(dataCols %in% c(defCols, column0))]
+  othCols   <- dataCols |> get_matches(y=defCols |> c(column0), matches=FALSE) |> first()
   # defCol0 |> print(); column0 |> print()
   rm(defCol0)
 
   ###### Format data
   # column0 |> print()
   data      <- data |> filter(!(column0 |> is.na()))
-  values0   <- data[[column0]]
-  years0    <- data[["year" ]]
+  years0    <- data |> pull(year) |> unique()
 
   ### Interpolation years
   doYears   <- years |> is.null()
@@ -186,42 +189,66 @@ interpolate_annual <- function(
   ### Filter to the region and then interpolate missing values
   cols0     <- c("x", "y")
   cols1     <- c("year") |> c(column0)
-  ### Iterate over states if byState=T
-  ### Otherwise, iterate over regions
-  if (byState) {
-    states0   <- data |> pull(state) |> unique()
-    df_interp <- states0 |> map(function(state_i){
-      ### Values
-      df_i     <- data |> filter(state==state_i)
-      x_i      <- df_i[["year" ]]
-      y_i      <- df_i[[column0]]
-      ### Approximate
-      new_i   <- approx(x=x_i, y=y_i, xout=years, rule=rule, method=method)
-      new_i   <- new_i |> as_tibble()
-      new_i   <- new_i |> rename_at(c(cols0), ~cols1)
-      new_i   <- new_i |> mutate(state=state_i)
-      # new_i |> names() |> print()
-      ### Return
-      return(new_i)
-    }) |> bind_rows()
-  } else { ### By state
-    regions0  <- data |> pull(region) |> unique()
-    df_interp <- regions0 |> map(function(region_i){
-      ### Values
-      df_i     <- data |> filter(region==region_i)
-      x_i      <- df_i[["year" ]]
-      y_i      <- df_i[[column0]]
-      ### Approximate
-      new_i   <- approx(x=x_i, y=y_i, xout=years, rule=rule, method=method)
-      new_i   <- new_i |> as_tibble()
-      new_i   <- new_i |> rename_at(c(cols0), ~cols1)
[…]
