### 2024-08-16 cf30f8d
Update utils.R
Edited `format_inputScenarios` function to work with methane module
