### 2024-08-16 acbf751
Update run_fredi_methane.R
Updated input process for methane function
