+++ FrEDI/R/run_fredi_methane.R
@@ -10,7 +10,7 @@
 ### This function creates a data frame of sector impacts for default values or scenario inputs.
 ### run_fredi relies on the following helper functions: "interpolate_annual", "match_scalarValues","get_econAdjValues" , "calcScalars", "interpolate_tempBin"
 run_fredi_methane <- function(
-    inputsList = list(o3 = NULL, ch4=NULL, nox=NULL, gdp=NULL, pop=NULL), ### List of inputs
+    inputsList = list(ch4=NULL, nox=NULL, gdp=NULL, pop=NULL), ### List of inputs
     elasticity = 1,    ### Override value for elasticity for economic values
     maxYear    = 2100,
     thru2300   = FALSE,
@@ -24,6 +24,9 @@ run_fredi_methane <- function(
   fredi_config <- rDataList[["fredi_config"]]
   for(name_i in fredi_config |> names()) {name_i |> assign(fredi_config[[name_i]]); rm(name_i)}
 
+  calc_NOx_factor <- listMethane$package$coefficients$NOx      [["fun0"]]
+  calc_mortality  <- listMethane$package$coefficients$Mortality[["fun0"]]
+  NOxFactor0      <- listMethane$package$coefficients$NOx$NOxFactor0
 
   ###### Set up the environment ######
   ### Level of messaging (default is to message the user)
@@ -92,103 +95,215 @@ run_fredi_methane <- function(
 
 
   ###### Scenarios ######
-  ###### ** Check Inputs ######
+  ###### ** Input Info ######
   paste0("Checking scenarios...") |> message()
   ### Add info to data
-  df_inputInfo <- listMethane[["package"]][["co_inputInfo"]]
+  co_inputInfo <- "co_inputInfo" |> get_frediDataObj(listSub="package", listName="listMethane")
+  # co_inputInfo <- co_inputInfo |> filter(!inputName %in% "o3")
+  co_inputInfo <- co_inputInfo |> mutate(ref_year = 2020)
+  co_inputInfo <- co_inputInfo |> mutate(min_year = 2020)
+  co_inputInfo <- co_inputInfo |> mutate(max_year = maxYear)
 
-  ### Input defaults
-  inputDefs    <- list()
-  gcamDefault  <- "gcam_default" |> get_frediDataObj("scenarioData")
-  inputDefs[["temp"]] <- gcamDefault |> select(c("year", "temp_C_conus")) |> rename_at(c("temp_C_conus"), ~"temp_C")
-  inputDefs[["slr" ]] <- gcamDefault |> select(c("year", "slr_cm"))
-  inputDefs[["gdp" ]] <- "gdp_default" |> get_frediDataObj("scenarioData")
-  inputDefs[["pop" ]] <- "pop_default" |> get_frediDataObj("scenarioData")
-  rm(gcamDefault)
+  ### Initialize subset
+  df_inputInfo <- co_inputInfo
 
   ### Input info
+  inNames0     <- co_inputInfo |> pull(inputName)
+  # inNames0 |> print()
+
+
+  ###### ** Input Columns ######
+  ### Get list with expected name of columns used for unique ids
+  ### Get list with expected name of column containing values
+  valCols0     <- co_inputInfo |> pull(valueCol) |> as.list() |> set_names(inNames0)
+  idCols0      <- list(valCols0=valCols0, df0=inputDefs[inNames0]) |> pmap(function(valCols0, df0){
+    df0 |> names() |> get_matches(y=valCols0, matches=F)
+  }) |> set_names(inNames0)
+
+
+
+  ###### ** Input Defaults ######
[…]
