### 2024-08-19 25f680b
Initialized methane function
Got methane function working and added methane utilities
