+++ FrEDI/R/run_fredi_methane.R
@@ -18,26 +18,28 @@ run_fredi_methane <- function(
     allCols    = FALSE, ### Whether to include additional columns in output
     silent     = TRUE   ### Whether to message the user
 ){
-  ###### Load Objects ######
+  ###### Set up the environment ######
+  ###### ** Messaging ######
+  ### Level of messaging (default is to message the user)
+  msgUser   <- !silent
+
+
+  ###### ** Load Data ######
   ### Assign data objects to objects in this namespace
   ### Assign FrEDI config
   fredi_config <- rDataList[["fredi_config"]]
   for(name_i in fredi_config |> names()) {name_i |> assign(fredi_config[[name_i]]); rm(name_i)}
 
-  calc_NOx_factor <- listMethane$package$coefficients$NOx      [["fun0"]]
-  calc_mortality  <- listMethane$package$coefficients$Mortality[["fun0"]]
-  NOxFactor0      <- listMethane$package$coefficients$NOx$NOxFactor0
+  ### Coefficients
+  minYear0   <- listMethane[["package"]][["coefficients"]][["minYear0"]]
+  maxYear0   <- listMethane[["package"]][["coefficients"]][["maxYear0"]]
 
-  ###### Set up the environment ######
-  ### Level of messaging (default is to message the user)
-  msgUser   <- !silent
-  ### Uncomment for allCols
-  # doPrimary <- F  ### whether to filter to primary impacts
   ### Model years and NPD (FrEDI past 2100)
-  minYear   <- listMethane[["package"]][["coefficients"]][["minYear0"]]
+  minYear   <- minYear0
   maxYear   <- thru2300 |> ifelse(npdYear0, maxYear)
   do_npd    <- maxYear > maxYear0
-  # "got here2" |> print()
+
+
 
   ###### ** Return List ######
   ### Initialize list to return
@@ -86,15 +88,16 @@ run_fredi_methane <- function(
   } ### End if
   rm(has_elasticity, elasticity0)
 
+
+
   ###### ** State Info ######
   byState        <- TRUE
-  popCol0        <- "pop"
+  popCol0        <- c("pop")
   stateCols0     <- c("state", "postal")
 
 
 
-
-  ###### Scenarios ######
+  ###### Input Scenarios ######
   ###### ** Input Info ######
   paste0("Checking scenarios...") |> message()
   ### Add info to data
@@ -116,37 +119,21 @@ run_fredi_methane <- function(
   ### Get list with expected name of columns used for unique ids
   ### Get list with expected name of column containing values
   valCols0     <- co_inputInfo |> pull(valueCol) |> as.list() |> set_names(inNames0)
-  idCols0      <- list(valCols0=valCols0, df0=inputDefs[inNames0]) |> pmap(function(valCols0, df0){
-    df0 |> names() |> get_matches(y=valCols0, matches=F)
-  }) |> set_names(inNames0)
+  idCols0      <- inNames0     |> map(get_import_inputs_idCols) |> set_names(inNames0)
+  # idCols0      <- list(valCols0=valCols0, df0=inputDefs[inNames0]) |> pmap(function(valCols0, df0){
+  #   df0 |> names() |> get_matches(y=valCols0, matches=F)
+  # }) |> set_names(inNames0)
 
 
 
   ###### ** Input Defaults ######
-  # listMethane$package$ch4_default <- listMethane$package$ch4_default |> rename(CH4_ppbv = delta_ch4)
[…]
