### 2024-08-19 991a119
Improved methane module
Improved methane module:

- fixed joins
- Formatted outputs
