+++ FrEDI/R/utils_methane.R
@@ -1,7 +1,7 @@
 ### Utility functions for the FrEDI methane module
 ### Function for mortality
 # calc_mortality  <- listMethane$package$coefficients[["Mortality"]][["fun0"]]
-calc_mortality <- function(
+calc_methane_mortality <- function(
     df0, ### Tibble with population and years
     df1      = listMethane$package$rff_nat_pop, ### Tibble with columns for mortality rate slope and mortality rate intercept
     pCol0    = "national_pop"      , ### Column with national population
@@ -9,17 +9,20 @@ calc_mortality <- function(
     iCol0    = "rffMrate_intercept", ### Column with mortality rate intercept,
     joinCols = c("year") ### Column to join df0 and df1
 ){
+  ### Select columns
+  select0 <- c("year", "ifRespScalar", "rffPop") |> c(sCol0, iCol0)
+  df1     <- df1 |> select(all_of(select0))
   ### Join df0 and df1
   # join0 <- c("year")
-  join0 <- joinCols
-  df0   <- df0 |> left_join(df1, by=join0)
+  join0   <- joinCols
+  df0     <- df0 |> left_join(df1, by=join0)
   rm(df1)
   ### Calculate intermediate populations
   # df0   <- df0 |> mutate(delta_rffPop = national_pop - rffPop)
   # df0   <- df0 |> mutate(rffFactor    = delta_rffPop * rffMrate_slope + rffMrate_intercept)
-  df0   <- df0 |> mutate(delta_rffPop = !!sym(pCol0) - rffPop)
-  df0   <- df0 |> mutate(rffFactor    = delta_rffPop * !!sym(sCol0) + !!sym(iCol0))
-  df0   <- df0 |> mutate(respMrate    = rffFactor    * ifRespScalar)
+  df0     <- df0 |> mutate(delta_rffPop = !!sym(pCol0) - rffPop)
+  df0     <- df0 |> mutate(rffFactor    = delta_rffPop * !!sym(sCol0) + !!sym(iCol0))
+  df0     <- df0 |> mutate(respMrate    = rffFactor    * ifRespScalar)
   ### Return data
   return(df0)
 }
@@ -67,6 +70,32 @@ calc_o3_conc <- function(
 }
 
 
+### Format methane drivers
+format_methane_drivers <- function(
+    df0, ### Tibble with scenarios
+    df1=listMethane$package$state_rrScalar
+){
+  ### Format data
+  idCols0  <- c("region", "state", "postal", "model")
+  sumCols0 <- c("state_o3response_pptv_per_ppbv", "state_mortScalar")
+  select0  <- idCols0 |> c(sumCols0)
+  df1      <- df1     |> select(all_of(select0))
+  ### Join data
+  join0    <- df0 |> names() |> get_matches(y=df1 |> names())
+  df0      <- df0 |> left_join(df1, by=join0, relationship="many-to-many")
+  rm(df1)
+  ### Calculate O3 if methane and nox present
+  do_o3    <- df0 |> names() |> get_matches(y=c("CH4_ppbv")) |> length()
+  if(do_o3) df0 <- df0 |> calc_o3_conc()
+  ### Reorganize values
+  move0   <- c("region", "state", "postal", "model") |> c("year")
+  df0     <- df0 |> relocate(all_of(move0))
+  df0     <- df0 |> arrange_at(c(move0))
+  ### Return
+  return(df0)
+}
+
+
 ### Function to calculate CONUS scenario
 calc_conus_scenario <- function(
     df0 ### Tibble with state population and GDP information
@@ -76,13 +105,18 @@ calc_conus_scenario <- function(
   ### Join with df0
   filter0 <- c("AK", "HI")
   group0  <- c("year")
[…]
