+++ FrEDI/R/run_fredi_methane.R
@@ -139,8 +139,7 @@ run_fredi_methane <- function(
     ### Return
     return(df0)
   }) |> set_names(inNames0)
-
-
+  # inputDefs$o3 |> glimpse()
 
 
 
@@ -165,25 +164,29 @@ run_fredi_methane <- function(
 
   ### Need scenario for CH4 & NOX or O3:
   ### If has O3, use O3. Otherwise, use CH4
-  has_o3     <- !(inputsList[["o3" ]] |> is.null())
-  has_o3 |> print()
-  # has_ch4    <- !(inputsList[["ch4"]] |> is.null())
+  has_o3     <- inputsList[["o3" ]] |> nrow() |> length()
+  has_ch4    <- inputsList[["ch4"]] |> nrow() |> length()
   if(has_o3) {
-    inputsList <- inputsList |> (function(list0, y=c("ch4", "nox")){list0[!((list0 |> names() %in% y))]})()
-    inputDefs  <- inputDefs  |> (function(list0, y=c("ch4", "nox")){list0[!((list0 |> names() %in% y))]})()
-    inNames0   <- inNames0   |> get_matches(y=c("ch4", "nox"), matches=F)
-    inNames    <- inNames    |> get_matches(y=c("ch4", "nox"), matches=F)
-  } else {
-    inputsList <- inputsList |> (function(list0, y=c("o3")){list0[!((list0 |> names() %in% y))]})()
-    inputDefs  <- inputDefs  |> (function(list0, y=c("o3")){list0[!((list0 |> names() %in% y))]})()
-    inNames0   <- inNames0   |> get_matches(y=c("o3"), matches=F)
-    inNames    <- inNames    |> get_matches(y=c("o3"), matches=F)
+    drop0      <- c("ch4", "nox")
+    inputsList <- inputsList |> (function(list0, y=drop0){list0[!((list0 |> names() %in% y))]})()
+    inputDefs  <- inputDefs  |> (function(list0, y=drop0){list0[!((list0 |> names() %in% y))]})()
+    inNames0   <- inNames0   |> get_matches(y=drop0, matches=F)
+    inNames    <- inNames    |> get_matches(y=drop0, matches=F)
+    rm(drop0)
+  } else if(has_ch4) {
+    drop0      <- c("o3")
+    inputsList <- inputsList |> (function(list0, y=drop0){list0[!((list0 |> names() %in% y))]})()
+    inputDefs  <- inputDefs  |> (function(list0, y=drop0){list0[!((list0 |> names() %in% y))]})()
+    inNames0   <- inNames0   |> get_matches(y=drop0, matches=F)
+    inNames    <- inNames    |> get_matches(y=drop0, matches=F)
+    rm(drop0)
   } ### End if(has_o3)
 
 
   ###### ** Check Inputs ######
   ### Filter to valid inputs & get info
   ### Reorganize inputs list
+  # inputDefs |> names() |> print()
   df_inputInfo <- df_inputInfo |> filter(inputName %in% inNames)
   inNames      <- df_inputInfo |> pull(inputName)
 
@@ -285,10 +288,9 @@ run_fredi_methane <- function(
 
   ###### Physical Driver Scenario  ######
   ### Need scenario for CH4 & NOX or O3
-  has_ch4    <- inputsList[["ch4"]] |> nrow() |> length()
-  has_nox    <- inputsList[["nox"]] |> nrow() |> length()
   has_o3     <- inputsList[["o3" ]] |> nrow() |> length()
-  has_driver <- (has_ch4 & has_nox) | has_o3
+  has_ch4    <- inputsList[["ch4"]] |> nrow() |> length()
+  has_driver <- has_o3 | has_ch4
[…]
