### 2024-08-20 84ad5f3
Fixed issue with interpolation
Fixed issue with selecting columns and interpolating annual values (fix due to update in grouping method for `interpolate_annual` function)
