+++ FrEDI/R/utils.R
@@ -147,13 +147,14 @@ interpolate_annual <- function(
   if (byModel) {modelCols0 <- c("model")} else{modelCols0 <- c()}
   ##### Columns
   dataCols  <- data |> names()
-  defCols   <- c("year", "region") |> c(stateCols0) |> c(modelCols0)
+  defCols   <- c("region") |> c(stateCols0) |> c(modelCols0) |> c("year")
   # defCol0   <- dataCols[!(dataCols %in% defCols)][1]
   defCol0   <- dataCols |> get_matches(y=defCols, matches=FALSE) |> first()
   column0   <- column   |> is.null() |> ifelse(defCol0, column)
   # othCols   <- dataCols[!(dataCols %in% c(defCols, column0))]
   othCols   <- dataCols |> get_matches(y=defCols |> c(column0), matches=FALSE) |> first()
   # defCol0 |> print(); column0 |> print()
+  # data |> glimpse(); dataCols |> print()
   rm(defCol0)
 
   ###### Format data
@@ -171,6 +172,7 @@ interpolate_annual <- function(
 
   ##### Regions
   addRegion <- !("region" %in% dataCols)
+  # addRegion |> print()
   if (addRegion) {data <- data |> mutate(region = region0)}
   rm(addRegion)
 
@@ -192,7 +194,9 @@ interpolate_annual <- function(
 
   ### Get IDs
   ### Subset data
-  group0 <- data[,"region" |> c(stateCols0, modelCols0)]
+  group0 <- "region" |> c(stateCols0, modelCols0)
+  group0 <- data |> select(any_of(group0))
+  # group0 <- data[,"region" |> c(stateCols0, modelCols0)]
   ### Get scenario IDs
   group0 <- group0 |> apply(1, function(x){x |> as.vector() |> paste(collapse ="_")}) |> unlist()
   data   <- data   |> mutate(group_id = group0)
@@ -213,66 +217,29 @@ interpolate_annual <- function(
     ### Return
     return(new_i)
   }) |> bind_rows()
-  # ### Iterate over states if byState=T
-  # ### Otherwise, iterate over regions
-  # if (byState) {
-  #   states0   <- data |> pull(state) |> unique()
-  #   df_interp <- states0 |> map(function(state_i){
-  #     ### Values
-  #     df_i     <- data |> filter(state==state_i)
-  #     x_i      <- df_i[["year" ]]
-  #     y_i      <- df_i[[column0]]
-  #     ### Approximate
-  #     new_i   <- approx(x=x_i, y=y_i, xout=years, rule=rule, method=method)
-  #     new_i   <- new_i |> as_tibble()
-  #     new_i   <- new_i |> rename_at(c(cols0), ~cols1)
-  #     new_i   <- new_i |> mutate(state=state_i)
-  #     # new_i |> names() |> print()
-  #     ### Return
-  #     return(new_i)
-  #   }) |> bind_rows()
-  # } else { ### By state
-  #   regions0  <- data |> pull(region) |> unique()
-  #   df_interp <- regions0 |> map(function(region_i){
-  #     ### Values
-  #     df_i     <- data |> filter(region==region_i)
-  #     x_i      <- df_i[["year" ]]
-  #     y_i      <- df_i[[column0]]
-  #     ### Approximate
-  #     new_i   <- approx(x=x_i, y=y_i, xout=years, rule=rule, method=method)
-  #     new_i   <- new_i |> as_tibble()
-  #     new_i   <- new_i |> rename_at(c(cols0), ~cols1)
-  #     new_i   <- new_i |> mutate(region = region_i)
-  #     # new_i |> names() |> print()
-  #     ### Return
-  #     return(new_i)
-  #   }) |> bind_rows()
[…]
