### 2024-08-20 e4c69ee
Fixed issue with mortality scalar
Fixed issue with base mortality scalar: previously, multiplying rrScalar by base mortality rate. Now, dividing base mortality rate by rrScalar
