### 2024-08-20 0267f8d
Fixed issue with standardizing regions
