### 2024-08-20 60b65b1
Additional fixes for methane
- Fixed issue with `interpolate_annual`: improved grouping
- Fixed issue with standardizing region and model names
