### 2024-08-20 4a4c1a0
Update run_fredi_methane.R
Added driver values, units to outputs
