### 2024-08-20 d35015d
Edited region names in population scenario
- Edited region names in population scenario (had dots instead of spaces)
- Added step to `check_regions()` function in "utils_import_inputs.R" to standardize region in user inputs
- Moved step to check regions for pop, o3 inputs before calculation of population
