### 2024-08-20 970ced6
Added steps to clear up memory
Added steps wtih `gc()` to free up memory
