### 2024-08-20 00426bf
Added steps to clear memory
Added lines with `gc()` to free up unused memory
