### 2024-08-20 0b9a071
Fixed issue with formatting input scenarios for SV
- Step where temps, SLR were zero out was causing issues with scenario ID for SV
- Added step to keep scenario ID
