+++ FrEDI/R/run_fredi_sv.R
@@ -230,8 +230,11 @@ run_fredi_sv <- function(
 
 
   ###### ** Input Columns ######
-  ### Get list with expected name of columns used for unique ids
+  ### Get list with minimum, maximum years associated with inputs
+  ### Get lists with expected name of columns used for unique ids
   ### Get list with expected name of column containing values
+  minYrs0      <- inNames0 |> map(function(name0, df0=co_inputInfo){df0 |> filter(inputName==name0) |> pull(min_year) |> unique()})
+  maxYrs0      <- inNames0 |> map(function(name0, df0=co_inputInfo){df0 |> filter(inputName==name0) |> pull(max_year) |> unique()})
   valCols0     <- co_inputInfo |> pull(valueCol) |> as.list() |> set_names(inNames0)
   idCols0      <- list(valCols0=valCols0, df0=inputDefs[inNames0]) |> pmap(function(valCols0, df0){
     df0 |> names() |> get_matches(y=valCols0, matches=F)
@@ -242,7 +245,6 @@ run_fredi_sv <- function(
   ### Figure out which inputs are not null, and filter to that list
   ### inputsList Names
   inNames      <- inputsList |> names()
-  # inWhich      <- inNames    |> map(function(name0, list0=inputsList){(!(list0[[name0]] |> is.null())) |> which()}) |> unlist() |> unique()
   inWhich      <- inNames    |> map(function(name0, list0=inputsList){!(list0[[name0]] |> is.null())}) |> unlist() |> which()
   ### Filter to values that are not NULL
   inputsList   <- inputsList[inWhich]
@@ -288,25 +290,24 @@ run_fredi_sv <- function(
 
     ### Unlist one level and format names
     inputsList <- inputsList |> unlist(recursive=FALSE)
-    inNames    <- inputsList |> names() |> str_replace(pattern=inNames |> names() |> paste0("\\.") |> paste(collapse="|"), "")
-    inputsList <- inputsList |> set_names(inNames)
-    idCols0    <- inNames |> map(function(name0, list0=idCols0     ){list0[[name0]]}) |> set_names(inNames)
-    valCols0   <- inNames |> map(function(name0, list0=valCols0    ){list0[[name0]]}) |> set_names(inNames)
-    minYrs0    <- inNames |> map(function(name0, df0=df_inputInfo){df0 |> filter(inputName==name0) |> pull(min_year) |> unique()}) |> set_names(inNames)
-    maxYrs0    <- inNames |> map(function(name0, df0=df_inputInfo){df0 |> filter(inputName==name0) |> pull(max_year) |> unique()}) |> set_names(inNames)
+    inputsList <- inputsList |> (function(list0, names0=inNames){
+      lNames0 <- list0   |> names()
+      str0    <- names0  |> paste0("\\.") |> paste(collapse="|")
+      lNames1 <- lNames0 |> str_replace(pattern=str0, "")
+      list0   <- list0   |> set_names(lNames1)
+      return(list0)
+    })()
+    ### Update names
+    inNames    <- inputsList |> names()
 
     ### Check input data
     inputsList <- list(
       inputName = inNames,
       inputDf   = inputsList,
-      # idCol     = idCols0 [inNames],
-      # valCol    = valCols0[inNames],
-      # yearMin   = df_inputInfo |> pull(min_year),
-      # yearMax   = df_inputInfo |> pull(max_year),
-      idCol     = idCols0 ,
-      valCol    = valCols0,
-      yearMin   = minYrs0,
-      yearMax   = maxYrs0,
+      idCol     = idCols0 [inNames],
+      valCol    = valCols0[inNames],
+      yearMin   = minYrs0 [inNames],
+      yearMax   = maxYrs0 [inNames],
[…]
