### 2024-08-20 3eb8ece
Fixed issue with columns in step to format input scenarios
Fixed issue with columns in step to format input scenarios:

- Added helper function to "utils.R" to help zero out temp and SLR scenarios
- Tested functions
