+++ FrEDI/R/utils.R
@@ -124,6 +124,47 @@ get_frediDataObj <- function(
   return(return_x)
 } ### End get_frediDataObj
 
+
+
+
+###### get_scenario_id ######
+### Function to standardize the creation of the scenario_id
+get_scenario_id <- function(
+    data_x,
+    include=c("region", "model") ### Character vector of column names to include
+){
+  ### Other vals
+  mnl0   <- "\n" ### Message new line
+  msg0   <- "\t" ### Message indent level 0
+  mcom   <- ", " ### Comma for collapsing lists
+  mqu0   <- "\'" ### Message quote
+  mqu1   <- mqu0 |> paste0(mcom, mqu0, collapse="")
+  mend0  <- "..."
+  ### Columns to include
+  main0  <- c("sector", "variant", "impactType", "impactYear")
+  cols0  <- main0  |> c(include)
+  ### Check names
+  names0 <- data_x |> names()
+  cCheck <- (cols0 %in% names0)
+  nCheck <- (!cCheck) |> which() |> length()
+  if(nCheck){
+    paste0("In get_scenario_id:") |> message()
+    msg0 |> paste0("Data is missing columns ", mqu0, cols0[!cCheck] |> paste(collapse=mqu1), mqu0, mend0) |> message()
+    paste0("Creating `scenario_id` from columns ", mqu0, cols0[cCheck] |> paste(collapse=mqu1), mqu0, mend0) |> message()
+    ### New names
+    cols0  <- cols0[cCheck]
+  } ### End if(nCheck)
+  ### Subset data
+  scen_x <- data_x[,cols0]
+  ### Get scenario IDs
+  scen_x <- scen_x |> apply(1, function(x){x |> as.vector() |> paste(collapse ="_")}) |> unlist()
+  data_x <- data_x |> mutate(scenario_id = scen_x)
+  ### Return
+  return(data_x)
+}
+
+
+
 ###### interpolate_annual ######
 ### Created 2021.02.05. Updated 2021.02.05
 ### Interpolate Annual Values
@@ -246,41 +287,54 @@ interpolate_annual <- function(
   return(data)
 } ### End function
 
-###### get_scenario_id ######
-### Function to standardize the creation of the scenario_id
-get_scenario_id <- function(
-    data_x,
-    include=c("region", "model") ### Character vector of column names to include
+
+
+### Function to zero out values for temperature and SLR scenarios
+zero_out_scenario     <- function(
+    df0,    ### Tibble with scenario and years, value column
+    type0   = "temp", ### Or SLR
+    valCol0 = NULL  , ### If NULL, function will use default
+    yrRef0  = NULL  , ### If NULL, function will use default
+    refVal0 = 0       ### Default to zero
 ){
-  ### Other vals
-  mnl0   <- "\n" ### Message new line
-  msg0   <- "\t" ### Message indent level 0
-  mcom   <- ", " ### Comma for collapsing lists
-  mqu0   <- "\'" ### Message quote
-  mqu1   <- mqu0 |> paste0(mcom, mqu0, collapse="")
-  mend0  <- "..."
-  ### Columns to include
-  main0  <- c("sector", "variant", "impactType", "impactYear")
-  cols0  <- main0  |> c(include)
-  ### Check names
-  names0 <- data_x |> names()
-  cCheck <- (cols0 %in% names0)
-  nCheck <- (!cCheck) |> which() |> length()
-  if(nCheck){
-    paste0("In get_scenario_id:") |> message()
[…]
