### 2024-08-20 e3e5845
Update build_fredi.yml
Updated build YML to update GDP, Population default scenarios
