### 2024-08-20 a2db49f
Deleted default results
- Deleted references to default results in "data.R" and "run_fredi.R".
- Updated examples in "run_fredi.R" and "run_fredi_sv.R" with new scenario data and file names
