+++ FrEDI/R/data.R
@@ -1,34 +1,3 @@
-###### defaultResults ######
-#' Default outputs of [FrEDI::run_fredi()]
-#
-#' A dataframe containing the default outputs of [FrEDI::run_fredi()]
-#
-#' @format A data frame with 1,501,500 rows and 20 columns:
-#' \describe{
-#'   \item{sector}{Name of the associated sector.}
-#'   \item{variant}{Name of the associated variant or adaptation (values are sector-specific).}
-#'   \item{impactYear}{Name of the impact year ("2010", "2090", "N/A", or "Interpolation").}
-#'   \item{impactType}{Name of the impact type ("all" or sector-specific values).}
-#'   \item{region}{Name of the associated region ("Midwest", "Northeast", "Northern Plains", "Northwest", "Southeast", "Southern Plains", "Southwest", or "National Total").}
-#'   \item{state}{Name of the associated state (or District of Columbia): "Alabama", "Arizona", "Arkansas", "California", "Colorado", "Connecticut", "Delaware", "District of Columbia", "Florida", "Georgia", "Idaho", "Illinois", "Indiana", "Iowa", "Kansas",  "Kentucky", "Louisiana", "Maine", "Maryland", "Massachusetts", "Michigan", "Minnesota",  "Mississippi", "Missouri", "Montana", "Nebraska", "Nevada", "New Hampshire",  "New Jersey", "New Mexico", "New York", "North Carolina", "North Dakota", "Ohio", "Oklahoma",  "Oregon", "Pennsylvania", "Rhode Island", "South Carolina", "South Dakota", "Tennessee", "Texas",  "Utah", "Vermont", "Virginia", "Washington", "West Virginia", "Wisconsin", "Wyoming". For the "National Total" region, `state = "All"`. For sectors with only has region-level data, `state = "N/A"`.}
-#'   \item{postal}{Postal code abbreviation of the associated state (e.g., `postal="AL"` for `state="Alabama"`). The postal code abbreviation for the District of Columbia is `"DC"`. For the "National Total" region, `postal = "US"`. For sectors with only has region-level data, `postal = "N/A"`.}
-#'   \item{model_type}{Type of model used to calculate impacts --- either "GCM" (for "Global Climate Model") or "SLR" for ("Sea Level Rise")}
-#'   \item{model}{Name of the GCM ("Average", "CanESM2", "CCSM4", "GCM Ensemble", "GFDL-CM3", "GISS-E2-R", "HadGEM2-ES", "MIROC5", "MRI-CGCM3") or SLR model ("Interpolation")}
-#'   \item{sectorprimary}{A `0` or `1` value indicating whether the sector is a primary CIRA sector (and whether it should be included when summing across sectors). `sectorprimary=1` if a primary sector and `sectorprimary=0` if not.}
-#'   \item{includeaggregate}{A `0` or `1` value indicating whether the variant is the primary variant for the sector (and whether it should be included when summing across sectors). `includeaggregate=1` if the variant is a primary variant and `includeaggregate=0` if not).}
-#'   \item{driverType}{Associated driver type ("Temperature" or "GMSL (SLR)").}
-#'   \item{driverUnit}{Unit for the associated driver type ("degrees Celsius" for temperature and "cm" for sea level rise).}
-#'   \item{driverValue}{Value for the associated driver type (in "degrees Celsius" for temperature and in "cm" for sea level rise).}
-#'   \item{gdp_usd}{National gross domestic product (GDP) for associated year in 2015 USD (U.S. dollars).}
-#'   \item{national_pop}{National population for associated year.}
-#'   \item{gdp_per_cap}{National GDP per capita for associated year in 2015 USD (U.S. dollars) per capita.}
[…]
