### 2024-08-20 7546cc4
Update build_fredi.yml
Added steps to save ozone, methane, and nox default scenarios
