+++ .github/workflows/build_fredi.yml
@@ -31,9 +31,10 @@ on:
         description: Update scenarios?
         required: true
         options:
-        - Temps & SLR
-        - Population
         - GDP
+        - Population
+        - Temps & SLR
+        - O3, CH4, & NOx
 
 jobs:
   compile_data:
@@ -65,10 +66,12 @@ jobs:
           Rscript -e '
             ###### Conditionals ######
             ### Which scenarios/documentation to update
-            do_temp  <- "Temps & SLR" %in% "${{ inputs.update_scenarios}}"
+            do_docs  <- "yes"         %in% "${{ inputs.update_scenarios}}"
             do_gdp   <- "GDP"         %in% "${{ inputs.update_scenarios}}"
             do_pop   <- "Pop"         %in% "${{ inputs.update_scenarios}}"
-            do_docs  <- "yes"         %in% "${{ inputs.update_scenarios}}"
+            do_temp  <- "Temps & SLR" %in% "${{ inputs.update_scenarios}}"
+            do_o3    <- "${{ inputs.update_scenarios}}" |> str_detect(pattern="O3") |> any()
+
 
 
             ###### File Paths ######
@@ -81,17 +84,6 @@ jobs:
 
 
             ###### Update Scenarios ######
-            ### Temperature & CSV
-            if(do_temp) {
-              gcamScenarios <- "gcam_scenarios" |> get_frediDataObj("scenarioData")
-              fTemps        <- "gcamScenarios"
-              dTemps        <- dPath0 |> paste0(fTemps, ".rda")
-              cTemps        <- cPath0 |> paste0(fTemps, ".csv")
-              gcamScenarios |> save(file=dTemps)
-              gcamScenarios |> write.csv(filename=cTemps, row.names=F)
-              rm(fTemps, dTemps, cTemps, gcamScenarios)
-            } ### End if(do_temp)
-
             ### GDP
             if(do_gdp) {
               gdpDefault    <- "gdp_default" |> get_frediDataObj("scenarioData")
@@ -113,6 +105,47 @@ jobs:
               popDefault |> write.csv(filename=cPop, row.names=F)
               rm(fPop, dPop, cPop, popDefault)
             } ### End if(do_pop)
+
+            ### Temperature & CSV
+            if(do_temp) {
+              gcamScenarios <- "gcam_scenarios" |> get_frediDataObj("scenarioData")
+              fTemps        <- "gcamScenarios"
+              dTemps        <- dPath0 |> paste0(fTemps, ".rda")
+              cTemps        <- cPath0 |> paste0(fTemps, ".csv")
+              gcamScenarios |> save(file=dTemps)
+              gcamScenarios |> write.csv(filename=cTemps, row.names=F)
+              rm(fTemps, dTemps, cTemps, gcamScenarios)
+            } ### End if(do_temp)
+
+            ### O3, CH4, and NOx
+            if(do_o3) {
+              ### Ozone
+              o3Default     <- "o3_default" |> get_frediDataObj("scenarioData")
+              fO3           <- "o3Default"
+              dO3           <- dPath0 |> paste0(fO3, ".rda")
+              cO3           <- cPath0 |> paste0(fO3, ".csv")
+              o3Default |> save(file=dO3)
+              o3Default |> write.csv(filename=cO3, row.names=F)
+              rm(fO3, dO3, cO3, o3Default)
+
+              ### CH4
+              ch4Default    <- "ch4_default" |> get_frediDataObj("scenarioData")
+              fCH4          <- "ch4Default"
+              dCH4          <- dPath0 |> paste0(fCH4, ".rda")
+              cCH4          <- cPath0 |> paste0(fCH4, ".csv")
[…]
