### 2024-08-20 aeb7962
Update run_fredi.R
- Added comments to some argument definitions
- Deleted deprecated comments
