### 2024-08-20 5786e5c
Update aggregate_impacts.R
Added steps to make columns more robust to FrEDI methane. Can't aggregate on FrEDI methane, however, unless driver scenario is dropped
