### 2024-08-21 2d2f240
Update build_fredi.yml
Added additional steps to standardize the columns and model names in the ozone scenario
