### 2024-08-21 759fa97
Update data.R
Deleted the default scenario from the data description and renamed `gdpScenario` and `popScenario` to `gdpDefault` and `popDefault`, respectively
