### 2024-08-21 d54cd56
Update run_fredi.R
Reorganized order of inputList item descriptions and corrected typos
