### 2024-08-21 2c8205e
Update import_inputs.R
Added additional information about regions in `import_inputs`
