### 2024-08-21 f2b6f5b
Update run_fredi_methane.R
Initialized function documentation for `run_fredi_methane`
