+++ FrEDI/R/run_fredi_methane.R
@@ -4,24 +4,149 @@
 #'
 #'
 #' @description
-#' This function allows users to estimate impacts from changes to methane, NOx, and ozone. Additional documentation to follow.
+#' This function allows users to estimate impacts from changes to concentrations of ozone, methane, and/or NOx in the atmosphere.
+#'
+#'
+#'
+#' @param inputsList=list(gdp=NULL,pop=NULL,ch4=NULL,nox=NULL,o3=NULL) A list with named elements (`gdp`, `pop`, `ch4`, `nox`, and/or `o3`), each containing data frames of custom scenarios for gross domestic product (GDP), state-level population, temperature, and/or global mean sea level rise (GMSL) trajectories, respectively, over a continuous period. Temperature and sea level rise inputs should start in 2000 or earlier. Values for population and GDP scenarios can start in 2010 or earlier. Values for each scenario type must be within reasonable ranges. For more information, see [FrEDI::import_inputs()].
+#'
+#' @param elasticity=1 A numeric value indicating an elasticity to use for adjusting VSL (defaults to `elasticity = 1`)
+#'
+#' @param maxYear=2100 A numeric value indicating the maximum year for the analysis. The range for `maxYear` is `[2011, 2300]`. Defaults to `maxYear = 2100`.
+#'
+#' @param thru2300=FALSE A ` TRUE/FALSE` shortcut that overrides the `maxYear` argument to run the model to 2300. Defaults to `thru2300 = FALSE`.
+#'
+#' @param outputList=FALSE A ` TRUE/FALSE` value indicating whether to output results as a data frame object (`outputList = FALSE`, default) or to return a list of objects (`outputList = TRUE`) that includes information about model provenance (including input arguments and input scenarios) along with the data frame of results.
+#'
+#' @param allCols=FALSE A `TRUE/FALSE` value indicating whether to include intermediate column values in results (e.g., physical and economic multipliers). Used in testing. Defaults to `allCols = FALSE`).
+#'
+#'
+#'
+#' @details This function allows users to estimate impacts from changes to concentrations of ozone, methane, and/or NOx in the atmosphere.
+#'
+#' Users can specify an optional list of custom scenarios with `inputsList` (for more information on the format of inputs, see [FrEDI::import_inputs()]). The function [FrEDI::import_inputs()] can be used to importing custom scenarios from CSV files. [FrEDI::import_inputs()] returns a list with named elements `gdp`, `pop`, `o3`, `ch4`, and/or `nox`, with each respectively containing a data frame with a custom scenario for GDP, state-level population, ozone concentration, methane concentration, and/or NOx concentration. If a user imports scenarios using [FrEDI::import_inputs()], they can pass the outputs of [FrEDI::import_inputs()] directly to the `run_fred_methane()` argument `inputsList`. Note that the documentation for [FrEDI::import_inputs()] can also provide additional guidance and specification on the formats for each scenario type.
+#'
[…]
