### 2024-08-23 74d5f69
Update utils_methane.R
Changed formula for `calc_methane_mortality` in "utils_methane.R" to use the natural log of population, rather than the relative population.  The rationale for the change from using relative population to population is that we only have one mortality rate for 2020, so we can’t take into account the fact that there are multiple different 2020 mortality rates in the original RFF data (despite constant population). The new formula uses coefficients for the “poplograw” (i.e., log of population) regression for all years.
