### 2024-08-23 995e469
Update utils_methane.R
Updated `calc_methane_mortality` with method to calculate mortality using log of population instead of population relative to RFF population
