+++ FrEDI/R/utils_methane.R
@@ -20,7 +20,7 @@ calc_methane_mortality <- function(
   # df0     <- df0 |> mutate(delta_rffPop = !!sym(pCol0) - rffPop)
   # df0     <- df0 |> mutate(rffFactor    = delta_rffPop * !!sym(sCol0) + !!sym(iCol0))
   df0     <- df0 |> mutate(logPop       = (!!sym(pCol0)) |> log())
-  df0     <- df0 |> mutate(rffFactor    = delta_rffPop * !!sym(sCol0) + !!sym(iCol0))
+  df0     <- df0 |> mutate(rffFactor    = logPop       * !!sym(sCol0) + !!sym(iCol0))
   df0     <- df0 |> mutate(respMrate    = rffFactor    * ifRespScalar)
   ### Return data
   return(df0)
@@ -80,6 +80,7 @@ format_methane_drivers <- function(
   select0  <- idCols0 |> c(sumCols0)
   df1      <- df1     |> select(all_of(select0))
   ### Join data
+  df0$model |> unique() |> print(); df1$model |> unique() |> print()
   join0    <- df0 |> names() |> get_matches(y=df1 |> names())
   df0      <- df0 |> left_join(df1, by=join0, relationship="many-to-many")
   rm(df1)
