### 2024-08-23 f9381ba
Standardized region, model labels
Added steps to standardize region and model in methane steps. Removed step in "build_fredi.yml" to standardize model name in O3 default (now done in pre-processing)
