### 2024-09-06 #5784 @pdebuyl
Hi @kthyng the package underwent important review within the administration recently from what I read at https://usepa.github.io/FrEDI/index.html . I emailed the author today and hope that they will be able to reply to us soon now.
